% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfp_utils.R
\name{build_xml_from_list}
\alias{build_xml_from_list}
\title{Build XML Request Body}
\usage{
build_xml_from_list(list, root_name = NULL, root = NULL,
  version = "v201508")
}
\arguments{
\item{list}{a \code{list} of data to fill the XML body}

\item{root_name}{a character string to be put in the 
topmost level of the created XML hierarchy}

\item{root}{a XML node to be placed as root 
in the returned XML document}

\item{version}{a character string matching a supported
API version}
}
\value{
a XML document
}
\description{
Parse data into XML format
}
\keyword{internal}

