% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ActivityGroupService.R
\name{dfp_createActivityGroups}
\alias{dfp_createActivityGroups}
\title{createActivityGroups}
\usage{
dfp_createActivityGroups(request_data)
}
\arguments{
\item{request_data}{a \code{list} or \code{data.frame} of data elements
to be formatted for a SOAP request (XML format, but passed as character string)}
}
\value{
a \code{list} containing all the elements of a createActivityGroupsResponse
}
\description{
Creates a new ActivityGroup objects.
}

