% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AdExclusionRuleService.R
\name{dfp_createAdExclusionRules}
\alias{dfp_createAdExclusionRules}
\title{createAdExclusionRules}
\usage{
dfp_createAdExclusionRules(request_data)
}
\arguments{
\item{request_data}{a \code{list} or \code{data.frame} of data elements
to be formatted for a SOAP request (XML format, but passed as character string)#' @return a \code{list} containing all the elements of a createAdExclusionRulesResponse}
}
\description{
Creates new AdExclusionRule objects.
}

