% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AdRuleService.R
\name{dfp_createAdRules}
\alias{dfp_createAdRules}
\title{createAdRules}
\usage{
dfp_createAdRules(request_data)
}
\arguments{
\item{request_data}{a \code{list} or \code{data.frame} of data elements
to be formatted for a SOAP request (XML format, but passed as character string)}
}
\value{
a \code{list} containing all the elements of a createAdRulesResponse
}
\description{
Creates new AdRule objects.
}
\seealso{
\href{https://developers.google.com/doubleclick-publishers/docs/reference/v201508/AdRuleService#createAdRules}{Google Documentation for createAdRules}
}

