% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BaseRateService.R
\name{dfp_createBaseRates}
\alias{dfp_createBaseRates}
\title{createBaseRates}
\usage{
dfp_createBaseRates(request_data)
}
\arguments{
\item{request_data}{a \code{list} or \code{data.frame} of data elements
to be formatted for a SOAP request (XML format, but passed as character string)}
}
\value{
a \code{list} containing all the elements of a createBaseRatesResponse
}
\description{
Creates a list of new BaseRate objects.
}

