% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CompanyService.R
\name{dfp_createCompanies}
\alias{dfp_createCompanies}
\title{createCompanies}
\usage{
dfp_createCompanies(request_data, as_df=FALSE)
}
\arguments{
\item{request_data}{a \code{list} or \code{data.frame} of data elements
to be formatted for a SOAP request (XML format, but passed as character string)}

\item{as_df}{a boolean indicating whether to attempt to parse the result into a \code{data.frame}}
}
\value{
a \code{list} or \code{data.frame} containing all the elements of a createCompaniesResponse
}
\description{
Creates new Company objects.
}
\seealso{
\href{https://developers.google.com/doubleclick-publishers/docs/reference/v201508/CompanyService#createCompanies}{Google Documentation for createCompanies}
}

