% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ContactService.R
\name{dfp_createContacts}
\alias{dfp_createContacts}
\title{createContacts}
\usage{
dfp_createContacts(request_data)
}
\arguments{
\item{request_data}{a \code{list} or \code{data.frame} of data elements
to be formatted for a SOAP request (XML format, but passed as character string)}
}
\value{
a \code{list} containing all the elements of a createContactsResponse
}
\description{
Creates new Contact objects.
}

