% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ContentBundleService.R
\name{dfp_createContentBundles}
\alias{dfp_createContentBundles}
\title{createContentBundles}
\usage{
dfp_createContentBundles(request_data)
}
\arguments{
\item{request_data}{a \code{list} or \code{data.frame} of data elements
to be formatted for a SOAP request (XML format, but passed as character string)}
}
\value{
a \code{list} containing all the elements of a createContentBundlesResponse
}
\description{
Creates new ContentBundle objects.
}

