% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ContentMetadataKeyHierarchyService.R
\name{dfp_createContentMetadataKeyHierarchies}
\alias{dfp_createContentMetadataKeyHierarchies}
\title{createContentMetadataKeyHierarchies}
\usage{
dfp_createContentMetadataKeyHierarchies(request_data)
}
\arguments{
\item{request_data}{a \code{list} or \code{data.frame} of data elements
to be formatted for a SOAP request (XML format, but passed as character string)}
}
\value{
a \code{list} containing all the elements of a createContentMetadataKeyHierarchiesResponse
}
\description{
Creates new ContentMetadataKeyHierarchy objects. The following fields are required: <ul> <li>ContentMetadataKeyHierarchy id</li> <li>ContentMetadataKeyHierarchy name</li> <li>ContentMetadataKeyHierarchy hierarchyLevels</li> </ul>
}

