% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreativeService.R
\name{dfp_createCreatives}
\alias{dfp_createCreatives}
\title{createCreatives}
\usage{
dfp_createCreatives(request_data, as_df=FALSE)
}
\arguments{
\item{request_data}{a \code{list} or \code{data.frame} of data elements
to be formatted for a SOAP request (XML format, but passed as character string)}

\item{as_df}{a boolean indicating whether to attempt to parse the result into a \code{data.frame}}
}
\value{
a \code{list} or \code{data.frame} containing all the elements of a createCreativesResponse
}
\description{
Creates new Creative objects.
}
\seealso{
\href{https://developers.google.com/doubleclick-publishers/docs/reference/v201508/CreativeService#createCreatives}{Google Documentation for createCreatives}
}

