% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreativeWrapperService.R
\name{dfp_createCreativeWrappers}
\alias{dfp_createCreativeWrappers}
\title{createCreativeWrappers}
\usage{
dfp_createCreativeWrappers(request_data, as_df=FALSE)
}
\arguments{
\item{request_data}{a \code{list} or \code{data.frame} of data elements
to be formatted for a SOAP request (XML format, but passed as character string)}

\item{as_df}{a boolean indicating whether to attempt to parse the result into a \code{data.frame}}
}
\value{
a \code{list} or \code{data.frame} containing all the elements of a createCreativeWrappersResponse
}
\description{
Creates a new CreativeWrapper objects. The following fields are required: <ul> <li>CreativeWrapper labelId</li> <li>CreativeWrapper ordering</li> <li>CreativeWrapper header or CreativeWrapper footer</li> </ul>
}
\seealso{
\href{https://developers.google.com/doubleclick-publishers/docs/reference/v201508/CreativeWrapperService#createCreativeWrappers}{Google Documentation for createCreativeWrappers}
}

