% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CustomFieldService.R
\name{dfp_createCustomFieldOptions}
\alias{dfp_createCustomFieldOptions}
\title{createCustomFieldOptions}
\usage{
dfp_createCustomFieldOptions(request_data)
}
\arguments{
\item{request_data}{a \code{list} or \code{data.frame} of data elements
to be formatted for a SOAP request (XML format, but passed as character string)#' @return a \code{list} containing all the elements of a createCustomFieldOptionsResponse}
}
\description{
Creates new CustomFieldOption objects. The following fields are required: <ul> <li>CustomFieldOption displayName</li> <li>CustomFieldOption customFieldId</li> </ul>
}

