% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CustomTargetingService.R
\name{dfp_createCustomTargetingKeys}
\alias{dfp_createCustomTargetingKeys}
\title{createCustomTargetingKeys}
\usage{
dfp_createCustomTargetingKeys(request_data)
}
\arguments{
\item{request_data}{a \code{list} or \code{data.frame} of data elements
to be formatted for a SOAP request (XML format, but passed as character string)#' @return a \code{list} containing all the elements of a createCustomTargetingKeysResponse}
}
\description{
Creates new CustomTargetingKey objects. The following fields are required: <ul> <li>CustomTargetingKey name</li> <li>CustomTargetingKey type</li> </ul>
}

