% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CustomTargetingService.R
\name{dfp_createCustomTargetingValues}
\alias{dfp_createCustomTargetingValues}
\title{createCustomTargetingValues}
\usage{
dfp_createCustomTargetingValues(request_data)
}
\arguments{
\item{request_data}{a \code{list} or \code{data.frame} of data elements
to be formatted for a SOAP request (XML format, but passed as character string)#' @return a \code{list} containing all the elements of a createCustomTargetingValuesResponse}
}
\description{
Creates new CustomTargetingValue objects. The following fields are required: <ul> <li>CustomTargetingValue customTargetingKeyId</li> <li>CustomTargetingValue name</li> </ul>
}

