% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LabelService.R
\name{dfp_createLabels}
\alias{dfp_createLabels}
\title{createLabels}
\usage{
dfp_createLabels(request_data)
}
\arguments{
\item{request_data}{a \code{list} or \code{data.frame} of data elements
to be formatted for a SOAP request (XML format, but passed as character string)#' @return a \code{list} containing all the elements of a createLabelsResponse}
}
\description{
Creates new Label objects.
}

