% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LineItemService.R
\name{dfp_createLineItems}
\alias{dfp_createLineItems}
\title{createLineItems}
\usage{
dfp_createLineItems(request_data)
}
\arguments{
\item{request_data}{a \code{list} or \code{data.frame} of data elements
to be formatted for a SOAP request (XML format, but passed as character string)#' @return a \code{list} containing all the elements of a createLineItemsResponse}
}
\description{
Creates new LineItem objects.
}

