% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PackageService.R
\name{dfp_createPackages}
\alias{dfp_createPackages}
\title{createPackages}
\usage{
dfp_createPackages(request_data)
}
\arguments{
\item{request_data}{a \code{list} or \code{data.frame} of data elements
to be formatted for a SOAP request (XML format, but passed as character string)#' @return a \code{list} containing all the elements of a createPackagesResponse}
}
\description{
Creates new Package objects. For each package, the following fields are required: <ul> <li>Package proposalId</li> <li>Package productPackageId</li> <li>Package name</li> </ul>
}

