% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProductPackageService.R
\name{dfp_createProductPackages}
\alias{dfp_createProductPackages}
\title{createProductPackages}
\usage{
dfp_createProductPackages(request_data, as_df=FALSE)
}
\arguments{
\item{request_data}{a \code{list} or \code{data.frame} of data elements
to be formatted for a SOAP request (XML format, but passed as character string)}

\item{as_df}{a boolean indicating whether to attempt to parse the result into a \code{data.frame}}
}
\value{
a \code{list} or \code{data.frame} containing all the elements of a createProductPackagesResponse
}
\description{
Creates new ProductPackage objects.
}
\seealso{
\href{https://developers.google.com/doubleclick-publishers/docs/reference/v201508/ProductPackageService#createProductPackages}{Google Documentation for createProductPackages}
}

