% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProductTemplateService.R
\name{dfp_createProductTemplates}
\alias{dfp_createProductTemplates}
\title{createProductTemplates}
\usage{
dfp_createProductTemplates(request_data)
}
\arguments{
\item{request_data}{a \code{list} or \code{data.frame} of data elements
to be formatted for a SOAP request (XML format, but passed as character string)}
}
\value{
a \code{list} containing all the elements of a createProductTemplatesResponse
}
\description{
Creates new ProductTemplate objects.
}

