% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProposalService.R
\name{dfp_createProposals}
\alias{dfp_createProposals}
\title{createProposals}
\usage{
dfp_createProposals(request_data, as_df=FALSE)
}
\arguments{
\item{request_data}{a \code{list} or \code{data.frame} of data elements
to be formatted for a SOAP request (XML format, but passed as character string)}

\item{as_df}{a boolean indicating whether to attempt to parse the result into a \code{data.frame}}
}
\value{
a \code{list} or \code{data.frame} containing all the elements of a createProposalsResponse
}
\description{
Creates new Proposal objects. For each proposal, the following fields are required: <ul> <li>Proposal name</li> </ul>
}
\seealso{
\href{https://developers.google.com/doubleclick-publishers/docs/reference/v201508/ProposalService#createProposals}{Google Documentation for createProposals}
}

