% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RateCardService.R
\name{dfp_createRateCards}
\alias{dfp_createRateCards}
\title{createRateCards}
\usage{
dfp_createRateCards(request_data, as_df=TRUE)
}
\arguments{
\item{request_data}{a \code{list} or \code{data.frame} of data elements
to be formatted for a SOAP request (XML format, but passed as character string)}

\item{as_df}{a boolean indicating whether to attempt to parse the result into a \code{data.frame}}
}
\value{
a \code{data.frame} or \code{list} containing all the elements of a createRateCardsResponse
}
\description{
Creates a list of RateCard objects. Rate cards must be activated before being associated with proposal line items and products.
}
\seealso{
\href{https://developers.google.com/doubleclick-publishers/docs/reference/v201508/RateCardService#createRateCards}{Google Documentation for createRateCards}
}

