% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UserTeamAssociationService.R
\name{dfp_createUserTeamAssociations}
\alias{dfp_createUserTeamAssociations}
\title{createUserTeamAssociations}
\usage{
dfp_createUserTeamAssociations(request_data)
}
\arguments{
\item{request_data}{a \code{list} or \code{data.frame} of data elements
to be formatted for a SOAP request (XML format, but passed as character string)}
}
\value{
a \code{list} containing all the elements of a createUserTeamAssociationsResponse
}
\description{
Creates new UserTeamAssociation objects.
}

