% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreativeSetService.R
\name{dfp_CreativeSetService_object_factory}
\alias{dfp_CreativeSetService_object_factory}
\title{CreativeSetService}
\usage{
dfp_CreativeSetService_object_factory(obj_type, obj_data)
}
\arguments{
\item{obj_type}{a string from one of the object types described below}

\item{obj_data}{a \code{list} of \code{lists} or a \code{data.frame} with the required components to create
the obj_type specified. All input names must match the expected attributes for the object.
If submitting a list of attributes to create 1 object make sure to nest it inside top-level list, so that
it is a list of length 1 with 1 list of attributes to create the object.}
}
\value{
a string of formatted XML that can be inserted into a SOAP request body to the API
}
\description{
Provides methods for adding, updating and retrieving CreativeSet objects.
}
\details{
CreativeSetService Object Factory

A function to create a variety of objects that are part of the CreativeSetService
Note: not all objects here are inputs to the functions of this service so 
please consult each function for the required inputs

Below is a list of objects and their required components to be created with this function:

\strong{ApiError}

The API error base class that provides details about an error that occurred
while processing a service request. The OGNL field path is provided for
parsers to identify the request data element that may have caused the error.
\href{https://developers.google.com/doubleclick-publishers/docs/reference/v201508/CreativeSetService.ApiError}{Google Documentation for ApiError}
\describe{
 \item{fieldPath}{a string -  The OGNL field path to identify cause of error.}
 \item{trigger}{a string -  The data that caused the error.}
 \item{errorString}{a string -  A simple string representation of the error and reason.}
}

\strong{ApiException}

Exception class for holding a list of service errors.
\href{https://developers.google.com/doubleclick-publishers/docs/reference/v201508/CreativeSetService.ApiException}{Google Documentation for ApiException}
\describe{
 \item{message}{a string (inherited from ApplicationException) -  Error message.}
 \item{errors}{a ApiError -  List of errors.}
}

\strong{ApiVersionError}

Errors related to the usage of API versions.
\href{https://developers.google.com/doubleclick-publishers/docs/reference/v201508/CreativeSetService.ApiVersionError}{Google Documentation for ApiVersionError}
\describe{
 \item{extends ApiError}{see documentation for ApiError}
 \item{reason}{a ApiVersionError.Reason - This can take one of the following values:
   \itemize{
     \item{UPDATE_TO_NEWER_VERSION - Indicates that the operation is not allowed in the version the request was
made in.}
     \item{UNKNOWN - The value returned if the actual value is not exposed by the requested API
version.}
   }
  }
}

\strong{ApplicationException}

Base class for exceptions.
\href{https://developers.google.com/doubleclick-publishers/docs/reference/v201508/CreativeSetService.ApplicationException}{Google Documentation for ApplicationException}
\describe{
 \item{message}{a string -  Error message.}
}

\strong{AssetError}

Lists all errors associated with assets.
\href{https://developers.google.com/doubleclick-publishers/docs/reference/v201508/CreativeSetService.AssetError}{Google Documentation for AssetError}
\describe{
 \item{extends ApiError}{see documentation for ApiError}
 \item{reason}{a AssetError.Reason - The error reason represented by an enum. This can take one of the following values:
   \itemize{
     \item{NON_UNIQUE_NAME - An asset name must be unique across advertiser.}
     \item{FILE_NAME_TOO_LONG - The file name is too long.}
     \item{FILE_SIZE_TOO_LARGE - The file size is too large.}
     \item{MISSING_REQUIRED_DYNAMIC_ALLOCATION_CLIENT - Required client code is not present in the code snippet.}
     \item{MISSING_REQUIRED_DYNAMIC_ALLOCATION_HEIGHT - Required height is not present in the code snippet.}
     \item{MISSING_REQUIRED_DYNAMIC_ALLOCATION_WIDTH - Required width is not present in the code snippet.}
     \item{MISSING_REQUIRED_DYNAMIC_ALLOCATION_FORMAT - Required format is not present in the mobile code snippet.}
     \item{INVALID_CODE_SNIPPET_PARAMETER_VALUE - The parameter value in the code snippet is invalid.}
     \item{INVALID_ASSET_ID - Invalid asset Id.}
     \item{UNKNOWN - The value returned if the actual value is not exposed by the requested API
version.}
   }
  }
}

\strong{AuthenticationError}

An error for an exception that occurred when authenticating.
\href{https://developers.google.com/doubleclick-publishers/docs/reference/v201508/CreativeSetService.AuthenticationError}{Google Documentation for AuthenticationError}
\describe{
 \item{extends ApiError}{see documentation for ApiError}
 \item{reason}{a AuthenticationError.Reason - This can take one of the following values:
   \itemize{
     \item{AMBIGUOUS_SOAP_REQUEST_HEADER - The SOAP message contains a request header with an ambiguous definition of
the authentication header fields. This means either the authToken and
oAuthToken fields were both null or both were specified. Exactly one value
should be specified with each request.}
     \item{INVALID_EMAIL - The login provided is invalid.}
     \item{AUTHENTICATION_FAILED - Tried to authenticate with provided information, but failed.}
     \item{INVALID_OAUTH_SIGNATURE - The OAuth provided is invalid.}
     \item{INVALID_SERVICE - The specified service to use was not recognized.}
     \item{MISSING_SOAP_REQUEST_HEADER - The SOAP message is missing a request header with an authToken and optional
networkCode.}
     \item{MISSING_AUTHENTICATION_HTTP_HEADER - The HTTP request is missing a request header with an authToken}
     \item{MISSING_AUTHENTICATION - The request is missing an authToken}
     \item{NOT_WHITELISTED_FOR_API_ACCESS - The customer is not whitelisted for API access.}
     \item{NO_NETWORKS_TO_ACCESS - The user is not associated with any network.}
     \item{NETWORK_NOT_FOUND - No network for the given networkCode was found.}
     \item{NETWORK_CODE_REQUIRED - The user has access to more than one network, but did not provide a
networkCode.}
     \item{CONNECTION_ERROR - An error happened on the server side during connection to authentication
service.}
     \item{GOOGLE_ACCOUNT_ALREADY_ASSOCIATED_WITH_NETWORK - The user tried to create a test network using an account that already is
associated with a network.}
     \item{UNKNOWN - The value returned if the actual value is not exposed by the requested API
version.}
   }
  }
}

\strong{BooleanValue}

Contains a boolean value.
\href{https://developers.google.com/doubleclick-publishers/docs/reference/v201508/CreativeSetService.BooleanValue}{Google Documentation for BooleanValue}
\describe{
 \item{extends Value}{see documentation for Value}
 \item{value}{a boolean -  The boolean value.}
}

\strong{CollectionSizeError}

Error for the size of the collection being too large
\href{https://developers.google.com/doubleclick-publishers/docs/reference/v201508/CreativeSetService.CollectionSizeError}{Google Documentation for CollectionSizeError}
\describe{
 \item{extends ApiError}{see documentation for ApiError}
 \item{reason}{a CollectionSizeError.Reason - This can take one of the following values:
   \itemize{
     \item{TOO_LARGE}
     \item{UNKNOWN - The value returned if the actual value is not exposed by the requested API
version.}
   }
  }
}

\strong{CommonError}

A place for common errors that can be used across services.
\href{https://developers.google.com/doubleclick-publishers/docs/reference/v201508/CreativeSetService.CommonError}{Google Documentation for CommonError}
\describe{
 \item{extends ApiError}{see documentation for ApiError}
 \item{reason}{a CommonError.Reason - This can take one of the following values:
   \itemize{
     \item{NOT_FOUND - Indicates that an attempt was made to retrieve an entity that does not
exist.}
     \item{ALREADY_EXISTS - Indicates that an attempt was made to create an entity that already exists.}
     \item{DUPLICATE_OBJECT - Indicates that two elements in the collection were identical.}
     \item{CANNOT_UPDATE - Indicates that an attempt was made to change an immutable field.}
     \item{CONCURRENT_MODIFICATION - Indicates that another request attempted to update the same data in the same
network at about the same time. Please wait and try the request again.}
     \item{UNKNOWN - The value returned if the actual value is not exposed by the requested API
version.}
   }
  }
}

\strong{CreativeAssetMacroError}

Lists all errors associated with creative asset macros.
\href{https://developers.google.com/doubleclick-publishers/docs/reference/v201508/CreativeSetService.CreativeAssetMacroError}{Google Documentation for CreativeAssetMacroError}
\describe{
 \item{extends ApiError}{see documentation for ApiError}
 \item{reason}{a CreativeAssetMacroError.Reason - The error reason represented by an enum. This can take one of the following values:
   \itemize{
     \item{INVALID_MACRO_NAME - Invalid macro name specified. Macro names must start with an alpha character
and consist only of alpha-numeric characters and underscores and be between
1 and 64 characters.}
     \item{UNKNOWN - The value returned if the actual value is not exposed by the requested API
version.}
   }
  }
}

\strong{CreativeError}

Lists all errors associated with creatives.
\href{https://developers.google.com/doubleclick-publishers/docs/reference/v201508/CreativeSetService.CreativeError}{Google Documentation for CreativeError}
\describe{
 \item{extends ApiError}{see documentation for ApiError}
 \item{reason}{a CreativeError.Reason - The error reason represented by an enum. This can take one of the following values:
   \itemize{
     \item{FLASH_AND_FALLBACK_URL_ARE_SAME - FlashRedirectCreative flashUrl and FlashRedirectCreative fallbackUrl are the
same. The fallback URL is used when the flash URL does not work and must be
different from it.}
     \item{INVALID_INTERNAL_REDIRECT_URL - The internal redirect URL was invalid. The URL must have the following
syntax http://ad.doubleclick.net/ad/sitename/;sz=size.}
     \item{DESTINATION_URL_REQUIRED - HasDestinationUrlCreative destinationUrl is required.}
     \item{CANNOT_CREATE_OR_UPDATE_LEGACY_DFP_CREATIVE - Cannot create or update legacy DART For Publishers creative.}
     \item{CANNOT_CREATE_OR_UPDATE_LEGACY_DFP_MOBILE_CREATIVE - Cannot create or update legacy mobile creative.}
     \item{MISSING_FEATURE - The user is missing a necessary feature.}
     \item{INVALID_COMPANY_TYPE - Company type should be one of Advertisers, House Advertisers and Ad
Networks.}
     \item{INVALID_ADSENSE_CREATIVE_SIZE - Invalid size for AdSense dynamic allocation creative. Only valid AFC sizes
are allowed.}
     \item{INVALID_AD_EXCHANGE_CREATIVE_SIZE - Invalid size for Ad Exchange dynamic allocation creative. Only valid Ad
Exchange sizes are allowed.}
     \item{DUPLICATE_ASSET_IN_CREATIVE - Assets associated with the same creative must be unique.}
     \item{CREATIVE_ASSET_CANNOT_HAVE_ID_AND_BYTE_ARRAY - A creative asset cannot contain an asset ID and a byte array.}
     \item{CANNOT_CREATE_OR_UPDATE_UNSUPPORTED_CREATIVE - Cannot create or update unsupported creative.}
     \item{CANNOT_CREATE_PROGRAMMATIC_CREATIVES - Cannot create programmatic creatives.}
     \item{UNKNOWN - The value returned if the actual value is not exposed by the requested API
version.}
   }
  }
}

\strong{CreativeSet}

A creative set is comprised of a master creative and its companion
creatives.
\href{https://developers.google.com/doubleclick-publishers/docs/reference/v201508/CreativeSetService.CreativeSet}{Google Documentation for CreativeSet}
\describe{
 \item{id}{a integer -  Uniquely identifies the CreativeSet. This attribute is read-only and is
assigned by Google when a creative set is created.}
 \item{name}{a string -  The name of the creative set. This attribute is required and has a maximum
length of 255 characters.}
 \item{masterCreativeId}{a integer -  The ID of the master creative associated with this creative set. This
attribute is required.}
 \item{companionCreativeIds}{a integer -  The IDs of the companion creatives associated with this creative set. This
attribute is required.}
 \item{lastModifiedDateTime}{a DateTime -  The date and time this creative set was last modified.}
}

\strong{CreativeSetError}

Errors relating to creative sets & subclasses.
\href{https://developers.google.com/doubleclick-publishers/docs/reference/v201508/CreativeSetService.CreativeSetError}{Google Documentation for CreativeSetError}
\describe{
 \item{extends ApiError}{see documentation for ApiError}
 \item{reason}{a CreativeSetError.Reason - The error reason represented by an enum. This can take one of the following values:
   \itemize{
     \item{VIDEO_FEATURE_REQUIRED - The 'video' feature is required but not enabled.}
     \item{CANNOT_CREATE_OR_UPDATE_VIDEO_CREATIVES - Video creatives (including overlays, VAST redirects, etc..) cannot be
created or updated through the API.}
     \item{ROADBLOCK_FEATURE_REQUIRED - The 'roadblock' feature is required but not enabled.}
     \item{MASTER_CREATIVE_CANNOT_BE_COMPANION - A master creative cannot be a companion creative in the same creative set.}
     \item{INVALID_ADVERTISER - Creatives in a creative set must be for the same advertiser.}
     \item{UPDATE_MASTER_CREATIVE_NOT_ALLOWED - Updating a master creative in a creative set is not allowed.}
     \item{UNKNOWN - The value returned if the actual value is not exposed by the requested API
version.}
   }
  }
}

\strong{CreativeSetPage}

Captures a page of CreativeSet objects.
\href{https://developers.google.com/doubleclick-publishers/docs/reference/v201508/CreativeSetService.CreativeSetPage}{Google Documentation for CreativeSetPage}
\describe{
 \item{totalResultSetSize}{a integer -  The size of the total result set to which this page belongs.}
 \item{startIndex}{a integer -  The absolute index in the total result set on which this page begins.}
 \item{results}{a CreativeSet -  The collection of creative sets contained within this page.}
}

\strong{CreativeTemplateError}

A catch-all error that lists all generic errors associated with
CreativeTemplate.
\href{https://developers.google.com/doubleclick-publishers/docs/reference/v201508/CreativeSetService.CreativeTemplateError}{Google Documentation for CreativeTemplateError}
\describe{
 \item{extends ApiError}{see documentation for ApiError}
 \item{reason}{a CreativeTemplateError.Reason - The error reason represented by an enum. This can take one of the following values:
   \itemize{
     \item{CANNOT_PARSE_CREATIVE_TEMPLATE - The XML of the creative template definition is malformed and cannot be
parsed.}
     \item{VARIABLE_DUPLICATE_UNIQUE_NAME - A creative template has multiple variables with the same uniqueName.}
     \item{VARIABLE_INVALID_UNIQUE_NAME - The creative template contains a variable with invalid characters. Valid
characters are letters, numbers, spaces, forward slashes, dashes, and
underscores.}
     \item{LIST_CHOICE_DUPLICATE_VALUE - Multiple choices for a CreativeTemplateListStringVariable have the same
value.}
     \item{LIST_CHOICE_NEEDS_DEFAULT - The choices for a CreativeTemplateListStringVariable do not contain the
default value.}
     \item{LIST_CHOICES_EMPTY - There are no choices specified for the list variable.}
     \item{NO_TARGET_PLATFORMS - No target platform is assigned to a creative template.}
     \item{MULTIPLE_TARGET_PLATFORMS - More than one target platform is assigned to a single creative template.}
     \item{UNRECOGNIZED_PLACEHOLDER - The formatter contains a placeholder which is not defined as a variable.}
     \item{PLACEHOLDERS_NOT_IN_FORMATTER - There are variables defined which are not being used in the formatter.}
     \item{MISSING_INTERSTITIAL_MACRO - The creative template is interstitial, but the formatter doesn't contain an
interstitial macro.}
     \item{UNKNOWN - The value returned if the actual value is not exposed by the requested API
version.}
   }
  }
}

\strong{CustomCreativeError}

Lists all errors associated with custom creatives.
\href{https://developers.google.com/doubleclick-publishers/docs/reference/v201508/CreativeSetService.CustomCreativeError}{Google Documentation for CustomCreativeError}
\describe{
 \item{extends ApiError}{see documentation for ApiError}
 \item{reason}{a CustomCreativeError.Reason - The error reason represented by an enum. This can take one of the following values:
   \itemize{
     \item{DUPLICATE_MACRO_NAME_FOR_CREATIVE - Macros associated with a single custom creative must have unique names.}
     \item{SNIPPET_REFERENCES_MISSING_MACRO - The file macro referenced in the snippet does not exist.}
     \item{UNRECOGNIZED_MACRO - The macro referenced in the snippet is not valid.}
     \item{CUSTOM_CREATIVE_NOT_ALLOWED - Custom creatives are not allowed in this context.}
     \item{MISSING_INTERSTITIAL_MACRO - The custom creative is an interstitial, but the snippet is missing an
interstitial macro.}
     \item{DUPLICATE_ASSET_IN_MACROS - Macros associated with the same custom creative cannot share the same asset.}
     \item{UNKNOWN - The value returned if the actual value is not exposed by the requested API
version.}
   }
  }
}

\strong{CustomFieldValueError}

Errors specific to editing custom field values
\href{https://developers.google.com/doubleclick-publishers/docs/reference/v201508/CreativeSetService.CustomFieldValueError}{Google Documentation for CustomFieldValueError}
\describe{
 \item{extends ApiError}{see documentation for ApiError}
 \item{reason}{a CustomFieldValueError.Reason - The error reason represented by an enum. This can take one of the following values:
   \itemize{
     \item{CUSTOM_FIELD_NOT_FOUND - An attempt was made to modify or create a CustomFieldValue for a CustomField
that does not exist.}
     \item{CUSTOM_FIELD_INACTIVE - An attempt was made to create a new value for a custom field that is
inactive.}
     \item{CUSTOM_FIELD_OPTION_NOT_FOUND - An attempt was made to modify or create a CustomFieldValue corresponding to
a CustomFieldOption that could not be found.}
     \item{UNKNOWN - The value returned if the actual value is not exposed by the requested API
version.}
   }
  }
}

\strong{Date}

Represents a date.
\href{https://developers.google.com/doubleclick-publishers/docs/reference/v201508/CreativeSetService.Date}{Google Documentation for Date}
\describe{
 \item{year}{a integer -  Year (e.g., 2009)}
 \item{month}{a integer -  Month (1..12)}
 \item{day}{a integer -  Day (1..31)}
}

\strong{DateTime}

Represents a date combined with the time of day.
\href{https://developers.google.com/doubleclick-publishers/docs/reference/v201508/CreativeSetService.DateTime}{Google Documentation for DateTime}
\describe{
 \item{date}{a Date}
 \item{hour}{a integer}
 \item{minute}{a integer}
 \item{second}{a integer}
 \item{timeZoneID}{a string}
}

\strong{DateTimeValue}

Contains a date-time value.
\href{https://developers.google.com/doubleclick-publishers/docs/reference/v201508/CreativeSetService.DateTimeValue}{Google Documentation for DateTimeValue}
\describe{
 \item{extends Value}{see documentation for Value}
 \item{value}{a DateTime -  The DateTime value.}
}

\strong{DateValue}

Contains a date value.
\href{https://developers.google.com/doubleclick-publishers/docs/reference/v201508/CreativeSetService.DateValue}{Google Documentation for DateValue}
\describe{
 \item{extends Value}{see documentation for Value}
 \item{value}{a Date -  The Date value.}
}

\strong{EntityLimitReachedError}

An error that occurs when creating an entity if the limit on the number of
allowed entities for a network has already been reached.
\href{https://developers.google.com/doubleclick-publishers/docs/reference/v201508/CreativeSetService.EntityLimitReachedError}{Google Documentation for EntityLimitReachedError}
\describe{
 \item{extends ApiError}{see documentation for ApiError}
 \item{reason}{a EntityLimitReachedError.Reason - This can take one of the following values:
   \itemize{
     \item{CUSTOM_TARGETING_VALUES_LIMIT_REACHED - The number of custom targeting values exceeds the max number allowed in the
network.}
     \item{AD_EXCLUSION_RULES_LIMIT_REACHED - The number of ad exclusion rules exceeds the max number allowed in the
network.}
     \item{FIRST_PARTY_AUDIENCE_SEGMENTS_LIMIT_REACHED - The number of first party audience segments exceeds the max number allowed
in the network.}
     \item{PLACEMENTS_LIMIT_REACHED - The number of active placements exceeds the max number allowed in the
network.}
     \item{LINE_ITEMS_LIMIT_REACHED - The number of line items excceeds the max number allowed in the network.}
     \item{UNKNOWN - The value returned if the actual value is not exposed by the requested API
version.}
   }
  }
}

\strong{FeatureError}

Errors related to feature management. If you attempt using a feature that is
not available to the current network you'll receive a FeatureError with the
missing feature as the trigger.
\href{https://developers.google.com/doubleclick-publishers/docs/reference/v201508/CreativeSetService.FeatureError}{Google Documentation for FeatureError}
\describe{
 \item{extends ApiError}{see documentation for ApiError}
 \item{reason}{a FeatureError.Reason - This can take one of the following values:
   \itemize{
     \item{MISSING_FEATURE - A feature is being used that is not enabled on the current network.}
     \item{UNKNOWN - The value returned if the actual value is not exposed by the requested API
version.}
   }
  }
}

\strong{FileError}

A list of all errors to be used for problems related to files.
\href{https://developers.google.com/doubleclick-publishers/docs/reference/v201508/CreativeSetService.FileError}{Google Documentation for FileError}
\describe{
 \item{extends ApiError}{see documentation for ApiError}
 \item{reason}{a FileError.Reason - This can take one of the following values:
   \itemize{
     \item{MISSING_CONTENTS - The provided byte array is empty.}
     \item{SIZE_TOO_LARGE - The provided file is larger than the maximum size defined for the network.}
     \item{UNKNOWN - The value returned if the actual value is not exposed by the requested API
version.}
   }
  }
}

\strong{ImageError}

Lists all errors associated with images.
\href{https://developers.google.com/doubleclick-publishers/docs/reference/v201508/CreativeSetService.ImageError}{Google Documentation for ImageError}
\describe{
 \item{extends ApiError}{see documentation for ApiError}
 \item{reason}{a ImageError.Reason - The error reason represented by an enum. This can take one of the following values:
   \itemize{
     \item{INVALID_IMAGE - The file's format is invalid.}
     \item{INVALID_SIZE - Size width and Size height cannot be negative.}
     \item{UNEXPECTED_SIZE - The actual image size does not match the expected image size.}
     \item{OVERLAY_SIZE_TOO_LARGE - The size of the asset is larger than that of the overlay creative.}
     \item{ANIMATED_NOT_ALLOWED - Animated images are not allowed.}
     \item{ANIMATION_TOO_LONG - Animation length exceeded the allowed policy limit.}
     \item{CMYK_JPEG_NOT_ALLOWED - Images in CMYK color formats are not allowed.}
     \item{FLASH_NOT_ALLOWED - Flash files are not allowed.}
     \item{FLASH_WITHOUT_CLICKTAG - If FlashCreative clickTagRequired is true, then the flash file is required
to have a click tag embedded in it.}
     \item{ANIMATED_VISUAL_EFFECT - Animated visual effect is not allowed.}
     \item{FLASH_ERROR - An error was encountered in the flash file.}
     \item{LAYOUT_PROBLEM - Incorrect image layout.}
     \item{FLASH_HAS_NETWORK_OBJECTS - Flash files with network objects are not allowed.}
     \item{FLASH_HAS_NETWORK_METHODS - Flash files with network methods are not allowed.}
     \item{FLASH_HAS_URL - Flash files with hard-coded click thru URLs are not allowed.}
     \item{FLASH_HAS_MOUSE_TRACKING - Flash files with mouse tracking are not allowed.}
     \item{FLASH_HAS_RANDOM_NUM - Flash files that generate or use random numbers are not allowed.}
     \item{FLASH_SELF_TARGETS - Flash files with self targets are not allowed.}
     \item{FLASH_BAD_GETURL_TARGET - Flash file contains a bad geturl target.}
     \item{FLASH_VERSION_NOT_SUPPORTED - Flash or ActionScript version in the submitted file is not supported.}
     \item{FILE_TOO_LARGE - The uploaded file is too large.}
     \item{SYSTEM_ERROR - A system error occured, please try again.}
     \item{UNEXPECTED_PRIMARY_ASSET_DENSITY - The image density for a primary asset was not one of the expected image
densities.}
     \item{DUPLICATE_ASSET_DENSITY - Two or more assets have the same image density.}
     \item{MISSING_DEFAULT_ASSET - The creative does not contain a primary asset. (For high-density creatives,
the primary asset must be a standard image file with 1x density.)}
     \item{UNKNOWN - The value returned if the actual value is not exposed by the requested API
version.}
   }
  }
}

\strong{InternalApiError}

Indicates that a server-side error has occured. InternalApiErrors are
generally not the result of an invalid request or message sent by the
client.
\href{https://developers.google.com/doubleclick-publishers/docs/reference/v201508/CreativeSetService.InternalApiError}{Google Documentation for InternalApiError}
\describe{
 \item{extends ApiError}{see documentation for ApiError}
 \item{reason}{a InternalApiError.Reason - The error reason represented by an enum. This can take one of the following values:
   \itemize{
     \item{UNEXPECTED_INTERNAL_API_ERROR - API encountered an unexpected internal error.}
     \item{TRANSIENT_ERROR - A temporary error occurred during the request. Please retry.}
     \item{UNKNOWN - The cause of the error is not known or only defined in newer versions.}
     \item{DOWNTIME - The API is currently unavailable for a planned downtime.}
   }
  }
}

\strong{InvalidPhoneNumberError}

Lists all errors associated with phone numbers.
\href{https://developers.google.com/doubleclick-publishers/docs/reference/v201508/CreativeSetService.InvalidPhoneNumberError}{Google Documentation for InvalidPhoneNumberError}
\describe{
 \item{extends ApiError}{see documentation for ApiError}
 \item{reason}{a InvalidPhoneNumberError.Reason - This can take one of the following values:
   \itemize{
     \item{INVALID_FORMAT - The phone number is invalid.}
     \item{TOO_SHORT - The phone number is too short.}
     \item{UNKNOWN - The value returned if the actual value is not exposed by the requested API
version.}
   }
  }
}

\strong{InvalidUrlError}

Lists all errors associated with URLs.
\href{https://developers.google.com/doubleclick-publishers/docs/reference/v201508/CreativeSetService.InvalidUrlError}{Google Documentation for InvalidUrlError}
\describe{
 \item{extends ApiError}{see documentation for ApiError}
 \item{reason}{a InvalidUrlError.Reason - This can take one of the following values:
   \itemize{
     \item{ILLEGAL_CHARACTERS - The URL contains invalid characters.}
     \item{INVALID_FORMAT - The format of the URL is not allowed. This could occur for a number of
reasons. For example, if an invalid scheme is specified (like "ftp://") or
if a port is specified when not required, or if a query was specified when
not required.}
     \item{INSECURE_SCHEME - URL contains insecure scheme.}
     \item{NO_SCHEME - The URL does not contain a scheme.}
     \item{UNKNOWN - The value returned if the actual value is not exposed by the requested API
version.}
   }
  }
}

\strong{LabelEntityAssociationError}

Errors specific to creating label entity associations.
\href{https://developers.google.com/doubleclick-publishers/docs/reference/v201508/CreativeSetService.LabelEntityAssociationError}{Google Documentation for LabelEntityAssociationError}
\describe{
 \item{extends ApiError}{see documentation for ApiError}
 \item{reason}{a LabelEntityAssociationError.Reason - The error reason represented by an enum. This can take one of the following values:
   \itemize{
     \item{DUPLICATE_ASSOCIATION - The label has already been attached to the entity.}
     \item{INVALID_ASSOCIATION - A label is being applied to an entity that does not support that entity
type.}
     \item{DUPLICATE_ASSOCIATION_WITH_NEGATION - The same label is being applied and negated to the same entity.}
     \item{UNKNOWN - The value returned if the actual value is not exposed by the requested API
version.}
   }
  }
}

\strong{NotNullError}

Caused by supplying a null value for an attribute that cannot be null.
\href{https://developers.google.com/doubleclick-publishers/docs/reference/v201508/CreativeSetService.NotNullError}{Google Documentation for NotNullError}
\describe{
 \item{extends ApiError}{see documentation for ApiError}
 \item{reason}{a NotNullError.Reason - The error reason represented by an enum. This can take one of the following values:
   \itemize{
     \item{ARG1_NULL - Assuming that a method will not have more than 3 arguments, if it does,
return NULL}
     \item{ARG2_NULL}
     \item{ARG3_NULL}
     \item{NULL}
     \item{UNKNOWN - The value returned if the actual value is not exposed by the requested API
version.}
   }
  }
}

\strong{NullError}

Errors associated with violation of a NOT NULL check.
\href{https://developers.google.com/doubleclick-publishers/docs/reference/v201508/CreativeSetService.NullError}{Google Documentation for NullError}
\describe{
 \item{extends ApiError}{see documentation for ApiError}
 \item{reason}{a NullError.Reason - The error reason represented by an enum. This can take one of the following values:
   \itemize{
     \item{NULL_CONTENT - Specified list/container must not contain any null elements}
   }
  }
}

\strong{NumberValue}

Contains a numeric value.
\href{https://developers.google.com/doubleclick-publishers/docs/reference/v201508/CreativeSetService.NumberValue}{Google Documentation for NumberValue}
\describe{
 \item{extends Value}{see documentation for Value}
 \item{value}{a string -  The numeric value represented as a string.}
}

\strong{ObjectValue}

Contains an object value.  This object is experimental! ObjectValue is an
experimental, innovative, and rapidly changing new feature for DFP.
Unfortunately, being on the bleeding edge means that we may make
backwards-incompatible changes to ObjectValue. We will inform the community
when this feature is no longer experimental.
\href{https://developers.google.com/doubleclick-publishers/docs/reference/v201508/CreativeSetService.ObjectValue}{Google Documentation for ObjectValue}
\describe{
 \item{extends Value}{see documentation for Value}
}

\strong{ParseError}

Lists errors related to parsing.
\href{https://developers.google.com/doubleclick-publishers/docs/reference/v201508/CreativeSetService.ParseError}{Google Documentation for ParseError}
\describe{
 \item{extends ApiError}{see documentation for ApiError}
 \item{reason}{a ParseError.Reason - The error reason represented by an enum. This can take one of the following values:
   \itemize{
     \item{UNPARSABLE - Indicates an error in parsing an attribute.}
     \item{UNKNOWN - The value returned if the actual value is not exposed by the requested API
version.}
   }
  }
}

\strong{PermissionError}

Errors related to incorrect permission.
\href{https://developers.google.com/doubleclick-publishers/docs/reference/v201508/CreativeSetService.PermissionError}{Google Documentation for PermissionError}
\describe{
 \item{extends ApiError}{see documentation for ApiError}
 \item{reason}{a PermissionError.Reason - This can take one of the following values:
   \itemize{
     \item{PERMISSION_DENIED - User does not have the required permission for the request.}
     \item{UNKNOWN - The value returned if the actual value is not exposed by the requested API
version.}
   }
  }
}

\strong{PublisherQueryLanguageContextError}

An error that occurs while executing a PQL query contained in a Statement
object.
\href{https://developers.google.com/doubleclick-publishers/docs/reference/v201508/CreativeSetService.PublisherQueryLanguageContextError}{Google Documentation for PublisherQueryLanguageContextError}
\describe{
 \item{extends ApiError}{see documentation for ApiError}
 \item{reason}{a PublisherQueryLanguageContextError.Reason - The error reason represented by an enum. This can take one of the following values:
   \itemize{
     \item{UNEXECUTABLE - Indicates that there was an error executing the PQL.}
     \item{UNKNOWN - The value returned if the actual value is not exposed by the requested API
version.}
   }
  }
}

\strong{PublisherQueryLanguageSyntaxError}

An error that occurs while parsing a PQL query contained in a Statement
object.
\href{https://developers.google.com/doubleclick-publishers/docs/reference/v201508/CreativeSetService.PublisherQueryLanguageSyntaxError}{Google Documentation for PublisherQueryLanguageSyntaxError}
\describe{
 \item{extends ApiError}{see documentation for ApiError}
 \item{reason}{a PublisherQueryLanguageSyntaxError.Reason - The error reason represented by an enum. This can take one of the following values:
   \itemize{
     \item{UNPARSABLE - Indicates that there was a PQL syntax error.}
     \item{UNKNOWN - The value returned if the actual value is not exposed by the requested API
version.}
   }
  }
}

\strong{QuotaError}

Describes a client-side error on which a user is attempting to perform an
action to which they have no quota remaining.
\href{https://developers.google.com/doubleclick-publishers/docs/reference/v201508/CreativeSetService.QuotaError}{Google Documentation for QuotaError}
\describe{
 \item{extends ApiError}{see documentation for ApiError}
 \item{reason}{a QuotaError.Reason - This can take one of the following values:
   \itemize{
     \item{EXCEEDED_QUOTA - The number of requests made per second is too high and has exceeded the
allowable limit. The recommended approach to handle this error is to wait
about 5 seconds and then retry the request. Note that this does not
guarantee the request will succeed. If it fails again, try increasing the
wait time.  Another way to mitigate this error is to limit requests to 2 per
second for Small Business networks, or 8 per second for Premium networks.
Once again this does not guarantee that every request will succeed, but may
help reduce the number of times you receive this error.}
     \item{UNKNOWN - The value returned if the actual value is not exposed by the requested API
version.}
   }
  }
}

\strong{RangeError}

A list of all errors associated with the Range constraint.
\href{https://developers.google.com/doubleclick-publishers/docs/reference/v201508/CreativeSetService.RangeError}{Google Documentation for RangeError}
\describe{
 \item{extends ApiError}{see documentation for ApiError}
 \item{reason}{a RangeError.Reason - This can take one of the following values:
   \itemize{
     \item{TOO_HIGH}
     \item{TOO_LOW}
     \item{UNKNOWN - The value returned if the actual value is not exposed by the requested API
version.}
   }
  }
}

\strong{RequiredCollectionError}

A list of all errors to be used for validating sizes of collections.
\href{https://developers.google.com/doubleclick-publishers/docs/reference/v201508/CreativeSetService.RequiredCollectionError}{Google Documentation for RequiredCollectionError}
\describe{
 \item{extends ApiError}{see documentation for ApiError}
 \item{reason}{a RequiredCollectionError.Reason - This can take one of the following values:
   \itemize{
     \item{REQUIRED - A required collection is missing.}
     \item{TOO_LARGE - Collection size is too large.}
     \item{TOO_SMALL - Collection size is too small.}
     \item{UNKNOWN - The value returned if the actual value is not exposed by the requested API
version.}
   }
  }
}

\strong{RequiredError}

Errors due to missing required field.
\href{https://developers.google.com/doubleclick-publishers/docs/reference/v201508/CreativeSetService.RequiredError}{Google Documentation for RequiredError}
\describe{
 \item{extends ApiError}{see documentation for ApiError}
 \item{reason}{a RequiredError.Reason - The error reason represented by an enum. This can take one of the following values:
   \itemize{
     \item{REQUIRED - Missing required field.}
   }
  }
}

\strong{RequiredNumberError}

A list of all errors to be used in conjunction with required number
validators.
\href{https://developers.google.com/doubleclick-publishers/docs/reference/v201508/CreativeSetService.RequiredNumberError}{Google Documentation for RequiredNumberError}
\describe{
 \item{extends ApiError}{see documentation for ApiError}
 \item{reason}{a RequiredNumberError.Reason - This can take one of the following values:
   \itemize{
     \item{REQUIRED}
     \item{TOO_LARGE}
     \item{TOO_SMALL}
     \item{TOO_LARGE_WITH_DETAILS}
     \item{TOO_SMALL_WITH_DETAILS}
     \item{UNKNOWN - The value returned if the actual value is not exposed by the requested API
version.}
   }
  }
}

\strong{RequiredSizeError}

A list of all errors to be used for validating Size.
\href{https://developers.google.com/doubleclick-publishers/docs/reference/v201508/CreativeSetService.RequiredSizeError}{Google Documentation for RequiredSizeError}
\describe{
 \item{extends ApiError}{see documentation for ApiError}
 \item{reason}{a RequiredSizeError.Reason - This can take one of the following values:
   \itemize{
     \item{REQUIRED - Creative size or LineItem creativeSizes is missing.}
     \item{NOT_ALLOWED - LineItemCreativeAssociation sizes must be a subset of LineItem
creativeSizes.}
     \item{UNKNOWN - The value returned if the actual value is not exposed by the requested API
version.}
   }
  }
}

\strong{RichMediaStudioCreativeError}

Lists all errors associated with Rich Media Studio creatives.
\href{https://developers.google.com/doubleclick-publishers/docs/reference/v201508/CreativeSetService.RichMediaStudioCreativeError}{Google Documentation for RichMediaStudioCreativeError}
\describe{
 \item{extends ApiError}{see documentation for ApiError}
 \item{reason}{a RichMediaStudioCreativeError.Reason - The error reason represented by an enum. This can take one of the following values:
   \itemize{
     \item{CREATION_NOT_ALLOWED - Only DoubleClick Rich Media Studio can create a RichMediaStudioCreative.}
     \item{UKNOWN_ERROR - Unknown error}
     \item{INVALID_CODE_GENERATION_REQUEST - Invalid request indicating missing/invalid request parameters.}
     \item{INVALID_CREATIVE_OBJECT - Invalid creative object.}
     \item{STUDIO_CONNECTION_ERROR - Unable to connect to Rich Media Studio to save the creative. Please try
again later.}
     \item{PUSHDOWN_DURATION_NOT_ALLOWED - The push down duration is not allowed}
     \item{INVALID_POSITION - The position is invalid}
     \item{INVALID_Z_INDEX - The Z-index is invalid}
     \item{INVALID_PUSHDOWN_DURATION - The push-down duration is invalid}
     \item{UNKNOWN - The value returned if the actual value is not exposed by the requested API
version.}
   }
  }
}

\strong{ServerError}

Errors related to the server.
\href{https://developers.google.com/doubleclick-publishers/docs/reference/v201508/CreativeSetService.ServerError}{Google Documentation for ServerError}
\describe{
 \item{extends ApiError}{see documentation for ApiError}
 \item{reason}{a ServerError.Reason - This can take one of the following values:
   \itemize{
     \item{SERVER_ERROR - Indicates that an unexpected error occured.}
     \item{SERVER_BUSY - Indicates that the server is currently experiencing a high load. Please wait
and try your request again.}
     \item{UNKNOWN - The value returned if the actual value is not exposed by the requested API
version.}
   }
  }
}

\strong{SetTopBoxCreativeError}

Errors associated with SetTopBoxCreative set-top box creatives.
\href{https://developers.google.com/doubleclick-publishers/docs/reference/v201508/CreativeSetService.SetTopBoxCreativeError}{Google Documentation for SetTopBoxCreativeError}
\describe{
 \item{extends ApiError}{see documentation for ApiError}
 \item{reason}{a SetTopBoxCreativeError.Reason - The error reason represented by an enum. This can take one of the following values:
   \itemize{
     \item{DURATION_IMMUTABLE - Set-top box creative durations are immutable after creation.}
     \item{EXTERNAL_ASSET_ID_IMMUTABLE - Set-top box creative external asset IDs are immutable after creation.}
     \item{EXTERNAL_ASSET_ID_REQUIRED - Set-top box creatives require an external asset ID.}
     \item{PROVIDER_ID_IMMUTABLE - Set-top box creative provider IDs are immutable after creation.}
     \item{UNKNOWN - The value returned if the actual value is not exposed by the requested API
version.}
   }
  }
}

\strong{SetValue}

Contains a set of Value Values. May not contain duplicates.
\href{https://developers.google.com/doubleclick-publishers/docs/reference/v201508/CreativeSetService.SetValue}{Google Documentation for SetValue}
\describe{
 \item{extends Value}{see documentation for Value}
 \item{values}{a Value -  The values. They must all be the same type of Value and not contain
duplicates.}
}

\strong{SoapRequestHeader}

Represents the SOAP request header used by API requests.
\href{https://developers.google.com/doubleclick-publishers/docs/reference/v201508/CreativeSetService.SoapRequestHeader}{Google Documentation for SoapRequestHeader}
\describe{
 \item{networkCode}{a string -  The network code to use in the context of a request.}
 \item{applicationName}{a string -  The name of client library application.}
}

\strong{SoapResponseHeader}

Represents the SOAP request header used by API responses.
\href{https://developers.google.com/doubleclick-publishers/docs/reference/v201508/CreativeSetService.SoapResponseHeader}{Google Documentation for SoapResponseHeader}
\describe{
 \item{requestId}{a string}
 \item{responseTime}{a integer}
}

\strong{Statement}

Captures the WHERE, ORDER BY and LIMIT clauses of a PQL query. Statements
are typically used to retrieve objects of a predefined domain type, which
makes SELECT clause unnecessary.  An example query text might be "WHERE
status = 'ACTIVE' ORDER BY id LIMIT 30".  Statements support bind variables.
These are substitutes for literals and can be thought of as input parameters
to a PQL query.  An example of such a query might be "WHERE id = :idValue".
Statements also support use of the LIKE keyword. This provides partial and
wildcard string matching.  An example of such a query might be "WHERE name
LIKE 'startswith\%'".  If using an API version newer than V201010, the value
for the variable idValue must then be set with an object of type Value and
is one of NumberValue, TextValue or BooleanValue.  If using an API version
older than or equal to V201010, the value for the variable idValue must then
be set with an object of type Param and is one of DoubleParam, LongParam or
StringParam.
\href{https://developers.google.com/doubleclick-publishers/docs/reference/v201508/CreativeSetService.Statement}{Google Documentation for Statement}
\describe{
 \item{query}{a string -  Holds the query in PQL syntax. See link for how to form PQL.}
 \item{values}{a String_ValueMapEntry -  Holds keys and values for bind variables and their values. The key is the
name of the bind variable. The value is the literal value of the variable.
In the example "WHERE status = :bindStatus ORDER BY id LIMIT 30", the bind
variable, represented by :bindStatus is named bindStatus, which would also
be the parameter map key. The bind variable's value would be represented by
a parameter map value of type TextValue. The final result, for example,
would be an entry of "bindStatus" => StringParam("ACTIVE").}
}

\strong{StatementError}

An error that occurs while parsing Statement objects.
\href{https://developers.google.com/doubleclick-publishers/docs/reference/v201508/CreativeSetService.StatementError}{Google Documentation for StatementError}
\describe{
 \item{extends ApiError}{see documentation for ApiError}
 \item{reason}{a StatementError.Reason - The error reason represented by an enum. This can take one of the following values:
   \itemize{
     \item{VARIABLE_NOT_BOUND_TO_VALUE - A bind variable has not been bound to a value.}
     \item{UNKNOWN - The value returned if the actual value is not exposed by the requested API
version.}
   }
  }
}

\strong{String_ValueMapEntry}

This represents an entry in a map with a key of type String and value of
type Value.
\href{https://developers.google.com/doubleclick-publishers/docs/reference/v201508/CreativeSetService.String_ValueMapEntry}{Google Documentation for String_ValueMapEntry}
\describe{
 \item{key}{a string}
 \item{value}{a Value}
}

\strong{StringLengthError}

Errors for Strings which do not meet given length constraints.
\href{https://developers.google.com/doubleclick-publishers/docs/reference/v201508/CreativeSetService.StringLengthError}{Google Documentation for StringLengthError}
\describe{
 \item{extends ApiError}{see documentation for ApiError}
 \item{reason}{a StringLengthError.Reason - This can take one of the following values:
   \itemize{
     \item{TOO_LONG}
     \item{TOO_SHORT}
     \item{UNKNOWN - The value returned if the actual value is not exposed by the requested API
version.}
   }
  }
}

\strong{SwiffyConversionError}

Error for converting flash to swiffy asset.
\href{https://developers.google.com/doubleclick-publishers/docs/reference/v201508/CreativeSetService.SwiffyConversionError}{Google Documentation for SwiffyConversionError}
\describe{
 \item{extends ApiError}{see documentation for ApiError}
 \item{reason}{a SwiffyConversionError.Reason - This can take one of the following values:
   \itemize{
     \item{SERVER_ERROR - Indicates the Swiffy service has an internal error that prevents the flash
asset being converted.}
     \item{INVALID_FLASH_FILE - Indicates the uploaded flash asset is not a valid flash file.}
     \item{UNSUPPORTED_FLASH - Indicates the Swiffy service currently does not support converting this
flash asset.}
     \item{UNKNOWN - The value returned if the actual value is not exposed by the requested API
version.}
   }
  }
}

\strong{TemplateInstantiatedCreativeError}

Lists all errors associated with template instantiated creatives.
\href{https://developers.google.com/doubleclick-publishers/docs/reference/v201508/CreativeSetService.TemplateInstantiatedCreativeError}{Google Documentation for TemplateInstantiatedCreativeError}
\describe{
 \item{extends ApiError}{see documentation for ApiError}
 \item{reason}{a TemplateInstantiatedCreativeError.Reason - The error reason represented by an enum. This can take one of the following values:
   \itemize{
     \item{INACTIVE_CREATIVE_TEMPLATE - A new creative cannot be created from an inactive creative template.}
     \item{FILE_TYPE_NOT_ALLOWED - An uploaded file type is not allowed}
     \item{UNKNOWN - The value returned if the actual value is not exposed by the requested API
version.}
   }
  }
}

\strong{TextValue}

Contains a string value.
\href{https://developers.google.com/doubleclick-publishers/docs/reference/v201508/CreativeSetService.TextValue}{Google Documentation for TextValue}
\describe{
 \item{extends Value}{see documentation for Value}
 \item{value}{a string -  The string value.}
}

\strong{TypeError}

An error for a field which is an invalid type.
\href{https://developers.google.com/doubleclick-publishers/docs/reference/v201508/CreativeSetService.TypeError}{Google Documentation for TypeError}
\describe{
 \item{extends ApiError}{see documentation for ApiError}
}

\strong{UniqueError}

An error for a field which must satisfy a uniqueness constraint
\href{https://developers.google.com/doubleclick-publishers/docs/reference/v201508/CreativeSetService.UniqueError}{Google Documentation for UniqueError}
\describe{
 \item{extends ApiError}{see documentation for ApiError}
}

\strong{Value}

Value represents a value.
\href{https://developers.google.com/doubleclick-publishers/docs/reference/v201508/CreativeSetService.Value}{Google Documentation for Value}
\describe{
}
}

