% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ActivityService.R
\name{dfp_getActivitiesByStatement}
\alias{dfp_getActivitiesByStatement}
\title{getActivitiesByStatement}
\usage{
dfp_getActivitiesByStatement(request_data)
}
\arguments{
\item{request_data}{a \code{list} or \code{data.frame} of data elements
to be formatted for a SOAP request (XML format, but passed as character string)}
}
\value{
a \code{list} containing all the elements of a getActivitiesByStatementResponse
}
\description{
Gets an ActivityPage of Activity objects that satisfy the given Statement query. The following fields are supported for filtering:  
\itemize{
  \item{id}
  \item{name}
  \item{expectedURL}
  \item{status}
  \item{activityGroupId}
}
}

