% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ActivityGroupService.R
\name{dfp_getActivityGroupsByStatement}
\alias{dfp_getActivityGroupsByStatement}
\title{getActivityGroupsByStatement}
\usage{
dfp_getActivityGroupsByStatement(request_data)
}
\arguments{
\item{request_data}{a \code{list} or \code{data.frame} of data elements
to be formatted for a SOAP request (XML format, but passed as character string)}
}
\value{
a \code{list} containing all the elements of a getActivityGroupsByStatementResponse
}
\description{
Gets an ActivityGroupPage of ActivityGroup objects that satisfy the given Statement query. The following fields are supported for filtering:  
\itemize{
  \item{id}
  \item{name}
  \item{impressionsLookback}
  \item{clicksLookback}
  \item{status}
}
}

