% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AdExclusionRuleService.R
\name{dfp_getAdExclusionRulesByStatement}
\alias{dfp_getAdExclusionRulesByStatement}
\title{getAdExclusionRulesByStatement}
\usage{
dfp_getAdExclusionRulesByStatement(request_data, as_df=TRUE)
}
\arguments{
\item{request_data}{a \code{list} or \code{data.frame} of data elements
to be formatted for a SOAP request (XML format, but passed as character string)}

\item{as_df}{a boolean indicating whether to attempt to parse the result into a \code{data.frame}}
}
\value{
a \code{data.frame} or \code{list} containing all the elements of a getAdExclusionRulesByStatementResponse
}
\description{
Gets a AdExclusionRulePage of AdExclusionRule objects that satisfy the given Statement query. The following fields are supported for filtering:  a set of rules 
\itemize{
  \item{id}
  \item{name}
  \item{status}
}
}
\seealso{
\href{https://developers.google.com/doubleclick-publishers/docs/reference/v201508/AdExclusionRuleService#getAdExclusionRulesByStatement}{Google Documentation for getAdExclusionRulesByStatement}
}

