% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InventoryService.R
\name{dfp_getAdUnitsByStatement}
\alias{dfp_getAdUnitsByStatement}
\title{getAdUnitsByStatement}
\usage{
dfp_getAdUnitsByStatement(request_data)
}
\arguments{
\item{request_data}{a \code{list} or \code{data.frame} of data elements
to be formatted for a SOAP request (XML format, but passed as character string)}
}
\value{
a \code{list} containing all the elements of a getAdUnitsByStatementResponse
}
\description{
Gets a AdUnitPage of AdUnit objects that satisfy the given Statement query. The following fields are supported for filtering:  a set of ad units 
\itemize{
  \item{adUnitCode}
  \item{id}
  \item{name}
  \item{parentId}
  \item{status}
  \item{lastModifiedDateTime}
}
}

