% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InventoryService.R
\name{dfp_getAdUnitSizesByStatement}
\alias{dfp_getAdUnitSizesByStatement}
\title{getAdUnitSizesByStatement}
\usage{
dfp_getAdUnitSizesByStatement(request_data, as_df=TRUE)
}
\arguments{
\item{request_data}{a \code{list} or \code{data.frame} of data elements
to be formatted for a SOAP request (XML format, but passed as character string)}

\item{as_df}{a boolean indicating whether to attempt to parse the result into a \code{data.frame}}
}
\value{
a \code{data.frame} or \code{list} containing all the elements of a getAdUnitSizesByStatementResponse
}
\description{
Gets a set of AdUnitSize objects that satisfy the given Statement query. The following fields are supported for filtering:  An exception will be thrown for queries with unsupported fields. Paging is not supported, as aren't the LIMIT and OFFSET PQL keywords. Only "=" operator is supported. a set of ad unit sizes 
\itemize{
  \item{targetPlatform}
}
}
\seealso{
\href{https://developers.google.com/doubleclick-publishers/docs/reference/v201508/InventoryService#getAdUnitSizesByStatement}{Google Documentation for getAdUnitSizesByStatement}
}

