% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UserService.R
\name{dfp_getAllRoles}
\alias{dfp_getAllRoles}
\title{getAllRoles}
\usage{
dfp_getAllRoles(as_df=FALSE)
}
\arguments{
\item{as_df}{a boolean indicating whether to attempt to parse the result into a \code{data.frame}}
}
\value{
a \code{list} or \code{data.frame} containing all the elements of a getAllRolesResponse
}
\description{
Returns the Role objects that are defined for the users of the network.
}
\seealso{
\href{https://developers.google.com/doubleclick-publishers/docs/reference/v201508/UserService#getAllRoles}{Google Documentation for getAllRoles}
}

