% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CompanyService.R
\name{dfp_getCompaniesByStatement}
\alias{dfp_getCompaniesByStatement}
\title{getCompaniesByStatement}
\usage{
dfp_getCompaniesByStatement(request_data, as_df = TRUE, verbose = FALSE)
}
\arguments{
\item{request_data}{a \code{list} or \code{data.frame} of data elements
to be formatted for a SOAP
request (XML format, but passed as character string)}

\item{as_df}{a boolean indicating whether to attempt to parse the result into
a \code{data.frame}}

\item{verbose}{a boolean indicating whether to print the service URL and POSTed XML}
}
\value{
a \code{data.frame} or \code{list} containing all the elements of a getCompaniesByStatementResponse
}
\description{
Gets a CompanyPage of Company objects that satisfy the given Statement query. The following fields are supported for filtering: 
\itemize{
  \item{id}
  \item{name}
  \item{type}
  \item{lastModifiedDateTime}
}
}
\examples{
\dontrun{
 dat <- list(filterStatement=list('query'="WHERE status='ACTIVE'")) 
 res <- dfp_getCompaniesByStatement(dat)
}
}
\seealso{
\href{https://developers.google.com/doubleclick-publishers/docs/reference/v201802/CompanyService#getCompaniesByStatement}{Google Documentation for getCompaniesByStatement}
}
