% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ContactService.R
\name{dfp_getContactsByStatement}
\alias{dfp_getContactsByStatement}
\title{getContactsByStatement}
\usage{
dfp_getContactsByStatement(request_data)
}
\arguments{
\item{request_data}{a \code{list} or \code{data.frame} of data elements
to be formatted for a SOAP request (XML format, but passed as character string)}
}
\value{
a \code{list} containing all the elements of a getContactsByStatementResponse
}
\description{
Gets a ContactPage of Contact objects that satisfy the given Statement query. The following fields are supported for filtering:  a set of contacts 
\itemize{
  \item{name}
  \item{email}
  \item{id}
  \item{comment}
  \item{companyId}
  \item{title}
  \item{cellPhone}
  \item{workPhone}
  \item{faxPhone}
  \item{status}
}
}

