% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ContentBundleService.R
\name{dfp_getContentBundlesByStatement}
\alias{dfp_getContentBundlesByStatement}
\title{getContentBundlesByStatement}
\usage{
dfp_getContentBundlesByStatement(request_data)
}
\arguments{
\item{request_data}{a \code{list} or \code{data.frame} of data elements
to be formatted for a SOAP request (XML format, but passed as character string)}
}
\value{
a \code{list} containing all the elements of a getContentBundlesByStatementResponse
}
\description{
Gets a ContentBundlePage of ContentBundle objects that satisfy the given Statement query. The following fields are supported for filtering:  a set of content bundles 
\itemize{
  \item{id}
  \item{name}
  \item{status}
}
}
\seealso{
\href{https://developers.google.com/doubleclick-publishers/docs/reference/v201508/ContentBundleService#getContentBundlesByStatement}{Google Documentation for getContentBundlesByStatement}
}

