% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ContentService.R
\name{dfp_getContentByStatement}
\alias{dfp_getContentByStatement}
\title{getContentByStatement}
\usage{
dfp_getContentByStatement(request_data)
}
\arguments{
\item{request_data}{a \code{list} or \code{data.frame} of data elements
to be formatted for a SOAP request (XML format, but passed as character string)#' @return a \code{list} containing all the elements of a getContentByStatementResponse}
}
\description{
Gets a ContentPage of Content objects that satisfy the given Statement query. The following fields are supported for filtering:  filter a set of content 
\itemize{
  \item{id}
  \item{status}
  \item{name}
  \item{lastModifiedDateTime}
}
}

