% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ContentService.R
\name{dfp_getContentByStatementAndCustomTargetingValue}
\alias{dfp_getContentByStatementAndCustomTargetingValue}
\title{getContentByStatementAndCustomTargetingValue}
\usage{
dfp_getContentByStatementAndCustomTargetingValue()
}
\value{
a \code{list} containing all the elements of a getContentByStatementAndCustomTargetingValueResponse
}
\description{
Gets a ContentPage of Content objects that satisfy the given Statement query. Additionally, filters on the given value ID and key ID that the value belongs to. The following fields are supported for filtering:  filter a set of content 
\itemize{
  \item{id}
  \item{status}
  \item{name}
  \item{lastModifiedDateTime}
}
}

