% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ContentMetadataKeyHierarchyService.R
\name{dfp_getContentMetadataKeyHierarchiesByStatement}
\alias{dfp_getContentMetadataKeyHierarchiesByStatement}
\title{getContentMetadataKeyHierarchiesByStatement}
\usage{
dfp_getContentMetadataKeyHierarchiesByStatement(request_data)
}
\arguments{
\item{request_data}{a \code{list} or \code{data.frame} of data elements
to be formatted for a SOAP request (XML format, but passed as character string)}
}
\value{
a \code{list} containing all the elements of a getContentMetadataKeyHierarchiesByStatementResponse
}
\description{
Gets a ContentMetadataKeyHierarchyPage of ContentMetadataKeyHierarchy objects that satisfy the given Statement query. The following fields are supported for filtering:  content metadata key hierarchies backend error 
\itemize{
  \item{id}
  \item{name}
  \item{status}
}
}

