% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreativeSetService.R
\name{dfp_getCreativeSetsByStatement}
\alias{dfp_getCreativeSetsByStatement}
\title{getCreativeSetsByStatement}
\usage{
dfp_getCreativeSetsByStatement(request_data, as_df=FALSE)
}
\arguments{
\item{request_data}{a \code{list} or \code{data.frame} of data elements
to be formatted for a SOAP request (XML format, but passed as character string)}

\item{as_df}{a boolean indicating whether to attempt to parse the result into a \code{data.frame}}
}
\value{
a \code{list} or \code{data.frame} containing all the elements of a getCreativeSetsByStatementResponse
}
\description{
Gets a CreativeSetPage of CreativeSet objects that satisfy the given Statement query. The following fields are supported for filtering:  a set of creative sets 
\itemize{
  \item{id}
  \item{name}
  \item{masterCreativeId}
  \item{lastModifiedDateTime}
}
}
\seealso{
\href{https://developers.google.com/doubleclick-publishers/docs/reference/v201508/CreativeSetService#getCreativeSetsByStatement}{Google Documentation for getCreativeSetsByStatement}
}

