% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UserService.R
\name{dfp_getCurrentUser}
\alias{dfp_getCurrentUser}
\title{getCurrentUser}
\usage{
dfp_getCurrentUser()
}
\value{
a \code{list} containing all the elements of a getCurrentUserResponse
}
\description{
Returns the current User.
}

