% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CustomFieldService.R
\name{dfp_getCustomFieldOption}
\alias{dfp_getCustomFieldOption}
\title{getCustomFieldOption}
\usage{
dfp_getCustomFieldOption()
}
\value{
a \code{list} containing all the elements of a getCustomFieldOptionResponse
}
\description{
Returns the CustomFieldOption uniquely identified by the given ID.
}
\seealso{
\href{https://developers.google.com/doubleclick-publishers/docs/reference/v201508/CustomFieldService#getCustomFieldOption}{Google Documentation for getCustomFieldOption}
}

