% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ForecastService.R
\name{dfp_getDeliveryForecast}
\alias{dfp_getDeliveryForecast}
\title{getDeliveryForecast}
\usage{
dfp_getDeliveryForecast()
}
\value{
a \code{list} containing all the elements of a getDeliveryForecastResponse
}
\description{
Gets the delivery forecast for a list of ProspectiveLineItem objects in a single delivery simulation with line items potentially contending with each other. A delivery forecast reports the number of units that will be delivered to each line item given the line item goals and contentions from other line items.
}

