% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ForecastService.R
\name{dfp_getDeliveryForecastByIds}
\alias{dfp_getDeliveryForecastByIds}
\title{getDeliveryForecastByIds}
\usage{
dfp_getDeliveryForecastByIds()
}
\value{
a \code{list} containing all the elements of a getDeliveryForecastByIdsResponse
}
\description{
Gets the delivery forecast for a list of existing LineItem objects in a single delivery simulation with line items potentially contending with each other. A delivery forecast reports the number of units that will be delivered to each line item given the line item goals and contentions from other line items.
}

