% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExchangeRateService.R
\name{dfp_getExchangeRatesByStatement}
\alias{dfp_getExchangeRatesByStatement}
\title{getExchangeRatesByStatement}
\usage{
dfp_getExchangeRatesByStatement(request_data)
}
\arguments{
\item{request_data}{a \code{list} or \code{data.frame} of data elements
to be formatted for a SOAP request (XML format, but passed as character string)#' @return a \code{list} containing all the elements of a getExchangeRatesByStatementResponse}
}
\description{
Gets a ExchangeRatePage of ExchangeRate objects that satisfy the given Statement query. The following fields are supported for filtering:  a set of exchange rates 
\itemize{
  \item{id}
  \item{currencyCode}
  \item{refreshRate}
  \item{direction}
  \item{exchangeRate}
}
}

