% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LabelService.R
\name{dfp_getLabelsByStatement}
\alias{dfp_getLabelsByStatement}
\title{getLabelsByStatement}
\usage{
dfp_getLabelsByStatement(request_data)
}
\arguments{
\item{request_data}{a \code{list} or \code{data.frame} of data elements
to be formatted for a SOAP request (XML format, but passed as character string)#' @return a \code{list} containing all the elements of a getLabelsByStatementResponse}
}
\description{
Gets a LabelPage of Label objects that satisfy the given Statement query. The following fields are supported for filtering:  a set of labels. 
\itemize{
  \item{id}
  \item{type}
  \item{name}
  \item{description}
  \item{isActive}
}
}

