% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LineItemCreativeAssociationService.R
\name{dfp_getLineItemCreativeAssociationsByStatement}
\alias{dfp_getLineItemCreativeAssociationsByStatement}
\title{getLineItemCreativeAssociationsByStatement}
\usage{
dfp_getLineItemCreativeAssociationsByStatement(request_data)
}
\arguments{
\item{request_data}{a \code{list} or \code{data.frame} of data elements
to be formatted for a SOAP request (XML format, but passed as character string)}
}
\value{
a \code{list} containing all the elements of a getLineItemCreativeAssociationsByStatementResponse
}
\description{
Gets a LineItemCreativeAssociationPage of LineItemCreativeAssociation objects that satisfy the given Statement query. The following fields are supported for filtering:  a set of line item creative associations 
\itemize{
  \item{creativeId}
  \item{manualCreativeRotationWeight}
  \item{destinationUrl}
  \item{lineItemId}
  \item{status}
  \item{lastModifiedDateTime}
}
}

