% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LiveStreamEventService.R
\name{dfp_getLiveStreamEventsByStatement}
\alias{dfp_getLiveStreamEventsByStatement}
\title{getLiveStreamEventsByStatement}
\usage{
dfp_getLiveStreamEventsByStatement(request_data)
}
\arguments{
\item{request_data}{a \code{list} or \code{data.frame} of data elements
to be formatted for a SOAP request (XML format, but passed as character string)#' @return a \code{list} containing all the elements of a getLiveStreamEventsByStatementResponse}
}
\description{
Gets a LiveStreamEventPage of LiveStreamEvent objects that satisfy the given Statement query. The following fields are supported for filtering:  list of live stream events 
\itemize{
  \item{id}
}
}

