% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PackageService.R
\name{dfp_getPackagesByStatement}
\alias{dfp_getPackagesByStatement}
\title{getPackagesByStatement}
\usage{
dfp_getPackagesByStatement(request_data, as_df = TRUE, verbose = FALSE)
}
\arguments{
\item{request_data}{a \code{list} or \code{data.frame} of data elements
to be formatted for a SOAP
request (XML format, but passed as character string)}

\item{as_df}{a boolean indicating whether to attempt to parse the result into
a \code{data.frame}}

\item{verbose}{a boolean indicating whether to print the service URL and POSTed XML}
}
\value{
a \code{data.frame} or \code{list} containing all the elements of a getPackagesByStatementResponse
}
\description{
Gets a PackagePage of Package objects that satisfy the given Statement query. The following fields are supported for filtering: 
\itemize{
  \item{id}
  \item{name}
  \item{proposalId}
  \item{productPackageId}
  \item{isArchived}
  \item{lastModifiedDateTime}
}
}
\examples{
\dontrun{
 dat <- list(filterStatement=list('query'="WHERE status='ACTIVE'")) 
 res <- dfp_getPackagesByStatement(dat)
}
}
\seealso{
\href{https://developers.google.com/ad-manager/api/reference/v201811/PackageService#getPackagesByStatement}{Google Documentation for getPackagesByStatement}
}
