% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlacementService.R
\name{dfp_getPlacementsByStatement}
\alias{dfp_getPlacementsByStatement}
\title{getPlacementsByStatement}
\usage{
dfp_getPlacementsByStatement(request_data)
}
\arguments{
\item{request_data}{a \code{list} or \code{data.frame} of data elements
to be formatted for a SOAP request (XML format, but passed as character string)#' @return a \code{list} containing all the elements of a getPlacementsByStatementResponse}
}
\description{
Gets a PlacementPage of Placement objects that satisfy the given Statement query. The following fields are supported for filtering:  a set of placements 
\itemize{
  \item{description}
  \item{id}
  \item{isAdSenseTargetingEnabled}
  \item{name}
  \item{placementCode}
  \item{status}
  \item{lastModifiedDateTime}
}
}

