% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PremiumRateService.R
\name{dfp_getPremiumRatesByStatement}
\alias{dfp_getPremiumRatesByStatement}
\title{getPremiumRatesByStatement}
\usage{
dfp_getPremiumRatesByStatement(request_data)
}
\arguments{
\item{request_data}{a \code{list} or \code{data.frame} of data elements
to be formatted for a SOAP request (XML format, but passed as character string)}
}
\value{
a \code{list} containing all the elements of a getPremiumRatesByStatementResponse
}
\description{
Gets a PremiumRatePage of PremiumRate objects that satisfy the given Statement query. The following fields are supported for filtering:  list of premium rates. 
\itemize{
  \item{id}
  \item{rateCardId}
  \item{pricingMethod}
}
}

