% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LineItemCreativeAssociationService.R
\name{dfp_getPreviewUrl}
\alias{dfp_getPreviewUrl}
\title{getPreviewUrl}
\usage{
dfp_getPreviewUrl()
}
\value{
a \code{list} containing all the elements of a getPreviewUrlResponse
}
\description{
Returns an insite preview URL that references the specified site URL with the specified creative from the association served to it. For Creative Set previewing you may specify the master creative Id. creative served to it
}

