% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LineItemCreativeAssociationService.R
\name{dfp_getPreviewUrlsForNativeStyles}
\alias{dfp_getPreviewUrlsForNativeStyles}
\title{getPreviewUrlsForNativeStyles}
\usage{
dfp_getPreviewUrlsForNativeStyles(as_df = TRUE, verbose = FALSE)
}
\arguments{
\item{as_df}{a boolean indicating whether to attempt to parse the result into
a \code{data.frame}}

\item{verbose}{a boolean indicating whether to print the service URL and POSTed XML}
}
\value{
a \code{data.frame} or \code{list} containing all the elements of a getPreviewUrlsForNativeStylesResponse
}
\description{
Returns a list of URLs that reference the specified site URL with the specified creative from the association served to it. For Creative Set previewing you may specify the master creative Id. Each URL corresponds to one available native style for previewing the specified creative. creative
}
\examples{
\dontrun{
 res <- dfp_getPreviewUrlsForNativeStyles()
}
}
\seealso{
\href{https://developers.google.com/ad-manager/api/reference/v201905/LineItemCreativeAssociationService#getPreviewUrlsForNativeStyles}{Google Documentation for getPreviewUrlsForNativeStyles}
}
