% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProductService.R
\name{dfp_getProductsByStatement}
\alias{dfp_getProductsByStatement}
\title{getProductsByStatement}
\usage{
dfp_getProductsByStatement(request_data)
}
\arguments{
\item{request_data}{a \code{list} or \code{data.frame} of data elements
to be formatted for a SOAP request (XML format, but passed as character string)}
}
\value{
a \code{list} containing all the elements of a getProductsByStatementResponse
}
\description{
Gets a ProductPage of Product objects that satisfy the criteria specified by given Statement query. The following fields are supported for filtering and/or sorting:  criteria over products 
\itemize{
  \item{rateCardId}
  \item{status}
  \item{lineItemType}
  \item{rateType}
  \item{productTemplateId}
  \item{name}
  \item{description}
  \item{id}
  \item{lastModifiedDateTime}
}
}

