% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProductTemplateService.R
\name{dfp_getProductTemplatesByStatement}
\alias{dfp_getProductTemplatesByStatement}
\title{getProductTemplatesByStatement}
\usage{
dfp_getProductTemplatesByStatement(request_data, as_df=FALSE)
}
\arguments{
\item{request_data}{a \code{list} or \code{data.frame} of data elements
to be formatted for a SOAP request (XML format, but passed as character string)}

\item{as_df}{a boolean indicating whether to attempt to parse the result into a \code{data.frame}}
}
\value{
a \code{list} or \code{data.frame} containing all the elements of a getProductTemplatesByStatementResponse
}
\description{
Gets a ProductTemplatePage of ProductTemplate objects that satisfy the filtering criteria specified by given Statement query. The following fields are supported for filtering:  filtering criteria over productTemplates 
\itemize{
  \item{id}
  \item{name}
  \item{nameMacro}
  \item{description}
  \item{status}
  \item{lastModifiedDateTime}
  \item{lineItemType}
  \item{productType}
  \item{rateType}
}
}
\seealso{
\href{https://developers.google.com/doubleclick-publishers/docs/reference/v201508/ProductTemplateService#getProductTemplatesByStatement}{Google Documentation for getProductTemplatesByStatement}
}

