% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProposalLineItemService.R
\name{dfp_getProposalLineItemsByStatement}
\alias{dfp_getProposalLineItemsByStatement}
\title{getProposalLineItemsByStatement}
\usage{
dfp_getProposalLineItemsByStatement(request_data)
}
\arguments{
\item{request_data}{a \code{list} or \code{data.frame} of data elements
to be formatted for a SOAP request (XML format, but passed as character string)#' @return a \code{list} containing all the elements of a getProposalLineItemsByStatementResponse}
}
\description{
Gets a ProposalLineItemPage of ProposalLineItem objects that satisfy the given Statement query. The following fields are supported for filtering:  a set of proposal line items 
\itemize{
  \item{id}
  \item{name}
  \item{proposalId}
  \item{isArchived}
  \item{lastModifiedDateTime}
  \item{useThirdPartyAdServerFromProposal}
  \item{thirdPartyAdServerId}
  \item{customThirdPartyAdServerName}
}
}

