% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProposalService.R
\name{dfp_getProposalsByStatement}
\alias{dfp_getProposalsByStatement}
\title{getProposalsByStatement}
\usage{
dfp_getProposalsByStatement(request_data)
}
\arguments{
\item{request_data}{a \code{list} or \code{data.frame} of data elements
to be formatted for a SOAP request (XML format, but passed as character string)#' @return a \code{list} containing all the elements of a getProposalsByStatementResponse}
}
\description{
Gets a ProposalPage of Proposal objects that satisfy the given Statement query. The following fields are supported for filtering:  a set of proposals 
\itemize{
  \item{id}
  \item{dfpOrderId}
  \item{name}
  \item{status}
  \item{isArchived}
  \item{approvalStatus}
  \item{lastModifiedDateTime}
  \item{thirdPartyAdServerId}
  \item{customThirdPartyAdServerName}
}
}

