% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReconciliationOrderReportService.R
\name{dfp_getReconciliationOrderReportsByStatement}
\alias{dfp_getReconciliationOrderReportsByStatement}
\title{getReconciliationOrderReportsByStatement}
\usage{
dfp_getReconciliationOrderReportsByStatement(request_data)
}
\arguments{
\item{request_data}{a \code{list} or \code{data.frame} of data elements
to be formatted for a SOAP request (XML format, but passed as character string)#' @return a \code{list} containing all the elements of a getReconciliationOrderReportsByStatementResponse}
}
\description{
Gets an ReconciliationOrderReportPage of ReconciliationOrderReport objects that satisfy the given Statement query. The following fields are supported for filtering:  The reconciliationReportId field is required and can only be combined with an AND to other conditions. Furthermore, the results may only belong to one ReconciliationReport. a set of reconciliation order reports. 
\itemize{
  \item{reconciliationReportId}
  \item{id}
  \item{orderId}
  \item{proposalId}
  \item{status}
  \item{submissionDateTime}
  \item{submitterId}
}
}

