% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReconciliationReportService.R
\name{dfp_getReconciliationReportsByStatement}
\alias{dfp_getReconciliationReportsByStatement}
\title{getReconciliationReportsByStatement}
\usage{
dfp_getReconciliationReportsByStatement(request_data)
}
\arguments{
\item{request_data}{a \code{list} or \code{data.frame} of data elements
to be formatted for a SOAP request (XML format, but passed as character string)}
}
\value{
a \code{list} containing all the elements of a getReconciliationReportsByStatementResponse
}
\description{
Gets an ReconciliationReportPage of ReconciliationReport objects that satisfy the given Statement query. The following fields are supported for filtering.  filter a set of reconciliation reports 
\itemize{
  \item{id}
  \item{status}
  \item{startDate}
}
}

