% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReportService.R
\name{dfp_getReportDownloadUrlWithOptions}
\alias{dfp_getReportDownloadUrlWithOptions}
\title{getReportDownloadUrlWithOptions}
\usage{
dfp_getReportDownloadUrlWithOptions()
}
\value{
a \code{list} containing all the elements of a getReportDownloadUrlWithOptionsResponse
}
\description{
Returns the URL at which the report file can be downloaded, and allows for customization of the downloaded report. <p> By default, the report will be generated as a gzip archive, containing the report file itself. This can be changed by setting ReportDownloadOptions useGzipCompression to false.
}
\seealso{
\href{https://developers.google.com/doubleclick-publishers/docs/reference/v201508/ReportService#getReportDownloadUrlWithOptions}{Google Documentation for getReportDownloadUrlWithOptions}
}

