% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReportService.R
\name{dfp_getReportJobStatus}
\alias{dfp_getReportJobStatus}
\title{getReportJobStatus}
\usage{
dfp_getReportJobStatus(request_data)
}
\arguments{
\item{request_data}{a \code{list} or \code{data.frame} of data elements
to be formatted for a SOAP request (XML format, but passed as character string)}
}
\value{
a \code{list} containing all the elements of a getReportJobStatusResponse
}
\description{
Returns the ReportJobStatus of the report job with the specified ID.
}
\seealso{
\href{https://developers.google.com/doubleclick-publishers/docs/reference/v201508/ReportService#getReportJobStatus}{Google Documentation for getReportJobStatus}
}

