% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SharedAdUnitService.R
\name{dfp_getSharedAdUnitsByStatement}
\alias{dfp_getSharedAdUnitsByStatement}
\title{getSharedAdUnitsByStatement}
\usage{
dfp_getSharedAdUnitsByStatement(request_data, as_df=TRUE, verbose=FALSE)
}
\arguments{
\item{request_data}{a \code{list} or \code{data.frame} of data elements
to be formatted for a SOAP request (XML format, but passed as character string)}

\item{as_df}{a boolean indicating whether to attempt to parse the result into a \code{data.frame}}

\item{verbose}{a boolean indicating whether to print the service URL and POSTed XML}
}
\value{
a \code{data.frame} or \code{list} containing all the elements of a getSharedAdUnitsByStatementResponse
}
\description{
Gets a SharedAdUnitPage of SharedAdUnit objects that satisfy the given Statement query. The following fields are supported for filtering:  a set of shared ad units 
\itemize{
  \item{id}
  \item{name}
}
}
\seealso{
\href{https://developers.google.com/doubleclick-publishers/docs/reference/v201508/SharedAdUnitService#getSharedAdUnitsByStatement}{Google Documentation for getSharedAdUnitsByStatement}
}

