% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SuggestedAdUnitService.R
\name{dfp_getSuggestedAdUnitsByStatement}
\alias{dfp_getSuggestedAdUnitsByStatement}
\title{getSuggestedAdUnitsByStatement}
\usage{
dfp_getSuggestedAdUnitsByStatement(request_data)
}
\arguments{
\item{request_data}{a \code{list} or \code{data.frame} of data elements
to be formatted for a SOAP request (XML format, but passed as character string)}
}
\value{
a \code{list} containing all the elements of a getSuggestedAdUnitsByStatementResponse
}
\description{
Gets a SuggestedAdUnitPage of SuggestedAdUnit objects that satisfy the filter query. There is a system-enforced limit of 1000 on the number of suggested ad units that are suggested at any one time.  <strong>Note:</strong> After API version 201311, the id field will only be numerical. a set of suggested ad units 
\itemize{
  \item{id}
  \item{numRequests}
}
}
\seealso{
\href{https://developers.google.com/doubleclick-publishers/docs/reference/v201508/SuggestedAdUnitService#getSuggestedAdUnitsByStatement}{Google Documentation for getSuggestedAdUnitsByStatement}
}

