% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UserService.R
\name{dfp_getUsersByStatement}
\alias{dfp_getUsersByStatement}
\title{getUsersByStatement}
\usage{
dfp_getUsersByStatement(request_data)
}
\arguments{
\item{request_data}{a \code{list} or \code{data.frame} of data elements
to be formatted for a SOAP request (XML format, but passed as character string)}
}
\value{
a \code{list} containing all the elements of a getUsersByStatementResponse
}
\description{
Gets a UserPage of User objects that satisfy the given Statement query. The following fields are supported for filtering:  a set of users 
\itemize{
  \item{email}
  \item{id}
  \item{name}
  \item{roleId}
  \item{rolename}
  \item{status}
}
}

