% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UserTeamAssociationService.R
\name{dfp_getUserTeamAssociationsByStatement}
\alias{dfp_getUserTeamAssociationsByStatement}
\title{getUserTeamAssociationsByStatement}
\usage{
dfp_getUserTeamAssociationsByStatement(request_data)
}
\arguments{
\item{request_data}{a \code{list} or \code{data.frame} of data elements
to be formatted for a SOAP request (XML format, but passed as character string)}
}
\value{
a \code{list} containing all the elements of a getUserTeamAssociationsByStatementResponse
}
\description{
Gets a UserTeamAssociationPage of UserTeamAssociation objects that satisfy the given Statement query. The following fields are supported for filtering:  a set of user team associations 
\itemize{
  \item{userId}
  \item{teamId}
}
}

