% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WorkflowRequestService.R
\name{dfp_getWorkflowRequestsByStatement}
\alias{dfp_getWorkflowRequestsByStatement}
\title{getWorkflowRequestsByStatement}
\usage{
dfp_getWorkflowRequestsByStatement(request_data, as_df=TRUE, verbose=FALSE)
}
\arguments{
\item{request_data}{a \code{list} or \code{data.frame} of data elements
to be formatted for a SOAP request (XML format, but passed as character string)}

\item{as_df}{a boolean indicating whether to attempt to parse the result into a \code{data.frame}}

\item{verbose}{a boolean indicating whether to print the service URL and POSTed XML}
}
\value{
a \code{data.frame} or \code{list} containing all the elements of a getWorkflowRequestsByStatementResponse
}
\description{
Gets a list of WorkflowRequest objects that satisfy the given Statement query. The following fields are supported for filtering:  Starting in V201405 a type filter must be used 
\itemize{
  \item{id}
  \item{workflowRuleName}
  \item{entityType}
  \item{entityId}
  \item{approvalStatus}
  \item{conditionStatus}
  \item{type}
}
}
\seealso{
\href{https://developers.google.com/doubleclick-publishers/docs/reference/v201508/WorkflowRequestService#getWorkflowRequestsByStatement}{Google Documentation for getWorkflowRequestsByStatement}
}

