% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LineItemService.R
\name{dfp_hasCustomPacingCurve}
\alias{dfp_hasCustomPacingCurve}
\title{hasCustomPacingCurve}
\usage{
dfp_hasCustomPacingCurve(as_df = TRUE, verbose = FALSE)
}
\arguments{
\item{as_df}{a boolean indicating whether to attempt to parse the result into
a \code{data.frame}}

\item{verbose}{a boolean indicating whether to print the service URL and POSTed XML}
}
\value{
a \code{data.frame} or \code{list} containing all the elements of a hasCustomPacingCurveResponse
}
\description{
Returns whether a custom pacing curve has been uploaded to Google Cloud Storage for a line item. Returns whether a custom pacing curve has been uploaded to Google Cloud Storage for a line item. @param lineItemId the ID of the line item Returns whether a custom pacing curve has been uploaded to Google Cloud Storage for a line item. @param lineItemId the ID of the line item
}
\examples{
\dontrun{
 res <- dfp_hasCustomPacingCurve()
}
}
\seealso{
\href{https://developers.google.com/ad-manager/api/reference/v201905/LineItemService#hasCustomPacingCurve}{Google Documentation for hasCustomPacingCurve}
}
