% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AdExclusionRuleService.R
\name{dfp_performAdExclusionRuleAction}
\alias{dfp_performAdExclusionRuleAction}
\title{performAdExclusionRuleAction}
\usage{
dfp_performAdExclusionRuleAction()
}
\value{
a \code{list} containing all the elements of a performAdExclusionRuleActionResponse
}
\description{
Performs action on AdExclusionRule objects that satisfy the given Statement query. a set of ad exclusion rules
}

