% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AdRuleService.R
\name{dfp_performAdRuleAction}
\alias{dfp_performAdRuleAction}
\title{performAdRuleAction}
\usage{
dfp_performAdRuleAction()
}
\value{
a \code{list} containing all the elements of a performAdRuleActionResponse
}
\description{
Performs actions on AdRule objects that match the given Statement query. a set of ad rules
}

