% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InventoryService.R
\name{dfp_performAdUnitAction}
\alias{dfp_performAdUnitAction}
\title{performAdUnitAction}
\usage{
dfp_performAdUnitAction()
}
\value{
a \code{list} containing all the elements of a performAdUnitActionResponse
}
\description{
Performs actions on AdUnit objects that match the given Statement query. a set of ad units
}

