% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BaseRateService.R
\name{dfp_performBaseRateAction}
\alias{dfp_performBaseRateAction}
\title{performBaseRateAction}
\usage{
dfp_performBaseRateAction()
}
\value{
a \code{list} containing all the elements of a performBaseRateActionResponse
}
\description{
Performs actions on BaseRate objects that satisfy the given Statement query. a set of base rates.
}

