% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ContentBundleService.R
\name{dfp_performContentBundleAction}
\alias{dfp_performContentBundleAction}
\title{performContentBundleAction}
\usage{
dfp_performContentBundleAction()
}
\value{
a \code{list} containing all the elements of a performContentBundleActionResponse
}
\description{
Performs actions on ContentBundle objects that match the given Statement query. a set of content bundles
}
\seealso{
\href{https://developers.google.com/doubleclick-publishers/docs/reference/v201508/ContentBundleService#performContentBundleAction}{Google Documentation for performContentBundleAction}
}

