% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreativeWrapperService.R
\name{dfp_performCreativeWrapperAction}
\alias{dfp_performCreativeWrapperAction}
\title{performCreativeWrapperAction}
\usage{
dfp_performCreativeWrapperAction()
}
\value{
a \code{list} containing all the elements of a performCreativeWrapperActionResponse
}
\description{
Performs actions on CreativeWrapper objects that match the given Statement query. a set of labels
}

