% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CustomFieldService.R
\name{dfp_performCustomFieldAction}
\alias{dfp_performCustomFieldAction}
\title{performCustomFieldAction}
\usage{
dfp_performCustomFieldAction()
}
\value{
a \code{list} containing all the elements of a performCustomFieldActionResponse
}
\description{
Performs actions on CustomField objects that match the given Statement query. a set of custom fields
}

