% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExchangeRateService.R
\name{dfp_performExchangeRateAction}
\alias{dfp_performExchangeRateAction}
\title{performExchangeRateAction}
\usage{
dfp_performExchangeRateAction()
}
\value{
a \code{list} containing all the elements of a performExchangeRateActionResponse
}
\description{
Performs an action on ExchangeRate objects that satisfy the given Statement query. The following fields are supported for filtering:  a set of exchange rates 
\itemize{
  \item{id}
  \item{currencyCode}
  \item{refreshRate}
  \item{direction}
  \item{exchangeRate}
}
}
\seealso{
\href{https://developers.google.com/doubleclick-publishers/docs/reference/v201508/ExchangeRateService#performExchangeRateAction}{Google Documentation for performExchangeRateAction}
}

