% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LabelService.R
\name{dfp_performLabelAction}
\alias{dfp_performLabelAction}
\title{performLabelAction}
\usage{
dfp_performLabelAction()
}
\value{
a \code{list} containing all the elements of a performLabelActionResponse
}
\description{
Performs actions on Label objects that match the given Statement query. a set of labels
}
\seealso{
\href{https://developers.google.com/doubleclick-publishers/docs/reference/v201508/LabelService#performLabelAction}{Google Documentation for performLabelAction}
}

