% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LineItemService.R
\name{dfp_performLineItemAction}
\alias{dfp_performLineItemAction}
\title{performLineItemAction}
\usage{
dfp_performLineItemAction()
}
\value{
a \code{list} containing all the elements of a performLineItemActionResponse
}
\description{
Performs actions on LineItem objects that match the given Statement query. a set of line items
}
\seealso{
\href{https://developers.google.com/doubleclick-publishers/docs/reference/v201508/LineItemService#performLineItemAction}{Google Documentation for performLineItemAction}
}

