% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LineItemCreativeAssociationService.R
\name{dfp_performLineItemCreativeAssociationAction}
\alias{dfp_performLineItemCreativeAssociationAction}
\title{performLineItemCreativeAssociationAction}
\usage{
dfp_performLineItemCreativeAssociationAction()
}
\value{
a \code{list} containing all the elements of a performLineItemCreativeAssociationActionResponse
}
\description{
Performs actions on LineItemCreativeAssociation objects that match the given Statement query. a set of line item creative associations
}
\seealso{
\href{https://developers.google.com/doubleclick-publishers/docs/reference/v201508/LineItemCreativeAssociationService#performLineItemCreativeAssociationAction}{Google Documentation for performLineItemCreativeAssociationAction}
}

