% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OrderService.R
\name{dfp_performOrderAction}
\alias{dfp_performOrderAction}
\title{performOrderAction}
\usage{
dfp_performOrderAction()
}
\value{
a \code{list} containing all the elements of a performOrderActionResponse
}
\description{
Performs actions on Order objects that match the given Statement query. a set of orders
}
\seealso{
\href{https://developers.google.com/doubleclick-publishers/docs/reference/v201508/OrderService#performOrderAction}{Google Documentation for performOrderAction}
}

