% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PackageService.R
\name{dfp_performPackageAction}
\alias{dfp_performPackageAction}
\title{performPackageAction}
\usage{
dfp_performPackageAction()
}
\value{
a \code{list} containing all the elements of a performPackageActionResponse
}
\description{
Performs actions on Package objects that match the given Statement.
}
\seealso{
\href{https://developers.google.com/doubleclick-publishers/docs/reference/v201508/PackageService#performPackageAction}{Google Documentation for performPackageAction}
}

