% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlacementService.R
\name{dfp_performPlacementAction}
\alias{dfp_performPlacementAction}
\title{performPlacementAction}
\usage{
dfp_performPlacementAction(request_data)
}
\arguments{
\item{request_data}{a \code{list} or \code{data.frame} of data elements
to be formatted for a SOAP request (XML format, but passed as character string)}
}
\value{
a \code{list} containing all the elements of a performPlacementActionResponse
}
\description{
Performs actions on Placement objects that match the given Statement query. a set of placements
}
\seealso{
\href{https://developers.google.com/doubleclick-publishers/docs/reference/v201508/PlacementService#performPlacementAction}{Google Documentation for performPlacementAction}
}

