% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProductService.R
\name{dfp_performProductAction}
\alias{dfp_performProductAction}
\title{performProductAction}
\usage{
dfp_performProductAction(request_data)
}
\arguments{
\item{request_data}{a \code{list} or \code{data.frame} of data elements
to be formatted for a SOAP request (XML format, but passed as character string)}
}
\value{
a \code{list} containing all the elements of a performProductActionResponse
}
\description{
Performs action on Product objects that satisfy the given Statement.
}
\seealso{
\href{https://developers.google.com/doubleclick-publishers/docs/reference/v201508/ProductService#performProductAction}{Google Documentation for performProductAction}
}

