% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProductPackageService.R
\name{dfp_performProductPackageAction}
\alias{dfp_performProductPackageAction}
\title{performProductPackageAction}
\usage{
dfp_performProductPackageAction()
}
\value{
a \code{list} containing all the elements of a performProductPackageActionResponse
}
\description{
Performs actions on ProductPackage objects that match the given Statement query.
}
\seealso{
\href{https://developers.google.com/doubleclick-publishers/docs/reference/v201508/ProductPackageService#performProductPackageAction}{Google Documentation for performProductPackageAction}
}

