% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProductPackageItemService.R
\name{dfp_performProductPackageItemAction}
\alias{dfp_performProductPackageItemAction}
\title{performProductPackageItemAction}
\usage{
dfp_performProductPackageItemAction(request_data, as_df=FALSE)
}
\arguments{
\item{request_data}{a \code{list} or \code{data.frame} of data elements
to be formatted for a SOAP request (XML format, but passed as character string)}

\item{as_df}{a boolean indicating whether to attempt to parse the result into a \code{data.frame}}
}
\value{
a \code{list} or \code{data.frame} containing all the elements of a performProductPackageItemActionResponse
}
\description{
Performs actions on ProductPackageItem objects that satisfy the given Statement query. a set of product package items
}
\seealso{
\href{https://developers.google.com/doubleclick-publishers/docs/reference/v201508/ProductPackageItemService#performProductPackageItemAction}{Google Documentation for performProductPackageItemAction}
}

