% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProposalLineItemService.R
\name{dfp_performProposalLineItemAction}
\alias{dfp_performProposalLineItemAction}
\title{performProposalLineItemAction}
\usage{
dfp_performProposalLineItemAction()
}
\value{
a \code{list} containing all the elements of a performProposalLineItemActionResponse
}
\description{
Performs actions on ProposalLineItem objects that match the given Statement query. proposal line items
}
\seealso{
\href{https://developers.google.com/doubleclick-publishers/docs/reference/v201508/ProposalLineItemService#performProposalLineItemAction}{Google Documentation for performProposalLineItemAction}
}

