% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RateCardService.R
\name{dfp_performRateCardAction}
\alias{dfp_performRateCardAction}
\title{performRateCardAction}
\usage{
dfp_performRateCardAction(request_data)
}
\arguments{
\item{request_data}{a \code{list} or \code{data.frame} of data elements
to be formatted for a SOAP request (XML format, but passed as character string)}
}
\value{
a \code{list} containing all the elements of a performRateCardActionResponse
}
\description{
Performs action on RateCard objects that satisfy the given Statement query. a set of rate cards.
}
\seealso{
\href{https://developers.google.com/doubleclick-publishers/docs/reference/v201508/RateCardService#performRateCardAction}{Google Documentation for performRateCardAction}
}

