% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReconciliationOrderReportService.R
\name{dfp_performReconciliationOrderReportAction}
\alias{dfp_performReconciliationOrderReportAction}
\title{performReconciliationOrderReportAction}
\usage{
dfp_performReconciliationOrderReportAction()
}
\value{
a \code{list} containing all the elements of a performReconciliationOrderReportActionResponse
}
\description{
Performs actions on the ReconciliationOrderReport objects that match the given Statement query. The following fields are supported for filtering:  The following statement patterns are supported:  reconciliationReportId = :reconciliationReportId AND orderId = :orderId reconciliationReportId = :reconciliationReportId AND proposalId = :proposalId reconciliationReportId = :reconciliationReportId AND (orderId IN (...) OR proposalId IN (...))  The IN clause could be expanded to multiple OR expressions like (orderId = :orderId OR orderId = :orderId OR ...) Only orders to which the API user has access will be included. one reconciliation report. 
\itemize{
  \item{orderId}
  \item{proposalId}
  \item{reconciliationReportId}
}
}
\seealso{
\href{https://developers.google.com/doubleclick-publishers/docs/reference/v201508/ReconciliationOrderReportService#performReconciliationOrderReportAction}{Google Documentation for performReconciliationOrderReportAction}
}

