% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SharedAdUnitService.R
\name{dfp_performSharedAdUnitAction}
\alias{dfp_performSharedAdUnitAction}
\title{performSharedAdUnitAction}
\usage{
dfp_performSharedAdUnitAction()
}
\value{
a \code{list} containing all the elements of a performSharedAdUnitActionResponse
}
\description{
Performs actions on shared ad unit objects that match the given Statement query. a set of shared ad units on which the action was performed.
}

