% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SuggestedAdUnitService.R
\name{dfp_performSuggestedAdUnitAction}
\alias{dfp_performSuggestedAdUnitAction}
\title{performSuggestedAdUnitAction}
\usage{
dfp_performSuggestedAdUnitAction(request_data, as_df=TRUE)
}
\arguments{
\item{request_data}{a \code{list} or \code{data.frame} of data elements
to be formatted for a SOAP request (XML format, but passed as character string)}

\item{as_df}{a boolean indicating whether to attempt to parse the result into a \code{data.frame}}
}
\value{
a \code{data.frame} or \code{list} containing all the elements of a performSuggestedAdUnitActionResponse
}
\description{
Performs actions on SuggestedAdUnit objects that match the given Statement query. The following fields are supported for filtering:  a set of suggested ad units 
\itemize{
  \item{id}
  \item{numRequests}
}
}
\seealso{
\href{https://developers.google.com/doubleclick-publishers/docs/reference/v201508/SuggestedAdUnitService#performSuggestedAdUnitAction}{Google Documentation for performSuggestedAdUnitAction}
}

