% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UserService.R
\name{dfp_performUserAction}
\alias{dfp_performUserAction}
\title{performUserAction}
\usage{
dfp_performUserAction()
}
\value{
a \code{list} containing all the elements of a performUserActionResponse
}
\description{
Performs actions on User objects that match the given Statement query. a set of users
}

