% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlacementService.R
\name{dfp_PlacementService_object_factory}
\alias{dfp_PlacementService_object_factory}
\title{PlacementService}
\usage{
dfp_PlacementService_object_factory(obj_type, obj_data)
}
\arguments{
\item{obj_type}{a string from one of the object types described above}

\item{obj_data}{a \code{list} of \code{lists} or a \code{data.frame} with the required components to create
the obj_type specified. All input names must match the expected attributes for the object.
If submitting a list of attributes to create 1 object make sure to nest it inside top-level list, so that
it is a list of length 1 with 1 list of attributes to create the object.}
}
\value{
a string of formatted XML that can be inserted into a SOAP request body to the APT
}
\description{
Provides methods for creating, updating and retrieving Placement objects.
}
\details{
You can use a placement to group ad units. For example, you might have a
placement that focuses on sports sites, which may be spread across
different branches of your inventory. You might also have a "fire sale"
placement that includes ad units that have not been selling and are
consequently priced very attractively.

PlacementService Object Factory

A function to create a variety of objects that are part of the PlacementService
Note: not all objects here are inputs to the functions of this service so 
please consult each function for the required inputs

Below is a list of objects and their required components to be created with this function:

\strong{ActivatePlacements}

The action used for activating Placement objects.
\href{https://developers.google.com/doubleclick-publishers/docs/reference/v201508/PlacementService.ActivatePlacements}{Google Documentation for ActivatePlacements}
\describe{
 \item{extends PlacementAction}{see documentation for PlacementAction}
}

\strong{ApiError}

The API error base class that provides details about an error that occurred
while processing a service request. The OGNL field path is provided for
parsers to identify the request data element that may have caused the error.
\href{https://developers.google.com/doubleclick-publishers/docs/reference/v201508/PlacementService.ApiError}{Google Documentation for ApiError}
\describe{
 \item{fieldPath}{a character - The OGNL field path to identify cause of error.}
 \item{trigger}{a character - The data that caused the error.}
 \item{errorString}{a character - A simple string representation of the error and reason.}
}

\strong{ApiException}

Exception class for holding a list of service errors.
\href{https://developers.google.com/doubleclick-publishers/docs/reference/v201508/PlacementService.ApiException}{Google Documentation for ApiException}
\describe{
 \item{message}{a character (inherited from ApplicationException) - Error message.}
 \item{errors}{a ApiError - List of errors.}
}

\strong{ApiVersionError}

Errors related to the usage of API versions.
\href{https://developers.google.com/doubleclick-publishers/docs/reference/v201508/PlacementService.ApiVersionError}{Google Documentation for ApiVersionError}
\describe{
 \item{extends ApiError}{see documentation for ApiError}
 \item{reason}{a ApiVersionError.Reason - This can take one of the following values: 
   \itemize{
     \item{UPDATE_TO_NEWER_VERSION - Indicates that the operation is not allowed in the version the request was
made in.}
     \item{UNKNOWN - The value returned if the actual value is not exposed by the requested API
version.}
   }
  }
}

\strong{ApplicationException}

Base class for exceptions.
\href{https://developers.google.com/doubleclick-publishers/docs/reference/v201508/PlacementService.ApplicationException}{Google Documentation for ApplicationException}
\describe{
 \item{message}{a character - Error message.}
}

\strong{ArchivePlacements}

The action used for archiving Placement objects.
\href{https://developers.google.com/doubleclick-publishers/docs/reference/v201508/PlacementService.ArchivePlacements}{Google Documentation for ArchivePlacements}
\describe{
 \item{extends PlacementAction}{see documentation for PlacementAction}
}

\strong{AuthenticationError}

An error for an exception that occurred when authenticating.
\href{https://developers.google.com/doubleclick-publishers/docs/reference/v201508/PlacementService.AuthenticationError}{Google Documentation for AuthenticationError}
\describe{
 \item{extends ApiError}{see documentation for ApiError}
 \item{reason}{a AuthenticationError.Reason - This can take one of the following values: 
   \itemize{
     \item{AMBIGUOUS_SOAP_REQUEST_HEADER - The SOAP message contains a request header with an ambiguous definition of
the authentication header fields. This means either the authToken and
oAuthToken fields were both null or both were specified. Exactly one value
should be specified with each request.}
     \item{INVALID_EMAIL - The login provided is invalid.}
     \item{AUTHENTICATION_FAILED - Tried to authenticate with provided information, but failed.}
     \item{INVALID_OAUTH_SIGNATURE - The OAuth provided is invalid.}
     \item{INVALID_SERVICE - The specified service to use was not recognized.}
     \item{MISSING_SOAP_REQUEST_HEADER - The SOAP message is missing a request header with an authToken and optional
networkCode.}
     \item{MISSING_AUTHENTICATION_HTTP_HEADER - The HTTP request is missing a request header with an authToken}
     \item{MISSING_AUTHENTICATION - The request is missing an authToken}
     \item{NOT_WHITELISTED_FOR_API_ACCESS - The customer is not whitelisted for API access.}
     \item{NO_NETWORKS_TO_ACCESS - The user is not associated with any network.}
     \item{NETWORK_NOT_FOUND - No network for the given networkCode was found.}
     \item{NETWORK_CODE_REQUIRED - The user has access to more than one network, but did not provide a
networkCode.}
     \item{CONNECTION_ERROR - An error happened on the server side during connection to authentication
service.}
     \item{GOOGLE_ACCOUNT_ALREADY_ASSOCIATED_WITH_NETWORK - The user tried to create a test network using an account that already is
associated with a network.}
     \item{UNKNOWN - The value returned if the actual value is not exposed by the requested API
version.}
   }
  }
}

\strong{BooleanValue}

Contains a boolean value.
\href{https://developers.google.com/doubleclick-publishers/docs/reference/v201508/PlacementService.BooleanValue}{Google Documentation for BooleanValue}
\describe{
 \item{extends Value}{see documentation for Value}
 \item{value}{a boolean - The boolean value.}
}

\strong{CollectionSizeError}

Error for the size of the collection being too large
\href{https://developers.google.com/doubleclick-publishers/docs/reference/v201508/PlacementService.CollectionSizeError}{Google Documentation for CollectionSizeError}
\describe{
 \item{extends ApiError}{see documentation for ApiError}
 \item{reason}{a CollectionSizeError.Reason - This can take one of the following values: 
   \itemize{
     \item{TOO_LARGE}
     \item{UNKNOWN - The value returned if the actual value is not exposed by the requested API
version.}
   }
  }
}

\strong{CommonError}

A place for common errors that can be used across services.
\href{https://developers.google.com/doubleclick-publishers/docs/reference/v201508/PlacementService.CommonError}{Google Documentation for CommonError}
\describe{
 \item{extends ApiError}{see documentation for ApiError}
 \item{reason}{a CommonError.Reason - This can take one of the following values: 
   \itemize{
     \item{NOT_FOUND - Indicates that an attempt was made to retrieve an entity that does not
exist.}
     \item{ALREADY_EXISTS - Indicates that an attempt was made to create an entity that already exists.}
     \item{DUPLICATE_OBJECT - Indicates that two elements in the collection were identical.}
     \item{CANNOT_UPDATE - Indicates that an attempt was made to change an immutable field.}
     \item{CONCURRENT_MODIFICATION - Indicates that another request attempted to update the same data in the same
network at about the same time. Please wait and try the request again.}
     \item{UNKNOWN - The value returned if the actual value is not exposed by the requested API
version.}
   }
  }
}

\strong{Date}

Represents a date.
\href{https://developers.google.com/doubleclick-publishers/docs/reference/v201508/PlacementService.Date}{Google Documentation for Date}
\describe{
 \item{year}{a integer - Year (e.g., 2009)}
 \item{month}{a integer - Month (1..12)}
 \item{day}{a integer - Day (1..31)}
}

\strong{DateTime}

Represents a date combined with the time of day.
\href{https://developers.google.com/doubleclick-publishers/docs/reference/v201508/PlacementService.DateTime}{Google Documentation for DateTime}
\describe{
 \item{date}{a Date}
 \item{hour}{a integer}
 \item{minute}{a integer}
 \item{second}{a integer}
 \item{timeZoneID}{a character}
}

\strong{DateTimeValue}

Contains a date-time value.
\href{https://developers.google.com/doubleclick-publishers/docs/reference/v201508/PlacementService.DateTimeValue}{Google Documentation for DateTimeValue}
\describe{
 \item{extends Value}{see documentation for Value}
 \item{value}{a DateTime - The DateTime value.}
}

\strong{DateValue}

Contains a date value.
\href{https://developers.google.com/doubleclick-publishers/docs/reference/v201508/PlacementService.DateValue}{Google Documentation for DateValue}
\describe{
 \item{extends Value}{see documentation for Value}
 \item{value}{a Date - The Date value.}
}

\strong{DeactivatePlacements}

The action used for deactivating Placement objects.
\href{https://developers.google.com/doubleclick-publishers/docs/reference/v201508/PlacementService.DeactivatePlacements}{Google Documentation for DeactivatePlacements}
\describe{
 \item{extends PlacementAction}{see documentation for PlacementAction}
}

\strong{EntityChildrenLimitReachedError}

Lists errors relating to having too many children on an entity.
\href{https://developers.google.com/doubleclick-publishers/docs/reference/v201508/PlacementService.EntityChildrenLimitReachedError}{Google Documentation for EntityChildrenLimitReachedError}
\describe{
 \item{extends ApiError}{see documentation for ApiError}
 \item{reason}{a EntityChildrenLimitReachedError.Reason - This can take one of the following values: 
   \itemize{
     \item{LINE_ITEM_LIMIT_FOR_ORDER_REACHED - The number of line items on the order exceeds the max number of line items
allowed per order in the network.}
     \item{CREATIVE_ASSOCIATION_LIMIT_FOR_LINE_ITEM_REACHED - The number of creatives associated with the line item exceeds the max number
of creatives allowed to be associated with a line item in the network.}
     \item{AD_UNIT_LIMIT_FOR_PLACEMENT_REACHED - The number of ad units on the placement exceeds the max number of ad units
allowed per placement in the network.}
     \item{TARGETING_EXPRESSION_LIMIT_FOR_LINE_ITEM_REACHED - The number of targeting expressions on the line item exceeds the max number
of targeting expressions allowed per line item in the network.}
     \item{CUSTOM_TARGETING_VALUES_FOR_KEY_LIMIT_REACHED - The number of custom targeting values for the free-form or predefined custom
targeting key exceeds the max number allowed.}
     \item{TARGETING_EXPRESSION_LIMIT_FOR_CREATIVES_ON_LINE_ITEM_REACHED - The total number of targeting expressions on the creatives for the line item
exceeds the max number allowed per line item in the network.}
     \item{ATTACHMENT_LIMIT_FOR_PROPOSAL_REACHED - The number of attachments added to the proposal exceeds the max number
allowed per proposal in the network.}
     \item{PROPOSAL_LINE_ITEM_LIMIT_FOR_PROPOSAL_REACHED - The number of proposal line items on the proposal exceeds the max number
allowed per proposal in the network.}
     \item{PRODUCT_LIMIT_FOR_PRODUCT_PACKAGE_REACHED - The number of product package items on the product package exceeds the max
number allowed per product package in the network.}
     \item{PRODUCT_TEMPLATE_AND_PRODUCT_BASE_RATE_LIMIT_FOR_RATE_CARD_REACHED - The number of product template and product base rates on the rate card
(including excluded product base rates) exceeds the max number allowed per
rate card in the network.}
     \item{PRODUCT_PACKAGE_ITEM_BASE_RATE_LIMIT_FOR_RATE_CARD_REACHED - The number of product package item base rates on the rate card exceeds the
max number allowed per rate card in the network.}
     \item{PREMIUM_LIMIT_FOR_RATE_CARD_REACHED - The number of premiums of the rate card exceeds the max number allowed per
rate card in the network.}
     \item{AD_UNIT_LIMIT_FOR_AD_EXCLUSION_RULE_TARGETING_REACHED - The number of ad units on AdExclusionRule inventoryTargeting exceeds the max
number of ad units allowed per ad exclusion rule inventory targeting in the
network.}
     \item{UNKNOWN - The value returned if the actual value is not exposed by the requested API
version.}
   }
  }
}

\strong{EntityLimitReachedError}

An error that occurs when creating an entity if the limit on the number of
allowed entities for a network has already been reached.
\href{https://developers.google.com/doubleclick-publishers/docs/reference/v201508/PlacementService.EntityLimitReachedError}{Google Documentation for EntityLimitReachedError}
\describe{
 \item{extends ApiError}{see documentation for ApiError}
 \item{reason}{a EntityLimitReachedError.Reason - This can take one of the following values: 
   \itemize{
     \item{CUSTOM_TARGETING_VALUES_LIMIT_REACHED - The number of custom targeting values exceeds the max number allowed in the
network.}
     \item{AD_EXCLUSION_RULES_LIMIT_REACHED - The number of ad exclusion rules exceeds the max number allowed in the
network.}
     \item{FIRST_PARTY_AUDIENCE_SEGMENTS_LIMIT_REACHED - The number of first party audience segments exceeds the max number allowed
in the network.}
     \item{PLACEMENTS_LIMIT_REACHED - The number of active placements exceeds the max number allowed in the
network.}
     \item{LINE_ITEMS_LIMIT_REACHED - The number of line items excceeds the max number allowed in the network.}
     \item{UNKNOWN - The value returned if the actual value is not exposed by the requested API
version.}
   }
  }
}

\strong{FeatureError}

Errors related to feature management. If you attempt using a feature that is
not available to the current network you'll receive a FeatureError with the
missing feature as the trigger.
\href{https://developers.google.com/doubleclick-publishers/docs/reference/v201508/PlacementService.FeatureError}{Google Documentation for FeatureError}
\describe{
 \item{extends ApiError}{see documentation for ApiError}
 \item{reason}{a FeatureError.Reason - This can take one of the following values: 
   \itemize{
     \item{MISSING_FEATURE - A feature is being used that is not enabled on the current network.}
     \item{UNKNOWN - The value returned if the actual value is not exposed by the requested API
version.}
   }
  }
}

\strong{InternalApiError}

Indicates that a server-side error has occured. InternalApiErrors are
generally not the result of an invalid request or message sent by the
client.
\href{https://developers.google.com/doubleclick-publishers/docs/reference/v201508/PlacementService.InternalApiError}{Google Documentation for InternalApiError}
\describe{
 \item{extends ApiError}{see documentation for ApiError}
 \item{reason}{a InternalApiError.Reason - The error reason represented by an enum. This can take one of the following values: 
   \itemize{
     \item{UNEXPECTED_INTERNAL_API_ERROR - API encountered an unexpected internal error.}
     \item{TRANSIENT_ERROR - A temporary error occurred during the request. Please retry.}
     \item{UNKNOWN - The cause of the error is not known or only defined in newer versions.}
     \item{DOWNTIME - The API is currently unavailable for a planned downtime.}
   }
  }
}

\strong{NotNullError}

Caused by supplying a null value for an attribute that cannot be null.
\href{https://developers.google.com/doubleclick-publishers/docs/reference/v201508/PlacementService.NotNullError}{Google Documentation for NotNullError}
\describe{
 \item{extends ApiError}{see documentation for ApiError}
 \item{reason}{a NotNullError.Reason - The error reason represented by an enum. This can take one of the following values: 
   \itemize{
     \item{ARG1_NULL - Assuming that a method will not have more than 3 arguments, if it does,
return NULL}
     \item{ARG2_NULL}
     \item{ARG3_NULL}
     \item{NULL}
     \item{UNKNOWN - The value returned if the actual value is not exposed by the requested API
version.}
   }
  }
}

\strong{NullError}

Errors associated with violation of a NOT NULL check.
\href{https://developers.google.com/doubleclick-publishers/docs/reference/v201508/PlacementService.NullError}{Google Documentation for NullError}
\describe{
 \item{extends ApiError}{see documentation for ApiError}
 \item{reason}{a NullError.Reason - The error reason represented by an enum. This can take one of the following values: 
   \itemize{
     \item{NULL_CONTENT - Specified list/container must not contain any null elements}
   }
  }
}

\strong{NumberValue}

Contains a numeric value.
\href{https://developers.google.com/doubleclick-publishers/docs/reference/v201508/PlacementService.NumberValue}{Google Documentation for NumberValue}
\describe{
 \item{extends Value}{see documentation for Value}
 \item{value}{a character - The numeric value represented as a string.}
}

\strong{ObjectValue}

Contains an object value.  This object is experimental! ObjectValue is an
experimental, innovative, and rapidly changing new feature for DFP.
Unfortunately, being on the bleeding edge means that we may make
backwards-incompatible changes to ObjectValue. We will inform the community
when this feature is no longer experimental.
\href{https://developers.google.com/doubleclick-publishers/docs/reference/v201508/PlacementService.ObjectValue}{Google Documentation for ObjectValue}
\describe{
 \item{extends Value}{see documentation for Value}
}

\strong{ParseError}

Lists errors related to parsing.
\href{https://developers.google.com/doubleclick-publishers/docs/reference/v201508/PlacementService.ParseError}{Google Documentation for ParseError}
\describe{
 \item{extends ApiError}{see documentation for ApiError}
 \item{reason}{a ParseError.Reason - The error reason represented by an enum. This can take one of the following values: 
   \itemize{
     \item{UNPARSABLE - Indicates an error in parsing an attribute.}
     \item{UNKNOWN - The value returned if the actual value is not exposed by the requested API
version.}
   }
  }
}

\strong{PermissionError}

Errors related to incorrect permission.
\href{https://developers.google.com/doubleclick-publishers/docs/reference/v201508/PlacementService.PermissionError}{Google Documentation for PermissionError}
\describe{
 \item{extends ApiError}{see documentation for ApiError}
 \item{reason}{a PermissionError.Reason - This can take one of the following values: 
   \itemize{
     \item{PERMISSION_DENIED - User does not have the required permission for the request.}
     \item{UNKNOWN - The value returned if the actual value is not exposed by the requested API
version.}
   }
  }
}

\strong{Placement}

A Placement groups related AdUnit objects.
\href{https://developers.google.com/doubleclick-publishers/docs/reference/v201508/PlacementService.Placement}{Google Documentation for Placement}
\describe{
 \item{targetingDescription}{a character (inherited from SiteTargetingInfo) - Description of the Placement visible to AdWords advertisers. This field is
required if Placement isAdSenseTargetingEnabled is true and has a maximum
length of 65535 characters.}
 \item{targetingSiteName}{a character (inherited from SiteTargetingInfo) - The name of the site targeted which is visible to AdWords advertisers. This
field is required if Placement isAdSenseTargetingEnabled is true and has a
maximum of 70 characters.}
 \item{targetingAdLocation}{a character (inherited from SiteTargetingInfo) - A description of the location of the Placement on the web page, visible to
AdWords users. This field is optional and has a maximum length of 50
characters.}
 \item{id}{a integer - Uniquely identifies the Placement. This attribute is read-only and is
assigned by Google when a placement is created.}
 \item{name}{a character - The name of the Placement. This value is required and has a maximum length
of 255 characters.}
 \item{description}{a character - A description of the Placement. This value is optional and its maximum
length is 65,535 characters.}
 \item{placementCode}{a character - A string used to uniquely identify the Placement for purposes of serving the
ad. This attribute is read-only and is assigned by Google when a placement
is created.}
 \item{status}{a InventoryStatus - The status of the Placement. This attribute is read-only. This can take one of the following values: 
   \itemize{
     \item{ACTIVE - The object is active.}
     \item{INACTIVE - The object is no longer active.}
     \item{ARCHIVED - The object has been archived.}
   }
  }
 \item{isAdSenseTargetingEnabled}{a boolean - Describes whether the Placement is available for site targeting by AdWords
advertisers.}
 \item{adSenseTargetingLocale}{a character - If isAdSenseTargetingEnabled is true, restricts AdWords ads to those only in
the placement's language.}
 \item{targetedAdUnitIds}{a character - The collection of AdUnit object IDs that constitute the Placement.}
 \item{lastModifiedDateTime}{a DateTime - The date and time this placement was last modified.}
}

\strong{PlacementAction}

Represents the actions that can be performed on Placement objects. May take one of the actions listed below.
\href{https://developers.google.com/doubleclick-publishers/docs/reference/v201508/PlacementService.PlacementAction}{Google Documentation for PlacementAction}
\describe{
 \item{ActivatePlacements}{The action used for activating Placement objects.}
 \item{ArchivePlacements}{The action used for archiving Placement objects.}
 \item{DeactivatePlacements}{The action used for deactivating Placement objects.}
}

\strong{PlacementError}

Class defining all validation errors for a placement.
\href{https://developers.google.com/doubleclick-publishers/docs/reference/v201508/PlacementService.PlacementError}{Google Documentation for PlacementError}
\describe{
 \item{extends ApiError}{see documentation for ApiError}
 \item{reason}{a PlacementError.Reason - This can take one of the following values: 
   \itemize{
     \item{INVALID_ENTITY_TYPE - Entity type is something other than inventory or content.}
     \item{SHARED_INVENTORY_ASSIGNED - Shared inventory cannot be assigned to a placement.}
     \item{PLACEMENTS_CANNOT_INCLUDE_INVENTORY_FROM_MULTIPLE_DISTRIBUTOR_NETWORKS - Shared inventory from one distributor network cannot be in the same
placement with inventory from another distributor.}
     \item{PLACEMENTS_CANNOT_INCLUDE_BOTH_LOCAL_AND_SHARED_INVENTORY - Shared inventory and local inventory cannot be in the same placement.}
     \item{UNKNOWN - The value returned if the actual value is not exposed by the requested API
version.}
   }
  }
}

\strong{PlacementPage}

Captures a page of Placement objects.
\href{https://developers.google.com/doubleclick-publishers/docs/reference/v201508/PlacementService.PlacementPage}{Google Documentation for PlacementPage}
\describe{
 \item{totalResultSetSize}{a integer - The size of the total result set to which this page belongs.}
 \item{startIndex}{a integer - The absolute index in the total result set on which this page begins.}
 \item{results}{a Placement - The collection of placements contained within this page.}
}

\strong{PublisherQueryLanguageContextError}

An error that occurs while executing a PQL query contained in a Statement
object.
\href{https://developers.google.com/doubleclick-publishers/docs/reference/v201508/PlacementService.PublisherQueryLanguageContextError}{Google Documentation for PublisherQueryLanguageContextError}
\describe{
 \item{extends ApiError}{see documentation for ApiError}
 \item{reason}{a PublisherQueryLanguageContextError.Reason - The error reason represented by an enum. This can take one of the following values: 
   \itemize{
     \item{UNEXECUTABLE - Indicates that there was an error executing the PQL.}
     \item{UNKNOWN - The value returned if the actual value is not exposed by the requested API
version.}
   }
  }
}

\strong{PublisherQueryLanguageSyntaxError}

An error that occurs while parsing a PQL query contained in a Statement
object.
\href{https://developers.google.com/doubleclick-publishers/docs/reference/v201508/PlacementService.PublisherQueryLanguageSyntaxError}{Google Documentation for PublisherQueryLanguageSyntaxError}
\describe{
 \item{extends ApiError}{see documentation for ApiError}
 \item{reason}{a PublisherQueryLanguageSyntaxError.Reason - The error reason represented by an enum. This can take one of the following values: 
   \itemize{
     \item{UNPARSABLE - Indicates that there was a PQL syntax error.}
     \item{UNKNOWN - The value returned if the actual value is not exposed by the requested API
version.}
   }
  }
}

\strong{QuotaError}

Describes a client-side error on which a user is attempting to perform an
action to which they have no quota remaining.
\href{https://developers.google.com/doubleclick-publishers/docs/reference/v201508/PlacementService.QuotaError}{Google Documentation for QuotaError}
\describe{
 \item{extends ApiError}{see documentation for ApiError}
 \item{reason}{a QuotaError.Reason - This can take one of the following values: 
   \itemize{
     \item{EXCEEDED_QUOTA - The number of requests made per second is too high and has exceeded the
allowable limit. The recommended approach to handle this error is to wait
about 5 seconds and then retry the request. Note that this does not
guarantee the request will succeed. If it fails again, try increasing the
wait time.  Another way to mitigate this error is to limit requests to 2 per
second for Small Business networks, or 8 per second for Premium networks.
Once again this does not guarantee that every request will succeed, but may
help reduce the number of times you receive this error.}
     \item{UNKNOWN - The value returned if the actual value is not exposed by the requested API
version.}
   }
  }
}

\strong{RangeError}

A list of all errors associated with the Range constraint.
\href{https://developers.google.com/doubleclick-publishers/docs/reference/v201508/PlacementService.RangeError}{Google Documentation for RangeError}
\describe{
 \item{extends ApiError}{see documentation for ApiError}
 \item{reason}{a RangeError.Reason - This can take one of the following values: 
   \itemize{
     \item{TOO_HIGH}
     \item{TOO_LOW}
     \item{UNKNOWN - The value returned if the actual value is not exposed by the requested API
version.}
   }
  }
}

\strong{RegExError}

Caused by supplying a value for an object attribute that does not conform to
a documented valid regular expression.
\href{https://developers.google.com/doubleclick-publishers/docs/reference/v201508/PlacementService.RegExError}{Google Documentation for RegExError}
\describe{
 \item{extends ApiError}{see documentation for ApiError}
 \item{reason}{a RegExError.Reason - The error reason represented by an enum. This can take one of the following values: 
   \itemize{
     \item{INVALID - Invalid value found.}
     \item{NULL - Null value found.}
     \item{UNKNOWN - The value returned if the actual value is not exposed by the requested API
version.}
   }
  }
}

\strong{RequiredCollectionError}

A list of all errors to be used for validating sizes of collections.
\href{https://developers.google.com/doubleclick-publishers/docs/reference/v201508/PlacementService.RequiredCollectionError}{Google Documentation for RequiredCollectionError}
\describe{
 \item{extends ApiError}{see documentation for ApiError}
 \item{reason}{a RequiredCollectionError.Reason - This can take one of the following values: 
   \itemize{
     \item{REQUIRED - A required collection is missing.}
     \item{TOO_LARGE - Collection size is too large.}
     \item{TOO_SMALL - Collection size is too small.}
     \item{UNKNOWN - The value returned if the actual value is not exposed by the requested API
version.}
   }
  }
}

\strong{RequiredError}

Errors due to missing required field.
\href{https://developers.google.com/doubleclick-publishers/docs/reference/v201508/PlacementService.RequiredError}{Google Documentation for RequiredError}
\describe{
 \item{extends ApiError}{see documentation for ApiError}
 \item{reason}{a RequiredError.Reason - The error reason represented by an enum. This can take one of the following values: 
   \itemize{
     \item{REQUIRED - Missing required field.}
   }
  }
}

\strong{ServerError}

Errors related to the server.
\href{https://developers.google.com/doubleclick-publishers/docs/reference/v201508/PlacementService.ServerError}{Google Documentation for ServerError}
\describe{
 \item{extends ApiError}{see documentation for ApiError}
 \item{reason}{a ServerError.Reason - This can take one of the following values: 
   \itemize{
     \item{SERVER_ERROR - Indicates that an unexpected error occured.}
     \item{SERVER_BUSY - Indicates that the server is currently experiencing a high load. Please wait
and try your request again.}
     \item{UNKNOWN - The value returned if the actual value is not exposed by the requested API
version.}
   }
  }
}

\strong{SetValue}

Contains a set of Value Values. May not contain duplicates.
\href{https://developers.google.com/doubleclick-publishers/docs/reference/v201508/PlacementService.SetValue}{Google Documentation for SetValue}
\describe{
 \item{extends Value}{see documentation for Value}
 \item{values}{a Value - The values. They must all be the same type of Value and not contain
duplicates.}
}

\strong{SiteTargetingInfo}

Contains information required for AdWords advertisers to place their ads.
\href{https://developers.google.com/doubleclick-publishers/docs/reference/v201508/PlacementService.SiteTargetingInfo}{Google Documentation for SiteTargetingInfo}
\describe{
 \item{targetingDescription}{a character - Description of the Placement visible to AdWords advertisers. This field is
required if Placement isAdSenseTargetingEnabled is true and has a maximum
length of 65535 characters.}
 \item{targetingSiteName}{a character - The name of the site targeted which is visible to AdWords advertisers. This
field is required if Placement isAdSenseTargetingEnabled is true and has a
maximum of 70 characters.}
 \item{targetingAdLocation}{a character - A description of the location of the Placement on the web page, visible to
AdWords users. This field is optional and has a maximum length of 50
characters.}
}

\strong{SoapRequestHeader}

Represents the SOAP request header used by API requests.
\href{https://developers.google.com/doubleclick-publishers/docs/reference/v201508/PlacementService.SoapRequestHeader}{Google Documentation for SoapRequestHeader}
\describe{
 \item{networkCode}{a character - The network code to use in the context of a request.}
 \item{applicationName}{a character - The name of client library application.}
}

\strong{SoapResponseHeader}

Represents the SOAP request header used by API responses.
\href{https://developers.google.com/doubleclick-publishers/docs/reference/v201508/PlacementService.SoapResponseHeader}{Google Documentation for SoapResponseHeader}
\describe{
 \item{requestId}{a character}
 \item{responseTime}{a integer}
}

\strong{Statement}

Holds the query in PQL syntax. See link for how to form PQL.
\href{https://developers.google.com/doubleclick-publishers/docs/reference/v201508/PlacementService.Statement}{Google Documentation for Statement}
\describe{
 \item{query}{a character - Holds the query in PQL syntax. See link for how to form PQL.}
 \item{values}{a String_ValueMapEntry - Holds keys and values for bind variables and their values. The key is the
name of the bind variable. The value is the literal value of the variable.
In the example "WHERE status = :bindStatus ORDER BY id LIMIT 30", the bind
variable, represented by :bindStatus is named bindStatus, which would also
be the parameter map key. The bind variable's value would be represented by
a parameter map value of type TextValue. The final result, for example,
would be an entry of "bindStatus" => StringParam("ACTIVE").}
}

\strong{StatementError}

An error that occurs while parsing Statement objects.
\href{https://developers.google.com/doubleclick-publishers/docs/reference/v201508/PlacementService.StatementError}{Google Documentation for StatementError}
\describe{
 \item{extends ApiError}{see documentation for ApiError}
 \item{reason}{a StatementError.Reason - The error reason represented by an enum. This can take one of the following values: 
   \itemize{
     \item{VARIABLE_NOT_BOUND_TO_VALUE - A bind variable has not been bound to a value.}
     \item{UNKNOWN - The value returned if the actual value is not exposed by the requested API
version.}
   }
  }
}

\strong{String_ValueMapEntry}

This represents an entry in a map with a key of type String and value of
type Value.
\href{https://developers.google.com/doubleclick-publishers/docs/reference/v201508/PlacementService.String_ValueMapEntry}{Google Documentation for String_ValueMapEntry}
\describe{
 \item{key}{a character}
 \item{value}{a Value}
}

\strong{StringLengthError}

Errors for Strings which do not meet given length constraints.
\href{https://developers.google.com/doubleclick-publishers/docs/reference/v201508/PlacementService.StringLengthError}{Google Documentation for StringLengthError}
\describe{
 \item{extends ApiError}{see documentation for ApiError}
 \item{reason}{a StringLengthError.Reason - This can take one of the following values: 
   \itemize{
     \item{TOO_LONG}
     \item{TOO_SHORT}
     \item{UNKNOWN - The value returned if the actual value is not exposed by the requested API
version.}
   }
  }
}

\strong{TextValue}

Contains a string value.
\href{https://developers.google.com/doubleclick-publishers/docs/reference/v201508/PlacementService.TextValue}{Google Documentation for TextValue}
\describe{
 \item{extends Value}{see documentation for Value}
 \item{value}{a character - The string value.}
}

\strong{TypeError}

An error for a field which is an invalid type.
\href{https://developers.google.com/doubleclick-publishers/docs/reference/v201508/PlacementService.TypeError}{Google Documentation for TypeError}
\describe{
 \item{extends ApiError}{see documentation for ApiError}
}

\strong{UniqueError}

An error for a field which must satisfy a uniqueness constraint
\href{https://developers.google.com/doubleclick-publishers/docs/reference/v201508/PlacementService.UniqueError}{Google Documentation for UniqueError}
\describe{
 \item{extends ApiError}{see documentation for ApiError}
}

\strong{UpdateResult}

Represents the result of performing an action on objects.
\href{https://developers.google.com/doubleclick-publishers/docs/reference/v201508/PlacementService.UpdateResult}{Google Documentation for UpdateResult}
\describe{
 \item{numChanges}{a integer - The number of objects that were changed as a result of performing the
action.}
}

\strong{Value}

Value represents a value.
\href{https://developers.google.com/doubleclick-publishers/docs/reference/v201508/PlacementService.Value}{Google Documentation for Value}
\describe{
}
}

