% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PremiumRateService.R
\name{dfp_PremiumRateService_object_factory}
\alias{dfp_PremiumRateService_object_factory}
\title{PremiumRateService}
\usage{
dfp_PremiumRateService_object_factory(obj_type, obj_data)
}
\arguments{
\item{obj_type}{a string from one of the object types described below}

\item{obj_data}{a \code{list} of \code{lists} or a \code{data.frame} with the required components to create
the obj_type specified. All input names must match the expected attributes for the object.
If submitting a list of attributes to create 1 object make sure to nest it inside top-level list, so that
it is a list of length 1 with 1 list of attributes to create the object.}
}
\value{
a string of formatted XML that can be inserted into a SOAP request body to the API
}
\description{
Provides methods for managing PremiumRate objects.
To use this service,
you need to have the new sales management solution enabled on your network.
If you do not see a "Sales" tab in <a
href="https://www.google.com/dfp">DoubleClick for Publishers (DFP)</a>, you
will not be able to use this service.
}
\details{
PremiumRateService Object Factory

A function to create a variety of objects that are part of the PremiumRateService
Note: not all objects here are inputs to the functions of this service so 
please consult each function for the required inputs

Below is a list of objects and their required components to be created with this function:

\strong{AdUnitPremiumFeature}

A premium feature applied to ad unit targeting.
\href{https://developers.google.com/doubleclick-publishers/docs/reference/v201508/PremiumRateService.AdUnitPremiumFeature}{Google Documentation for AdUnitPremiumFeature}
\describe{
 \item{extends PremiumFeature}{see documentation for PremiumFeature}
}

\strong{ApiError}

The API error base class that provides details about an error that occurred
while processing a service request. The OGNL field path is provided for
parsers to identify the request data element that may have caused the error.
\href{https://developers.google.com/doubleclick-publishers/docs/reference/v201508/PremiumRateService.ApiError}{Google Documentation for ApiError}
\describe{
 \item{fieldPath}{a string -  The OGNL field path to identify cause of error.}
 \item{trigger}{a string -  The data that caused the error.}
 \item{errorString}{a string -  A simple string representation of the error and reason.}
}

\strong{ApiException}

Exception class for holding a list of service errors.
\href{https://developers.google.com/doubleclick-publishers/docs/reference/v201508/PremiumRateService.ApiException}{Google Documentation for ApiException}
\describe{
 \item{message}{a string (inherited from ApplicationException) -  Error message.}
 \item{errors}{a ApiError -  List of errors.}
}

\strong{ApiVersionError}

Errors related to the usage of API versions.
\href{https://developers.google.com/doubleclick-publishers/docs/reference/v201508/PremiumRateService.ApiVersionError}{Google Documentation for ApiVersionError}
\describe{
 \item{extends ApiError}{see documentation for ApiError}
 \item{reason}{a ApiVersionError.Reason - This can take one of the following values:
   \itemize{
     \item{UPDATE_TO_NEWER_VERSION - Indicates that the operation is not allowed in the version the request was
made in.}
     \item{UNKNOWN - The value returned if the actual value is not exposed by the requested API
version.}
   }
  }
}

\strong{ApplicationException}

Base class for exceptions.
\href{https://developers.google.com/doubleclick-publishers/docs/reference/v201508/PremiumRateService.ApplicationException}{Google Documentation for ApplicationException}
\describe{
 \item{message}{a string -  Error message.}
}

\strong{AudienceSegmentPremiumFeature}

A premium feature applied to audience segment targeting.
\href{https://developers.google.com/doubleclick-publishers/docs/reference/v201508/PremiumRateService.AudienceSegmentPremiumFeature}{Google Documentation for AudienceSegmentPremiumFeature}
\describe{
 \item{extends PremiumFeature}{see documentation for PremiumFeature}
 \item{audienceSegmentId}{a integer -  The ID of the AudienceSegment id. This attribute is required and cannot be
changed after creation.}
}

\strong{AuthenticationError}

An error for an exception that occurred when authenticating.
\href{https://developers.google.com/doubleclick-publishers/docs/reference/v201508/PremiumRateService.AuthenticationError}{Google Documentation for AuthenticationError}
\describe{
 \item{extends ApiError}{see documentation for ApiError}
 \item{reason}{a AuthenticationError.Reason - This can take one of the following values:
   \itemize{
     \item{AMBIGUOUS_SOAP_REQUEST_HEADER - The SOAP message contains a request header with an ambiguous definition of
the authentication header fields. This means either the authToken and
oAuthToken fields were both null or both were specified. Exactly one value
should be specified with each request.}
     \item{INVALID_EMAIL - The login provided is invalid.}
     \item{AUTHENTICATION_FAILED - Tried to authenticate with provided information, but failed.}
     \item{INVALID_OAUTH_SIGNATURE - The OAuth provided is invalid.}
     \item{INVALID_SERVICE - The specified service to use was not recognized.}
     \item{MISSING_SOAP_REQUEST_HEADER - The SOAP message is missing a request header with an authToken and optional
networkCode.}
     \item{MISSING_AUTHENTICATION_HTTP_HEADER - The HTTP request is missing a request header with an authToken}
     \item{MISSING_AUTHENTICATION - The request is missing an authToken}
     \item{NOT_WHITELISTED_FOR_API_ACCESS - The customer is not whitelisted for API access.}
     \item{NO_NETWORKS_TO_ACCESS - The user is not associated with any network.}
     \item{NETWORK_NOT_FOUND - No network for the given networkCode was found.}
     \item{NETWORK_CODE_REQUIRED - The user has access to more than one network, but did not provide a
networkCode.}
     \item{CONNECTION_ERROR - An error happened on the server side during connection to authentication
service.}
     \item{GOOGLE_ACCOUNT_ALREADY_ASSOCIATED_WITH_NETWORK - The user tried to create a test network using an account that already is
associated with a network.}
     \item{UNKNOWN - The value returned if the actual value is not exposed by the requested API
version.}
   }
  }
}

\strong{BandwidthPremiumFeature}

A premium feature applied to bandwidth targeting.
\href{https://developers.google.com/doubleclick-publishers/docs/reference/v201508/PremiumRateService.BandwidthPremiumFeature}{Google Documentation for BandwidthPremiumFeature}
\describe{
 \item{extends PremiumFeature}{see documentation for PremiumFeature}
}

\strong{BooleanValue}

Contains a boolean value.
\href{https://developers.google.com/doubleclick-publishers/docs/reference/v201508/PremiumRateService.BooleanValue}{Google Documentation for BooleanValue}
\describe{
 \item{extends Value}{see documentation for Value}
 \item{value}{a boolean -  The boolean value.}
}

\strong{BrowserLanguagePremiumFeature}

A premium feature applied to browser language targeting.
\href{https://developers.google.com/doubleclick-publishers/docs/reference/v201508/PremiumRateService.BrowserLanguagePremiumFeature}{Google Documentation for BrowserLanguagePremiumFeature}
\describe{
 \item{extends PremiumFeature}{see documentation for PremiumFeature}
}

\strong{BrowserPremiumFeature}

A premium feature applied to browser targeting.
\href{https://developers.google.com/doubleclick-publishers/docs/reference/v201508/PremiumRateService.BrowserPremiumFeature}{Google Documentation for BrowserPremiumFeature}
\describe{
 \item{extends PremiumFeature}{see documentation for PremiumFeature}
}

\strong{CollectionSizeError}

Error for the size of the collection being too large
\href{https://developers.google.com/doubleclick-publishers/docs/reference/v201508/PremiumRateService.CollectionSizeError}{Google Documentation for CollectionSizeError}
\describe{
 \item{extends ApiError}{see documentation for ApiError}
 \item{reason}{a CollectionSizeError.Reason - This can take one of the following values:
   \itemize{
     \item{TOO_LARGE}
     \item{UNKNOWN - The value returned if the actual value is not exposed by the requested API
version.}
   }
  }
}

\strong{CommonError}

A place for common errors that can be used across services.
\href{https://developers.google.com/doubleclick-publishers/docs/reference/v201508/PremiumRateService.CommonError}{Google Documentation for CommonError}
\describe{
 \item{extends ApiError}{see documentation for ApiError}
 \item{reason}{a CommonError.Reason - This can take one of the following values:
   \itemize{
     \item{NOT_FOUND - Indicates that an attempt was made to retrieve an entity that does not
exist.}
     \item{ALREADY_EXISTS - Indicates that an attempt was made to create an entity that already exists.}
     \item{DUPLICATE_OBJECT - Indicates that two elements in the collection were identical.}
     \item{CANNOT_UPDATE - Indicates that an attempt was made to change an immutable field.}
     \item{CONCURRENT_MODIFICATION - Indicates that another request attempted to update the same data in the same
network at about the same time. Please wait and try the request again.}
     \item{UNKNOWN - The value returned if the actual value is not exposed by the requested API
version.}
   }
  }
}

\strong{ContentBundlePremiumFeature}

A premium feature applied to content bundle targeting.
\href{https://developers.google.com/doubleclick-publishers/docs/reference/v201508/PremiumRateService.ContentBundlePremiumFeature}{Google Documentation for ContentBundlePremiumFeature}
\describe{
 \item{extends PremiumFeature}{see documentation for PremiumFeature}
}

\strong{CustomTargetingPremiumFeature}

A premium feature applied to custom criteria targeting.
\href{https://developers.google.com/doubleclick-publishers/docs/reference/v201508/PremiumRateService.CustomTargetingPremiumFeature}{Google Documentation for CustomTargetingPremiumFeature}
\describe{
 \item{extends PremiumFeature}{see documentation for PremiumFeature}
 \item{customTargetingKeyId}{a integer -  The ID of the CustomTargetingKey. This attribute is required and cannot be
changed after creation.}
 \item{customTargetingValueId}{a integer -  The ID of the CustomTargetingValue. null if all custom targeting values of
the customTargetingKeyId should be matched. This attribute cannot be changed
after creation.}
}

\strong{Date}

Represents a date.
\href{https://developers.google.com/doubleclick-publishers/docs/reference/v201508/PremiumRateService.Date}{Google Documentation for Date}
\describe{
 \item{year}{a integer -  Year (e.g., 2009)}
 \item{month}{a integer -  Month (1..12)}
 \item{day}{a integer -  Day (1..31)}
}

\strong{DateTime}

Represents a date combined with the time of day.
\href{https://developers.google.com/doubleclick-publishers/docs/reference/v201508/PremiumRateService.DateTime}{Google Documentation for DateTime}
\describe{
 \item{date}{a Date}
 \item{hour}{a integer}
 \item{minute}{a integer}
 \item{second}{a integer}
 \item{timeZoneID}{a string}
}

\strong{DateTimeValue}

Contains a date-time value.
\href{https://developers.google.com/doubleclick-publishers/docs/reference/v201508/PremiumRateService.DateTimeValue}{Google Documentation for DateTimeValue}
\describe{
 \item{extends Value}{see documentation for Value}
 \item{value}{a DateTime -  The DateTime value.}
}

\strong{DateValue}

Contains a date value.
\href{https://developers.google.com/doubleclick-publishers/docs/reference/v201508/PremiumRateService.DateValue}{Google Documentation for DateValue}
\describe{
 \item{extends Value}{see documentation for Value}
 \item{value}{a Date -  The Date value.}
}

\strong{DaypartPremiumFeature}

A premium feature applied to daypart targeting.
\href{https://developers.google.com/doubleclick-publishers/docs/reference/v201508/PremiumRateService.DaypartPremiumFeature}{Google Documentation for DaypartPremiumFeature}
\describe{
 \item{extends PremiumFeature}{see documentation for PremiumFeature}
}

\strong{DeviceCapabilityPremiumFeature}

A premium feature applied to device capability targeting.
\href{https://developers.google.com/doubleclick-publishers/docs/reference/v201508/PremiumRateService.DeviceCapabilityPremiumFeature}{Google Documentation for DeviceCapabilityPremiumFeature}
\describe{
 \item{extends PremiumFeature}{see documentation for PremiumFeature}
}

\strong{DeviceCategoryPremiumFeature}

A premium feature applied to device category targeting.
\href{https://developers.google.com/doubleclick-publishers/docs/reference/v201508/PremiumRateService.DeviceCategoryPremiumFeature}{Google Documentation for DeviceCategoryPremiumFeature}
\describe{
 \item{extends PremiumFeature}{see documentation for PremiumFeature}
}

\strong{DeviceManufacturerPremiumFeature}

A premium feature applied to device manufacturer targeting.
\href{https://developers.google.com/doubleclick-publishers/docs/reference/v201508/PremiumRateService.DeviceManufacturerPremiumFeature}{Google Documentation for DeviceManufacturerPremiumFeature}
\describe{
 \item{extends PremiumFeature}{see documentation for PremiumFeature}
}

\strong{EntityChildrenLimitReachedError}

Lists errors relating to having too many children on an entity.
\href{https://developers.google.com/doubleclick-publishers/docs/reference/v201508/PremiumRateService.EntityChildrenLimitReachedError}{Google Documentation for EntityChildrenLimitReachedError}
\describe{
 \item{extends ApiError}{see documentation for ApiError}
 \item{reason}{a EntityChildrenLimitReachedError.Reason - This can take one of the following values:
   \itemize{
     \item{LINE_ITEM_LIMIT_FOR_ORDER_REACHED - The number of line items on the order exceeds the max number of line items
allowed per order in the network.}
     \item{CREATIVE_ASSOCIATION_LIMIT_FOR_LINE_ITEM_REACHED - The number of creatives associated with the line item exceeds the max number
of creatives allowed to be associated with a line item in the network.}
     \item{AD_UNIT_LIMIT_FOR_PLACEMENT_REACHED - The number of ad units on the placement exceeds the max number of ad units
allowed per placement in the network.}
     \item{TARGETING_EXPRESSION_LIMIT_FOR_LINE_ITEM_REACHED - The number of targeting expressions on the line item exceeds the max number
of targeting expressions allowed per line item in the network.}
     \item{CUSTOM_TARGETING_VALUES_FOR_KEY_LIMIT_REACHED - The number of custom targeting values for the free-form or predefined custom
targeting key exceeds the max number allowed.}
     \item{TARGETING_EXPRESSION_LIMIT_FOR_CREATIVES_ON_LINE_ITEM_REACHED - The total number of targeting expressions on the creatives for the line item
exceeds the max number allowed per line item in the network.}
     \item{ATTACHMENT_LIMIT_FOR_PROPOSAL_REACHED - The number of attachments added to the proposal exceeds the max number
allowed per proposal in the network.}
     \item{PROPOSAL_LINE_ITEM_LIMIT_FOR_PROPOSAL_REACHED - The number of proposal line items on the proposal exceeds the max number
allowed per proposal in the network.}
     \item{PRODUCT_LIMIT_FOR_PRODUCT_PACKAGE_REACHED - The number of product package items on the product package exceeds the max
number allowed per product package in the network.}
     \item{PRODUCT_TEMPLATE_AND_PRODUCT_BASE_RATE_LIMIT_FOR_RATE_CARD_REACHED - The number of product template and product base rates on the rate card
(including excluded product base rates) exceeds the max number allowed per
rate card in the network.}
     \item{PRODUCT_PACKAGE_ITEM_BASE_RATE_LIMIT_FOR_RATE_CARD_REACHED - The number of product package item base rates on the rate card exceeds the
max number allowed per rate card in the network.}
     \item{PREMIUM_LIMIT_FOR_RATE_CARD_REACHED - The number of premiums of the rate card exceeds the max number allowed per
rate card in the network.}
     \item{AD_UNIT_LIMIT_FOR_AD_EXCLUSION_RULE_TARGETING_REACHED - The number of ad units on AdExclusionRule inventoryTargeting exceeds the max
number of ad units allowed per ad exclusion rule inventory targeting in the
network.}
     \item{UNKNOWN - The value returned if the actual value is not exposed by the requested API
version.}
   }
  }
}

\strong{FeatureError}

Errors related to feature management. If you attempt using a feature that is
not available to the current network you'll receive a FeatureError with the
missing feature as the trigger.
\href{https://developers.google.com/doubleclick-publishers/docs/reference/v201508/PremiumRateService.FeatureError}{Google Documentation for FeatureError}
\describe{
 \item{extends ApiError}{see documentation for ApiError}
 \item{reason}{a FeatureError.Reason - This can take one of the following values:
   \itemize{
     \item{MISSING_FEATURE - A feature is being used that is not enabled on the current network.}
     \item{UNKNOWN - The value returned if the actual value is not exposed by the requested API
version.}
   }
  }
}

\strong{FrequencyCapPremiumFeature}

A premium feature applied to frequency caps.
\href{https://developers.google.com/doubleclick-publishers/docs/reference/v201508/PremiumRateService.FrequencyCapPremiumFeature}{Google Documentation for FrequencyCapPremiumFeature}
\describe{
 \item{extends PremiumFeature}{see documentation for PremiumFeature}
}

\strong{GeographyPremiumFeature}

A premium feature applied to geography targeting.
\href{https://developers.google.com/doubleclick-publishers/docs/reference/v201508/PremiumRateService.GeographyPremiumFeature}{Google Documentation for GeographyPremiumFeature}
\describe{
 \item{extends PremiumFeature}{see documentation for PremiumFeature}
}

\strong{InternalApiError}

Indicates that a server-side error has occured. InternalApiErrors are
generally not the result of an invalid request or message sent by the
client.
\href{https://developers.google.com/doubleclick-publishers/docs/reference/v201508/PremiumRateService.InternalApiError}{Google Documentation for InternalApiError}
\describe{
 \item{extends ApiError}{see documentation for ApiError}
 \item{reason}{a InternalApiError.Reason - The error reason represented by an enum. This can take one of the following values:
   \itemize{
     \item{UNEXPECTED_INTERNAL_API_ERROR - API encountered an unexpected internal error.}
     \item{TRANSIENT_ERROR - A temporary error occurred during the request. Please retry.}
     \item{UNKNOWN - The cause of the error is not known or only defined in newer versions.}
     \item{DOWNTIME - The API is currently unavailable for a planned downtime.}
   }
  }
}

\strong{MobileCarrierPremiumFeature}

A premium feature applied to mobile carrier targeting.
\href{https://developers.google.com/doubleclick-publishers/docs/reference/v201508/PremiumRateService.MobileCarrierPremiumFeature}{Google Documentation for MobileCarrierPremiumFeature}
\describe{
 \item{extends PremiumFeature}{see documentation for PremiumFeature}
}

\strong{NotNullError}

Caused by supplying a null value for an attribute that cannot be null.
\href{https://developers.google.com/doubleclick-publishers/docs/reference/v201508/PremiumRateService.NotNullError}{Google Documentation for NotNullError}
\describe{
 \item{extends ApiError}{see documentation for ApiError}
 \item{reason}{a NotNullError.Reason - The error reason represented by an enum. This can take one of the following values:
   \itemize{
     \item{ARG1_NULL - Assuming that a method will not have more than 3 arguments, if it does,
return NULL}
     \item{ARG2_NULL}
     \item{ARG3_NULL}
     \item{NULL}
     \item{UNKNOWN - The value returned if the actual value is not exposed by the requested API
version.}
   }
  }
}

\strong{NullError}

Errors associated with violation of a NOT NULL check.
\href{https://developers.google.com/doubleclick-publishers/docs/reference/v201508/PremiumRateService.NullError}{Google Documentation for NullError}
\describe{
 \item{extends ApiError}{see documentation for ApiError}
 \item{reason}{a NullError.Reason - The error reason represented by an enum. This can take one of the following values:
   \itemize{
     \item{NULL_CONTENT - Specified list/container must not contain any null elements}
   }
  }
}

\strong{NumberValue}

Contains a numeric value.
\href{https://developers.google.com/doubleclick-publishers/docs/reference/v201508/PremiumRateService.NumberValue}{Google Documentation for NumberValue}
\describe{
 \item{extends Value}{see documentation for Value}
 \item{value}{a string -  The numeric value represented as a string.}
}

\strong{ObjectValue}

Contains an object value.  This object is experimental! ObjectValue is an
experimental, innovative, and rapidly changing new feature for DFP.
Unfortunately, being on the bleeding edge means that we may make
backwards-incompatible changes to ObjectValue. We will inform the community
when this feature is no longer experimental.
\href{https://developers.google.com/doubleclick-publishers/docs/reference/v201508/PremiumRateService.ObjectValue}{Google Documentation for ObjectValue}
\describe{
 \item{extends Value}{see documentation for Value}
}

\strong{OperatingSystemPremiumFeature}

A premium feature applied to operating system targeting.
\href{https://developers.google.com/doubleclick-publishers/docs/reference/v201508/PremiumRateService.OperatingSystemPremiumFeature}{Google Documentation for OperatingSystemPremiumFeature}
\describe{
 \item{extends PremiumFeature}{see documentation for PremiumFeature}
}

\strong{ParseError}

Lists errors related to parsing.
\href{https://developers.google.com/doubleclick-publishers/docs/reference/v201508/PremiumRateService.ParseError}{Google Documentation for ParseError}
\describe{
 \item{extends ApiError}{see documentation for ApiError}
 \item{reason}{a ParseError.Reason - The error reason represented by an enum. This can take one of the following values:
   \itemize{
     \item{UNPARSABLE - Indicates an error in parsing an attribute.}
     \item{UNKNOWN - The value returned if the actual value is not exposed by the requested API
version.}
   }
  }
}

\strong{PermissionError}

Errors related to incorrect permission.
\href{https://developers.google.com/doubleclick-publishers/docs/reference/v201508/PremiumRateService.PermissionError}{Google Documentation for PermissionError}
\describe{
 \item{extends ApiError}{see documentation for ApiError}
 \item{reason}{a PermissionError.Reason - This can take one of the following values:
   \itemize{
     \item{PERMISSION_DENIED - User does not have the required permission for the request.}
     \item{UNKNOWN - The value returned if the actual value is not exposed by the requested API
version.}
   }
  }
}

\strong{PlacementPremiumFeature}

A premium feature applied to placement targeting.
\href{https://developers.google.com/doubleclick-publishers/docs/reference/v201508/PremiumRateService.PlacementPremiumFeature}{Google Documentation for PlacementPremiumFeature}
\describe{
 \item{extends PremiumFeature}{see documentation for PremiumFeature}
}

\strong{PremiumFeature}

A PremiumFeature represents the feature type to be applied as a premium on a
RateCard.
\href{https://developers.google.com/doubleclick-publishers/docs/reference/v201508/PremiumRateService.PremiumFeature}{Google Documentation for PremiumFeature}
\describe{
}

\strong{PremiumRate}

A premium rate holding a set of PremiumRateValue objects with the same
PremiumFeature.
\href{https://developers.google.com/doubleclick-publishers/docs/reference/v201508/PremiumRateService.PremiumRate}{Google Documentation for PremiumRate}
\describe{
 \item{id}{a integer -  Uniquely identifies the PremiumRate object. This attribute is read-only and
is assigned by Google when a premium rate is created.}
 \item{rateCardId}{a integer -  The ID of the RateCard object to which this premium rate belongs. This
attribute is required and cannot be changed after creation.}
 \item{pricingMethod}{a PricingMethod - The method of deciding which PremiumRateValue objects from this PremiumRate
to apply to a ProposalLineItem. This attribute is required. This can take one of the following values:
   \itemize{
     \item{SUM - Applies sum of all matched PremiumRateValue objects in the PremiumRate.}
     \item{HIGHEST - Applies the matched PremiumRateValue with highest adjustment size.}
     \item{ANY_VALUE - Only PremiumRateValue objects with 'Any' matching value are allowed to be
added to this PremiumRate.}
     \item{UNKNOWN - The value returned if the actual value is not exposed by the requested API
version.}
   }
  }
 \item{premiumFeature}{a PremiumFeature -  The feature type of this premium rate. This attribute is required and cannot
be changed after creation.}
 \item{premiumRateValues}{a PremiumRateValue -  The premiums rate values in this rate.}
}

\strong{PremiumRateError}

An error having to do with PremiumRate.
\href{https://developers.google.com/doubleclick-publishers/docs/reference/v201508/PremiumRateService.PremiumRateError}{Google Documentation for PremiumRateError}
\describe{
 \item{extends ApiError}{see documentation for ApiError}
 \item{reason}{a PremiumRateError.Reason - This can take one of the following values:
   \itemize{
     \item{INVALID_RATE_TYPE - The rate type of the PremiumRateValue is invalid.}
     \item{INVALID_PRICING_METHOD - The pricing method of the PremiumRate is invalid.}
     \item{INVALID_PREMIUM_FEATURE - The premium feature of the PremiumRate is invalid.}
     \item{UNKNOWN - The value returned if the actual value is not exposed by the requested API
version.}
   }
  }
}

\strong{PremiumRatePage}

Captures a page of PremiumRate objects.
\href{https://developers.google.com/doubleclick-publishers/docs/reference/v201508/PremiumRateService.PremiumRatePage}{Google Documentation for PremiumRatePage}
\describe{
 \item{results}{a PremiumRate -  The collection of premium rates contained within this page.}
 \item{startIndex}{a integer -  The absolute index in the total result set on which this page begins.}
 \item{totalResultSetSize}{a integer -  The size of the total result set to which this page belongs.}
}

\strong{PremiumRateValue}

A premium of a RateCard which could be applied to the line item and charges
extra.
\href{https://developers.google.com/doubleclick-publishers/docs/reference/v201508/PremiumRateService.PremiumRateValue}{Google Documentation for PremiumRateValue}
\describe{
 \item{premiumRateId}{a integer -  The ID of the PremiumRate object to which this premium rate value belongs.
This attribute is readonly.}
 \item{premiumFeature}{a PremiumFeature -  The feature type of this premium rate value. This attribute is required and
cannot be changed after creation.}
 \item{rateType}{a RateType - The rate type this premium rate value would be applied to. This attribute is
required. This can take one of the following values:
   \itemize{
     \item{CPM - The rate applies to cost per mille (CPM) revenue.}
     \item{CPC - The rate applies to cost per click (CPC) revenue.}
     \item{CPD - The rate applies to cost per day (CPD) revenue.}
     \item{CPU - The rate applies to cost per unit (CPU) revenue.}
     \item{FLAT_FEE - The rate applies to flat fee revenue.}
     \item{UNKNOWN - The value returned if the actual value is not exposed by the requested API
version.}
   }
  }
 \item{adjustmentType}{a PremiumAdjustmentType - Whether the premium rate value is a percentage or absolute value to the
product rate. This attribute is required. This can take one of the following values:
   \itemize{
     \item{PERCENTAGE - The adjustment size to the base rate is a percentage.}
     \item{ABSOLUTE_VALUE - The adjustment size to the base rate is an absolute value.}
     \item{UNKNOWN - The value returned if the actual value is not exposed by the requested API
version.}
   }
  }
 \item{adjustmentSize}{a integer -  If adjustmentType is PremiumAdjustmentType ABSOLUTE_VALUE, this value is in
micros. If adjustmentType is PremiumAdjustmentType PERCENTAGE, this value is
in millipercent. This attribute is required.}
}

\strong{PublisherQueryLanguageContextError}

An error that occurs while executing a PQL query contained in a Statement
object.
\href{https://developers.google.com/doubleclick-publishers/docs/reference/v201508/PremiumRateService.PublisherQueryLanguageContextError}{Google Documentation for PublisherQueryLanguageContextError}
\describe{
 \item{extends ApiError}{see documentation for ApiError}
 \item{reason}{a PublisherQueryLanguageContextError.Reason - The error reason represented by an enum. This can take one of the following values:
   \itemize{
     \item{UNEXECUTABLE - Indicates that there was an error executing the PQL.}
     \item{UNKNOWN - The value returned if the actual value is not exposed by the requested API
version.}
   }
  }
}

\strong{PublisherQueryLanguageSyntaxError}

An error that occurs while parsing a PQL query contained in a Statement
object.
\href{https://developers.google.com/doubleclick-publishers/docs/reference/v201508/PremiumRateService.PublisherQueryLanguageSyntaxError}{Google Documentation for PublisherQueryLanguageSyntaxError}
\describe{
 \item{extends ApiError}{see documentation for ApiError}
 \item{reason}{a PublisherQueryLanguageSyntaxError.Reason - The error reason represented by an enum. This can take one of the following values:
   \itemize{
     \item{UNPARSABLE - Indicates that there was a PQL syntax error.}
     \item{UNKNOWN - The value returned if the actual value is not exposed by the requested API
version.}
   }
  }
}

\strong{QuotaError}

Describes a client-side error on which a user is attempting to perform an
action to which they have no quota remaining.
\href{https://developers.google.com/doubleclick-publishers/docs/reference/v201508/PremiumRateService.QuotaError}{Google Documentation for QuotaError}
\describe{
 \item{extends ApiError}{see documentation for ApiError}
 \item{reason}{a QuotaError.Reason - This can take one of the following values:
   \itemize{
     \item{EXCEEDED_QUOTA - The number of requests made per second is too high and has exceeded the
allowable limit. The recommended approach to handle this error is to wait
about 5 seconds and then retry the request. Note that this does not
guarantee the request will succeed. If it fails again, try increasing the
wait time.  Another way to mitigate this error is to limit requests to 2 per
second for Small Business networks, or 8 per second for Premium networks.
Once again this does not guarantee that every request will succeed, but may
help reduce the number of times you receive this error.}
     \item{UNKNOWN - The value returned if the actual value is not exposed by the requested API
version.}
   }
  }
}

\strong{RangeError}

A list of all errors associated with the Range constraint.
\href{https://developers.google.com/doubleclick-publishers/docs/reference/v201508/PremiumRateService.RangeError}{Google Documentation for RangeError}
\describe{
 \item{extends ApiError}{see documentation for ApiError}
 \item{reason}{a RangeError.Reason - This can take one of the following values:
   \itemize{
     \item{TOO_HIGH}
     \item{TOO_LOW}
     \item{UNKNOWN - The value returned if the actual value is not exposed by the requested API
version.}
   }
  }
}

\strong{RequiredError}

Errors due to missing required field.
\href{https://developers.google.com/doubleclick-publishers/docs/reference/v201508/PremiumRateService.RequiredError}{Google Documentation for RequiredError}
\describe{
 \item{extends ApiError}{see documentation for ApiError}
 \item{reason}{a RequiredError.Reason - The error reason represented by an enum. This can take one of the following values:
   \itemize{
     \item{REQUIRED - Missing required field.}
   }
  }
}

\strong{ServerError}

Errors related to the server.
\href{https://developers.google.com/doubleclick-publishers/docs/reference/v201508/PremiumRateService.ServerError}{Google Documentation for ServerError}
\describe{
 \item{extends ApiError}{see documentation for ApiError}
 \item{reason}{a ServerError.Reason - This can take one of the following values:
   \itemize{
     \item{SERVER_ERROR - Indicates that an unexpected error occured.}
     \item{SERVER_BUSY - Indicates that the server is currently experiencing a high load. Please wait
and try your request again.}
     \item{UNKNOWN - The value returned if the actual value is not exposed by the requested API
version.}
   }
  }
}

\strong{SetValue}

Contains a set of Value Values. May not contain duplicates.
\href{https://developers.google.com/doubleclick-publishers/docs/reference/v201508/PremiumRateService.SetValue}{Google Documentation for SetValue}
\describe{
 \item{extends Value}{see documentation for Value}
 \item{values}{a Value -  The values. They must all be the same type of Value and not contain
duplicates.}
}

\strong{SoapRequestHeader}

Represents the SOAP request header used by API requests.
\href{https://developers.google.com/doubleclick-publishers/docs/reference/v201508/PremiumRateService.SoapRequestHeader}{Google Documentation for SoapRequestHeader}
\describe{
 \item{networkCode}{a string -  The network code to use in the context of a request.}
 \item{applicationName}{a string -  The name of client library application.}
}

\strong{SoapResponseHeader}

Represents the SOAP request header used by API responses.
\href{https://developers.google.com/doubleclick-publishers/docs/reference/v201508/PremiumRateService.SoapResponseHeader}{Google Documentation for SoapResponseHeader}
\describe{
 \item{requestId}{a string}
 \item{responseTime}{a integer}
}

\strong{Statement}

Captures the WHERE, ORDER BY and LIMIT clauses of a PQL query. Statements
are typically used to retrieve objects of a predefined domain type, which
makes SELECT clause unnecessary.  An example query text might be "WHERE
status = 'ACTIVE' ORDER BY id LIMIT 30".  Statements support bind variables.
These are substitutes for literals and can be thought of as input parameters
to a PQL query.  An example of such a query might be "WHERE id = :idValue".
Statements also support use of the LIKE keyword. This provides partial and
wildcard string matching.  An example of such a query might be "WHERE name
LIKE 'startswith\%'".  If using an API version newer than V201010, the value
for the variable idValue must then be set with an object of type Value and
is one of NumberValue, TextValue or BooleanValue.  If using an API version
older than or equal to V201010, the value for the variable idValue must then
be set with an object of type Param and is one of DoubleParam, LongParam or
StringParam.
\href{https://developers.google.com/doubleclick-publishers/docs/reference/v201508/PremiumRateService.Statement}{Google Documentation for Statement}
\describe{
 \item{query}{a string -  Holds the query in PQL syntax. See link for how to form PQL.}
 \item{values}{a String_ValueMapEntry -  Holds keys and values for bind variables and their values. The key is the
name of the bind variable. The value is the literal value of the variable.
In the example "WHERE status = :bindStatus ORDER BY id LIMIT 30", the bind
variable, represented by :bindStatus is named bindStatus, which would also
be the parameter map key. The bind variable's value would be represented by
a parameter map value of type TextValue. The final result, for example,
would be an entry of "bindStatus" => StringParam("ACTIVE").}
}

\strong{StatementError}

An error that occurs while parsing Statement objects.
\href{https://developers.google.com/doubleclick-publishers/docs/reference/v201508/PremiumRateService.StatementError}{Google Documentation for StatementError}
\describe{
 \item{extends ApiError}{see documentation for ApiError}
 \item{reason}{a StatementError.Reason - The error reason represented by an enum. This can take one of the following values:
   \itemize{
     \item{VARIABLE_NOT_BOUND_TO_VALUE - A bind variable has not been bound to a value.}
     \item{UNKNOWN - The value returned if the actual value is not exposed by the requested API
version.}
   }
  }
}

\strong{String_ValueMapEntry}

This represents an entry in a map with a key of type String and value of
type Value.
\href{https://developers.google.com/doubleclick-publishers/docs/reference/v201508/PremiumRateService.String_ValueMapEntry}{Google Documentation for String_ValueMapEntry}
\describe{
 \item{key}{a string}
 \item{value}{a Value}
}

\strong{TextValue}

Contains a string value.
\href{https://developers.google.com/doubleclick-publishers/docs/reference/v201508/PremiumRateService.TextValue}{Google Documentation for TextValue}
\describe{
 \item{extends Value}{see documentation for Value}
 \item{value}{a string -  The string value.}
}

\strong{UnknownPremiumFeature}

The PremiumFeature returned if the actual feature is not exposed by the
requested API version.
\href{https://developers.google.com/doubleclick-publishers/docs/reference/v201508/PremiumRateService.UnknownPremiumFeature}{Google Documentation for UnknownPremiumFeature}
\describe{
 \item{extends PremiumFeature}{see documentation for PremiumFeature}
}

\strong{UserDomainPremiumFeature}

A premium feature applied to user domain targeting.
\href{https://developers.google.com/doubleclick-publishers/docs/reference/v201508/PremiumRateService.UserDomainPremiumFeature}{Google Documentation for UserDomainPremiumFeature}
\describe{
 \item{extends PremiumFeature}{see documentation for PremiumFeature}
}

\strong{Value}

Value represents a value.
\href{https://developers.google.com/doubleclick-publishers/docs/reference/v201508/PremiumRateService.Value}{Google Documentation for Value}
\describe{
}

\strong{VideoPositionPremiumFeature}

A premium feature applied to video position targeting.
\href{https://developers.google.com/doubleclick-publishers/docs/reference/v201508/PremiumRateService.VideoPositionPremiumFeature}{Google Documentation for VideoPositionPremiumFeature}
\describe{
 \item{extends PremiumFeature}{see documentation for PremiumFeature}
}
}

