% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProductPackageService.R
\name{dfp_ProductPackageService_object_factory}
\alias{dfp_ProductPackageService_object_factory}
\title{ProductPackageService}
\usage{
dfp_ProductPackageService_object_factory(obj_type, obj_data)
}
\arguments{
\item{obj_type}{a string from one of the object types described below}

\item{obj_data}{a \code{list} of \code{lists} or a \code{data.frame} with the required components to create
the obj_type specified. All input names must match the expected attributes for the object.
If submitting a list of attributes to create 1 object make sure to nest it inside top-level list, so that
it is a list of length 1 with 1 list of attributes to create the object.}
}
\value{
a string of formatted XML that can be inserted into a SOAP request body to the API
}
\description{
Provides methods for updating and retrieving ProductPackage objects.
A
ProductPackage represents a group of products which will be sold together.
}
\details{
To use this service, you need to have the new sales management solution
enabled on your network. If you do not see a "Sales" tab in <a
href="https://www.google.com/dfp">DoubleClick for Publishers (DFP)</a>, you
will not be able to use this service.

ProductPackageService Object Factory

A function to create a variety of objects that are part of the ProductPackageService
Note: not all objects here are inputs to the functions of this service so 
please consult each function for the required inputs

Below is a list of objects and their required components to be created with this function:

\strong{ActivateProductPackages}

The action used to activate product packages.
\href{https://developers.google.com/doubleclick-publishers/docs/reference/v201508/ProductPackageService.ActivateProductPackages}{Google Documentation for ActivateProductPackages}
\describe{
 \item{extends ProductPackageAction}{see documentation for ProductPackageAction}
}

\strong{ApiError}

The API error base class that provides details about an error that occurred
while processing a service request. The OGNL field path is provided for
parsers to identify the request data element that may have caused the error.
\href{https://developers.google.com/doubleclick-publishers/docs/reference/v201508/ProductPackageService.ApiError}{Google Documentation for ApiError}
\describe{
 \item{fieldPath}{a string -  The OGNL field path to identify cause of error.}
 \item{trigger}{a string -  The data that caused the error.}
 \item{errorString}{a string -  A simple string representation of the error and reason.}
}

\strong{ApiException}

Exception class for holding a list of service errors.
\href{https://developers.google.com/doubleclick-publishers/docs/reference/v201508/ProductPackageService.ApiException}{Google Documentation for ApiException}
\describe{
 \item{message}{a string (inherited from ApplicationException) -  Error message.}
 \item{errors}{a ApiError -  List of errors.}
}

\strong{ApiVersionError}

Errors related to the usage of API versions.
\href{https://developers.google.com/doubleclick-publishers/docs/reference/v201508/ProductPackageService.ApiVersionError}{Google Documentation for ApiVersionError}
\describe{
 \item{extends ApiError}{see documentation for ApiError}
 \item{reason}{a ApiVersionError.Reason - This can take one of the following values:
   \itemize{
     \item{UPDATE_TO_NEWER_VERSION - Indicates that the operation is not allowed in the version the request was
made in.}
     \item{UNKNOWN - The value returned if the actual value is not exposed by the requested API
version.}
   }
  }
}

\strong{ApplicationException}

Base class for exceptions.
\href{https://developers.google.com/doubleclick-publishers/docs/reference/v201508/ProductPackageService.ApplicationException}{Google Documentation for ApplicationException}
\describe{
 \item{message}{a string -  Error message.}
}

\strong{ArchiveProductPackages}

The action used to archive product packages.
\href{https://developers.google.com/doubleclick-publishers/docs/reference/v201508/ProductPackageService.ArchiveProductPackages}{Google Documentation for ArchiveProductPackages}
\describe{
 \item{extends ProductPackageAction}{see documentation for ProductPackageAction}
}

\strong{AuthenticationError}

An error for an exception that occurred when authenticating.
\href{https://developers.google.com/doubleclick-publishers/docs/reference/v201508/ProductPackageService.AuthenticationError}{Google Documentation for AuthenticationError}
\describe{
 \item{extends ApiError}{see documentation for ApiError}
 \item{reason}{a AuthenticationError.Reason - This can take one of the following values:
   \itemize{
     \item{AMBIGUOUS_SOAP_REQUEST_HEADER - The SOAP message contains a request header with an ambiguous definition of
the authentication header fields. This means either the authToken and
oAuthToken fields were both null or both were specified. Exactly one value
should be specified with each request.}
     \item{INVALID_EMAIL - The login provided is invalid.}
     \item{AUTHENTICATION_FAILED - Tried to authenticate with provided information, but failed.}
     \item{INVALID_OAUTH_SIGNATURE - The OAuth provided is invalid.}
     \item{INVALID_SERVICE - The specified service to use was not recognized.}
     \item{MISSING_SOAP_REQUEST_HEADER - The SOAP message is missing a request header with an authToken and optional
networkCode.}
     \item{MISSING_AUTHENTICATION_HTTP_HEADER - The HTTP request is missing a request header with an authToken}
     \item{MISSING_AUTHENTICATION - The request is missing an authToken}
     \item{NOT_WHITELISTED_FOR_API_ACCESS - The customer is not whitelisted for API access.}
     \item{NO_NETWORKS_TO_ACCESS - The user is not associated with any network.}
     \item{NETWORK_NOT_FOUND - No network for the given networkCode was found.}
     \item{NETWORK_CODE_REQUIRED - The user has access to more than one network, but did not provide a
networkCode.}
     \item{CONNECTION_ERROR - An error happened on the server side during connection to authentication
service.}
     \item{GOOGLE_ACCOUNT_ALREADY_ASSOCIATED_WITH_NETWORK - The user tried to create a test network using an account that already is
associated with a network.}
     \item{UNKNOWN - The value returned if the actual value is not exposed by the requested API
version.}
   }
  }
}

\strong{BaseCustomFieldValue}

The value of a CustomField for a particular entity.
\href{https://developers.google.com/doubleclick-publishers/docs/reference/v201508/ProductPackageService.BaseCustomFieldValue}{Google Documentation for BaseCustomFieldValue}
\describe{
 \item{customFieldId}{a integer -  Id of the CustomField to which this value belongs. This attribute is
required.}
}

\strong{BaseRateError}

An error having to do with BaseRate.
\href{https://developers.google.com/doubleclick-publishers/docs/reference/v201508/ProductPackageService.BaseRateError}{Google Documentation for BaseRateError}
\describe{
 \item{extends ApiError}{see documentation for ApiError}
 \item{reason}{a BaseRateError.Reason - This can take one of the following values:
   \itemize{
     \item{CANNOT_QUERY_BOTH_PRODUCT_TEMPLATE_ID_AND_PRODUCT_ID - The PQL statement contains both productTemplateId and productId fields.}
     \item{INVALID_CURRENCY_CODE - The currency code is invalid.}
     \item{PRODUCT_TEMPLATE_ARCHIVED - Cannot create or activate a base rate if the product template is archived.}
     \item{UNSUPPORTED_OPERATION - Indicates that the requested operation is not supported.}
     \item{PRODUCT_TEMPLATE_BASE_RATE_NOT_FOUND - Cannot create a base rate to a product if its product template does not have
a base rate on this rate card.}
     \item{PRODUCT_BASE_RATE_EXISTS - Cannot delete a base rate on a product template if its products still have
base rates on this rate card.}
     \item{UNKNOWN - The value returned if the actual value is not exposed by the requested API
version.}
   }
  }
}

\strong{BooleanValue}

Contains a boolean value.
\href{https://developers.google.com/doubleclick-publishers/docs/reference/v201508/ProductPackageService.BooleanValue}{Google Documentation for BooleanValue}
\describe{
 \item{extends Value}{see documentation for Value}
 \item{value}{a boolean -  The boolean value.}
}

\strong{CollectionSizeError}

Error for the size of the collection being too large
\href{https://developers.google.com/doubleclick-publishers/docs/reference/v201508/ProductPackageService.CollectionSizeError}{Google Documentation for CollectionSizeError}
\describe{
 \item{extends ApiError}{see documentation for ApiError}
 \item{reason}{a CollectionSizeError.Reason - This can take one of the following values:
   \itemize{
     \item{TOO_LARGE}
     \item{UNKNOWN - The value returned if the actual value is not exposed by the requested API
version.}
   }
  }
}

\strong{CommonError}

A place for common errors that can be used across services.
\href{https://developers.google.com/doubleclick-publishers/docs/reference/v201508/ProductPackageService.CommonError}{Google Documentation for CommonError}
\describe{
 \item{extends ApiError}{see documentation for ApiError}
 \item{reason}{a CommonError.Reason - This can take one of the following values:
   \itemize{
     \item{NOT_FOUND - Indicates that an attempt was made to retrieve an entity that does not
exist.}
     \item{ALREADY_EXISTS - Indicates that an attempt was made to create an entity that already exists.}
     \item{DUPLICATE_OBJECT - Indicates that two elements in the collection were identical.}
     \item{CANNOT_UPDATE - Indicates that an attempt was made to change an immutable field.}
     \item{CONCURRENT_MODIFICATION - Indicates that another request attempted to update the same data in the same
network at about the same time. Please wait and try the request again.}
     \item{UNKNOWN - The value returned if the actual value is not exposed by the requested API
version.}
   }
  }
}

\strong{CustomFieldValue}

The value of a CustomField that does not have a CustomField dataType of
CustomFieldDataType DROP_DOWN.
\href{https://developers.google.com/doubleclick-publishers/docs/reference/v201508/ProductPackageService.CustomFieldValue}{Google Documentation for CustomFieldValue}
\describe{
 \item{customFieldId}{a integer (inherited from BaseCustomFieldValue) -  Id of the CustomField to which this value belongs. This attribute is
required.}
 \item{value}{a Value -  The value for this field. The appropriate type of Value is determined by the
CustomField dataType of the CustomField that this conforms to.}
}

\strong{Date}

Represents a date.
\href{https://developers.google.com/doubleclick-publishers/docs/reference/v201508/ProductPackageService.Date}{Google Documentation for Date}
\describe{
 \item{year}{a integer -  Year (e.g., 2009)}
 \item{month}{a integer -  Month (1..12)}
 \item{day}{a integer -  Day (1..31)}
}

\strong{DateTime}

Represents a date combined with the time of day.
\href{https://developers.google.com/doubleclick-publishers/docs/reference/v201508/ProductPackageService.DateTime}{Google Documentation for DateTime}
\describe{
 \item{date}{a Date}
 \item{hour}{a integer}
 \item{minute}{a integer}
 \item{second}{a integer}
 \item{timeZoneID}{a string}
}

\strong{DateTimeValue}

Contains a date-time value.
\href{https://developers.google.com/doubleclick-publishers/docs/reference/v201508/ProductPackageService.DateTimeValue}{Google Documentation for DateTimeValue}
\describe{
 \item{extends Value}{see documentation for Value}
 \item{value}{a DateTime -  The DateTime value.}
}

\strong{DateValue}

Contains a date value.
\href{https://developers.google.com/doubleclick-publishers/docs/reference/v201508/ProductPackageService.DateValue}{Google Documentation for DateValue}
\describe{
 \item{extends Value}{see documentation for Value}
 \item{value}{a Date -  The Date value.}
}

\strong{DeactivateProductPackages}

The action used to de-activate product packages.
\href{https://developers.google.com/doubleclick-publishers/docs/reference/v201508/ProductPackageService.DeactivateProductPackages}{Google Documentation for DeactivateProductPackages}
\describe{
 \item{extends ProductPackageAction}{see documentation for ProductPackageAction}
}

\strong{DropDownCustomFieldValue}

A CustomFieldValue for a CustomField that has a CustomField dataType of
CustomFieldDataType DROP_DOWN
\href{https://developers.google.com/doubleclick-publishers/docs/reference/v201508/ProductPackageService.DropDownCustomFieldValue}{Google Documentation for DropDownCustomFieldValue}
\describe{
 \item{customFieldId}{a integer (inherited from BaseCustomFieldValue) -  Id of the CustomField to which this value belongs. This attribute is
required.}
 \item{customFieldOptionId}{a integer -  The CustomFieldOption id ID of the CustomFieldOption for this value.}
}

\strong{EntityChildrenLimitReachedError}

Lists errors relating to having too many children on an entity.
\href{https://developers.google.com/doubleclick-publishers/docs/reference/v201508/ProductPackageService.EntityChildrenLimitReachedError}{Google Documentation for EntityChildrenLimitReachedError}
\describe{
 \item{extends ApiError}{see documentation for ApiError}
 \item{reason}{a EntityChildrenLimitReachedError.Reason - This can take one of the following values:
   \itemize{
     \item{LINE_ITEM_LIMIT_FOR_ORDER_REACHED - The number of line items on the order exceeds the max number of line items
allowed per order in the network.}
     \item{CREATIVE_ASSOCIATION_LIMIT_FOR_LINE_ITEM_REACHED - The number of creatives associated with the line item exceeds the max number
of creatives allowed to be associated with a line item in the network.}
     \item{AD_UNIT_LIMIT_FOR_PLACEMENT_REACHED - The number of ad units on the placement exceeds the max number of ad units
allowed per placement in the network.}
     \item{TARGETING_EXPRESSION_LIMIT_FOR_LINE_ITEM_REACHED - The number of targeting expressions on the line item exceeds the max number
of targeting expressions allowed per line item in the network.}
     \item{CUSTOM_TARGETING_VALUES_FOR_KEY_LIMIT_REACHED - The number of custom targeting values for the free-form or predefined custom
targeting key exceeds the max number allowed.}
     \item{TARGETING_EXPRESSION_LIMIT_FOR_CREATIVES_ON_LINE_ITEM_REACHED - The total number of targeting expressions on the creatives for the line item
exceeds the max number allowed per line item in the network.}
     \item{ATTACHMENT_LIMIT_FOR_PROPOSAL_REACHED - The number of attachments added to the proposal exceeds the max number
allowed per proposal in the network.}
     \item{PROPOSAL_LINE_ITEM_LIMIT_FOR_PROPOSAL_REACHED - The number of proposal line items on the proposal exceeds the max number
allowed per proposal in the network.}
     \item{PRODUCT_LIMIT_FOR_PRODUCT_PACKAGE_REACHED - The number of product package items on the product package exceeds the max
number allowed per product package in the network.}
     \item{PRODUCT_TEMPLATE_AND_PRODUCT_BASE_RATE_LIMIT_FOR_RATE_CARD_REACHED - The number of product template and product base rates on the rate card
(including excluded product base rates) exceeds the max number allowed per
rate card in the network.}
     \item{PRODUCT_PACKAGE_ITEM_BASE_RATE_LIMIT_FOR_RATE_CARD_REACHED - The number of product package item base rates on the rate card exceeds the
max number allowed per rate card in the network.}
     \item{PREMIUM_LIMIT_FOR_RATE_CARD_REACHED - The number of premiums of the rate card exceeds the max number allowed per
rate card in the network.}
     \item{AD_UNIT_LIMIT_FOR_AD_EXCLUSION_RULE_TARGETING_REACHED - The number of ad units on AdExclusionRule inventoryTargeting exceeds the max
number of ad units allowed per ad exclusion rule inventory targeting in the
network.}
     \item{UNKNOWN - The value returned if the actual value is not exposed by the requested API
version.}
   }
  }
}

\strong{EntityLimitReachedError}

An error that occurs when creating an entity if the limit on the number of
allowed entities for a network has already been reached.
\href{https://developers.google.com/doubleclick-publishers/docs/reference/v201508/ProductPackageService.EntityLimitReachedError}{Google Documentation for EntityLimitReachedError}
\describe{
 \item{extends ApiError}{see documentation for ApiError}
 \item{reason}{a EntityLimitReachedError.Reason - This can take one of the following values:
   \itemize{
     \item{CUSTOM_TARGETING_VALUES_LIMIT_REACHED - The number of custom targeting values exceeds the max number allowed in the
network.}
     \item{AD_EXCLUSION_RULES_LIMIT_REACHED - The number of ad exclusion rules exceeds the max number allowed in the
network.}
     \item{FIRST_PARTY_AUDIENCE_SEGMENTS_LIMIT_REACHED - The number of first party audience segments exceeds the max number allowed
in the network.}
     \item{PLACEMENTS_LIMIT_REACHED - The number of active placements exceeds the max number allowed in the
network.}
     \item{LINE_ITEMS_LIMIT_REACHED - The number of line items excceeds the max number allowed in the network.}
     \item{UNKNOWN - The value returned if the actual value is not exposed by the requested API
version.}
   }
  }
}

\strong{FeatureError}

Errors related to feature management. If you attempt using a feature that is
not available to the current network you'll receive a FeatureError with the
missing feature as the trigger.
\href{https://developers.google.com/doubleclick-publishers/docs/reference/v201508/ProductPackageService.FeatureError}{Google Documentation for FeatureError}
\describe{
 \item{extends ApiError}{see documentation for ApiError}
 \item{reason}{a FeatureError.Reason - This can take one of the following values:
   \itemize{
     \item{MISSING_FEATURE - A feature is being used that is not enabled on the current network.}
     \item{UNKNOWN - The value returned if the actual value is not exposed by the requested API
version.}
   }
  }
}

\strong{InternalApiError}

Indicates that a server-side error has occured. InternalApiErrors are
generally not the result of an invalid request or message sent by the
client.
\href{https://developers.google.com/doubleclick-publishers/docs/reference/v201508/ProductPackageService.InternalApiError}{Google Documentation for InternalApiError}
\describe{
 \item{extends ApiError}{see documentation for ApiError}
 \item{reason}{a InternalApiError.Reason - The error reason represented by an enum. This can take one of the following values:
   \itemize{
     \item{UNEXPECTED_INTERNAL_API_ERROR - API encountered an unexpected internal error.}
     \item{TRANSIENT_ERROR - A temporary error occurred during the request. Please retry.}
     \item{UNKNOWN - The cause of the error is not known or only defined in newer versions.}
     \item{DOWNTIME - The API is currently unavailable for a planned downtime.}
   }
  }
}

\strong{NotNullError}

Caused by supplying a null value for an attribute that cannot be null.
\href{https://developers.google.com/doubleclick-publishers/docs/reference/v201508/ProductPackageService.NotNullError}{Google Documentation for NotNullError}
\describe{
 \item{extends ApiError}{see documentation for ApiError}
 \item{reason}{a NotNullError.Reason - The error reason represented by an enum. This can take one of the following values:
   \itemize{
     \item{ARG1_NULL - Assuming that a method will not have more than 3 arguments, if it does,
return NULL}
     \item{ARG2_NULL}
     \item{ARG3_NULL}
     \item{NULL}
     \item{UNKNOWN - The value returned if the actual value is not exposed by the requested API
version.}
   }
  }
}

\strong{NullError}

Errors associated with violation of a NOT NULL check.
\href{https://developers.google.com/doubleclick-publishers/docs/reference/v201508/ProductPackageService.NullError}{Google Documentation for NullError}
\describe{
 \item{extends ApiError}{see documentation for ApiError}
 \item{reason}{a NullError.Reason - The error reason represented by an enum. This can take one of the following values:
   \itemize{
     \item{NULL_CONTENT - Specified list/container must not contain any null elements}
   }
  }
}

\strong{NumberValue}

Contains a numeric value.
\href{https://developers.google.com/doubleclick-publishers/docs/reference/v201508/ProductPackageService.NumberValue}{Google Documentation for NumberValue}
\describe{
 \item{extends Value}{see documentation for Value}
 \item{value}{a string -  The numeric value represented as a string.}
}

\strong{ObjectValue}

Contains an object value.  This object is experimental! ObjectValue is an
experimental, innovative, and rapidly changing new feature for DFP.
Unfortunately, being on the bleeding edge means that we may make
backwards-incompatible changes to ObjectValue. We will inform the community
when this feature is no longer experimental.
\href{https://developers.google.com/doubleclick-publishers/docs/reference/v201508/ProductPackageService.ObjectValue}{Google Documentation for ObjectValue}
\describe{
 \item{extends Value}{see documentation for Value}
}

\strong{ParseError}

Lists errors related to parsing.
\href{https://developers.google.com/doubleclick-publishers/docs/reference/v201508/ProductPackageService.ParseError}{Google Documentation for ParseError}
\describe{
 \item{extends ApiError}{see documentation for ApiError}
 \item{reason}{a ParseError.Reason - The error reason represented by an enum. This can take one of the following values:
   \itemize{
     \item{UNPARSABLE - Indicates an error in parsing an attribute.}
     \item{UNKNOWN - The value returned if the actual value is not exposed by the requested API
version.}
   }
  }
}

\strong{PermissionError}

Errors related to incorrect permission.
\href{https://developers.google.com/doubleclick-publishers/docs/reference/v201508/ProductPackageService.PermissionError}{Google Documentation for PermissionError}
\describe{
 \item{extends ApiError}{see documentation for ApiError}
 \item{reason}{a PermissionError.Reason - This can take one of the following values:
   \itemize{
     \item{PERMISSION_DENIED - User does not have the required permission for the request.}
     \item{UNKNOWN - The value returned if the actual value is not exposed by the requested API
version.}
   }
  }
}

\strong{ProductPackage}

A ProductPackage represents a group of products which will be sold together.
\href{https://developers.google.com/doubleclick-publishers/docs/reference/v201508/ProductPackageService.ProductPackage}{Google Documentation for ProductPackage}
\describe{
 \item{id}{a integer -  Uniquely identifies the ProductPackage. This attribute is read-only and is
assigned by Google when a ProductPackage is created.}
 \item{name}{a string -  The name of the ProductPackage. This attribute is required and has maximum
length of 255 characters.}
 \item{notes}{a string -  The notes of the ProductPackage. This attribute has maximum length of 511
characters. This attribute is optional.}
 \item{status}{a ProductPackageStatus - The status of the ProductPackage. This attribute is read-only and is
assigned by Google. This can take one of the following values:
   \itemize{
     \item{ACTIVE}
     \item{INACTIVE}
     \item{UNKNOWN - The value returned if the actual value is not exposed by the requested API
version.}
   }
  }
 \item{isArchived}{a boolean -  The archival status of the ProductPackage. This attribute is read-only and
is assigned by Google.}
 \item{rateCardIds}{a integer -  RateCard IDs associated with the ProductPackage. This attribute is optional.}
 \item{customFieldValues}{a BaseCustomFieldValue -  The CustomFieldValue objects associated with this ProductPackage. This
attribute is optional.}
 \item{lastModifiedDateTime}{a DateTime -  The date and time this ProductPackage was last modified. This attribute is
read-only and is assigned by Google when a ProductPackage is updated.}
}

\strong{ProductPackageAction}

Represents the actions that can be performed on product packages. May take one of the actions listed below.
\href{https://developers.google.com/doubleclick-publishers/docs/reference/v201508/ProductPackageService.ProductPackageAction}{Google Documentation for ProductPackageAction}
\describe{
 \item{ActivateProductPackages}{The action used to activate product packages.}
 \item{ArchiveProductPackages}{The action used to archive product packages.}
 \item{DeactivateProductPackages}{The action used to de-activate product packages.}
 \item{UnarchiveProductPackages}{The action used to un-archive product packages.}
}

\strong{ProductPackageActionError}

Lists all errors for executing operations on product packages.
\href{https://developers.google.com/doubleclick-publishers/docs/reference/v201508/ProductPackageService.ProductPackageActionError}{Google Documentation for ProductPackageActionError}
\describe{
 \item{extends ApiError}{see documentation for ApiError}
 \item{reason}{a ProductPackageActionError.Reason - The error reason represented by an enum. This can take one of the following values:
   \itemize{
     \item{ACTIVATE_WITH_INACTIVE_MANDATORY_PRODUCT - The activate operation is not applicable if there's any inactive mandatory
product.}
     \item{NOT_APPLICABLE - The operation is not applicable to the current state.}
     \item{UNKNOWN - The value returned if the actual value is not exposed by the requested API
version.}
   }
  }
}

\strong{ProductPackageItemError}

Lists all errors associated with product package items.
\href{https://developers.google.com/doubleclick-publishers/docs/reference/v201508/ProductPackageService.ProductPackageItemError}{Google Documentation for ProductPackageItemError}
\describe{
 \item{extends ApiError}{see documentation for ApiError}
 \item{reason}{a ProductPackageItemError.Reason - The error reason represented by an enum. This can take one of the following values:
   \itemize{
     \item{ARCHIVED_PRODUCT_NOT_ALLOWED - Add a archived product to product package is not allowed.}
     \item{INACTIVE_MANDATORY_PRODUCT_NOT_ALLOWED - Inactive mandatory product is not allowed in active product package.}
     \item{UNKNOWN - The value returned if the actual value is not exposed by the requested API
version.}
   }
  }
}

\strong{ProductPackagePage}

Captures a page of ProductPackageDto objects.
\href{https://developers.google.com/doubleclick-publishers/docs/reference/v201508/ProductPackageService.ProductPackagePage}{Google Documentation for ProductPackagePage}
\describe{
 \item{totalResultSetSize}{a integer -  The size of the total result set to which this page belongs.}
 \item{startIndex}{a integer -  The absolute index in the total result set on which this page begins.}
 \item{results}{a ProductPackage -  The collection of product packages contained within this page.}
}

\strong{ProductPackageRateCardAssociationError}

Lists all errors associated with product packages rate card associations.
\href{https://developers.google.com/doubleclick-publishers/docs/reference/v201508/ProductPackageService.ProductPackageRateCardAssociationError}{Google Documentation for ProductPackageRateCardAssociationError}
\describe{
 \item{extends ApiError}{see documentation for ApiError}
 \item{reason}{a ProductPackageRateCardAssociationError.Reason - The error reason represented by an enum. This can take one of the following values:
   \itemize{
     \item{INVALID_RATE_CARD_ID - The id of associated rate card is invalid or invisible to user.}
     \item{INVALID_PRODUCT_PACKAGE_ID - The id of associated package is invalid.}
     \item{UNKNOWN - The value returned if the actual value is not exposed by the requested API
version.}
   }
  }
}

\strong{PublisherQueryLanguageContextError}

An error that occurs while executing a PQL query contained in a Statement
object.
\href{https://developers.google.com/doubleclick-publishers/docs/reference/v201508/ProductPackageService.PublisherQueryLanguageContextError}{Google Documentation for PublisherQueryLanguageContextError}
\describe{
 \item{extends ApiError}{see documentation for ApiError}
 \item{reason}{a PublisherQueryLanguageContextError.Reason - The error reason represented by an enum. This can take one of the following values:
   \itemize{
     \item{UNEXECUTABLE - Indicates that there was an error executing the PQL.}
     \item{UNKNOWN - The value returned if the actual value is not exposed by the requested API
version.}
   }
  }
}

\strong{PublisherQueryLanguageSyntaxError}

An error that occurs while parsing a PQL query contained in a Statement
object.
\href{https://developers.google.com/doubleclick-publishers/docs/reference/v201508/ProductPackageService.PublisherQueryLanguageSyntaxError}{Google Documentation for PublisherQueryLanguageSyntaxError}
\describe{
 \item{extends ApiError}{see documentation for ApiError}
 \item{reason}{a PublisherQueryLanguageSyntaxError.Reason - The error reason represented by an enum. This can take one of the following values:
   \itemize{
     \item{UNPARSABLE - Indicates that there was a PQL syntax error.}
     \item{UNKNOWN - The value returned if the actual value is not exposed by the requested API
version.}
   }
  }
}

\strong{QuotaError}

Describes a client-side error on which a user is attempting to perform an
action to which they have no quota remaining.
\href{https://developers.google.com/doubleclick-publishers/docs/reference/v201508/ProductPackageService.QuotaError}{Google Documentation for QuotaError}
\describe{
 \item{extends ApiError}{see documentation for ApiError}
 \item{reason}{a QuotaError.Reason - This can take one of the following values:
   \itemize{
     \item{EXCEEDED_QUOTA - The number of requests made per second is too high and has exceeded the
allowable limit. The recommended approach to handle this error is to wait
about 5 seconds and then retry the request. Note that this does not
guarantee the request will succeed. If it fails again, try increasing the
wait time.  Another way to mitigate this error is to limit requests to 2 per
second for Small Business networks, or 8 per second for Premium networks.
Once again this does not guarantee that every request will succeed, but may
help reduce the number of times you receive this error.}
     \item{UNKNOWN - The value returned if the actual value is not exposed by the requested API
version.}
   }
  }
}

\strong{RangeError}

A list of all errors associated with the Range constraint.
\href{https://developers.google.com/doubleclick-publishers/docs/reference/v201508/ProductPackageService.RangeError}{Google Documentation for RangeError}
\describe{
 \item{extends ApiError}{see documentation for ApiError}
 \item{reason}{a RangeError.Reason - This can take one of the following values:
   \itemize{
     \item{TOO_HIGH}
     \item{TOO_LOW}
     \item{UNKNOWN - The value returned if the actual value is not exposed by the requested API
version.}
   }
  }
}

\strong{ServerError}

Errors related to the server.
\href{https://developers.google.com/doubleclick-publishers/docs/reference/v201508/ProductPackageService.ServerError}{Google Documentation for ServerError}
\describe{
 \item{extends ApiError}{see documentation for ApiError}
 \item{reason}{a ServerError.Reason - This can take one of the following values:
   \itemize{
     \item{SERVER_ERROR - Indicates that an unexpected error occured.}
     \item{SERVER_BUSY - Indicates that the server is currently experiencing a high load. Please wait
and try your request again.}
     \item{UNKNOWN - The value returned if the actual value is not exposed by the requested API
version.}
   }
  }
}

\strong{SetValue}

Contains a set of Value Values. May not contain duplicates.
\href{https://developers.google.com/doubleclick-publishers/docs/reference/v201508/ProductPackageService.SetValue}{Google Documentation for SetValue}
\describe{
 \item{extends Value}{see documentation for Value}
 \item{values}{a Value -  The values. They must all be the same type of Value and not contain
duplicates.}
}

\strong{SoapRequestHeader}

Represents the SOAP request header used by API requests.
\href{https://developers.google.com/doubleclick-publishers/docs/reference/v201508/ProductPackageService.SoapRequestHeader}{Google Documentation for SoapRequestHeader}
\describe{
 \item{networkCode}{a string -  The network code to use in the context of a request.}
 \item{applicationName}{a string -  The name of client library application.}
}

\strong{SoapResponseHeader}

Represents the SOAP request header used by API responses.
\href{https://developers.google.com/doubleclick-publishers/docs/reference/v201508/ProductPackageService.SoapResponseHeader}{Google Documentation for SoapResponseHeader}
\describe{
 \item{requestId}{a string}
 \item{responseTime}{a integer}
}

\strong{Statement}

Captures the WHERE, ORDER BY and LIMIT clauses of a PQL query. Statements
are typically used to retrieve objects of a predefined domain type, which
makes SELECT clause unnecessary.  An example query text might be "WHERE
status = 'ACTIVE' ORDER BY id LIMIT 30".  Statements support bind variables.
These are substitutes for literals and can be thought of as input parameters
to a PQL query.  An example of such a query might be "WHERE id = :idValue".
Statements also support use of the LIKE keyword. This provides partial and
wildcard string matching.  An example of such a query might be "WHERE name
LIKE 'startswith\%'".  If using an API version newer than V201010, the value
for the variable idValue must then be set with an object of type Value and
is one of NumberValue, TextValue or BooleanValue.  If using an API version
older than or equal to V201010, the value for the variable idValue must then
be set with an object of type Param and is one of DoubleParam, LongParam or
StringParam.
\href{https://developers.google.com/doubleclick-publishers/docs/reference/v201508/ProductPackageService.Statement}{Google Documentation for Statement}
\describe{
 \item{query}{a string -  Holds the query in PQL syntax. See link for how to form PQL.}
 \item{values}{a String_ValueMapEntry -  Holds keys and values for bind variables and their values. The key is the
name of the bind variable. The value is the literal value of the variable.
In the example "WHERE status = :bindStatus ORDER BY id LIMIT 30", the bind
variable, represented by :bindStatus is named bindStatus, which would also
be the parameter map key. The bind variable's value would be represented by
a parameter map value of type TextValue. The final result, for example,
would be an entry of "bindStatus" => StringParam("ACTIVE").}
}

\strong{StatementError}

An error that occurs while parsing Statement objects.
\href{https://developers.google.com/doubleclick-publishers/docs/reference/v201508/ProductPackageService.StatementError}{Google Documentation for StatementError}
\describe{
 \item{extends ApiError}{see documentation for ApiError}
 \item{reason}{a StatementError.Reason - The error reason represented by an enum. This can take one of the following values:
   \itemize{
     \item{VARIABLE_NOT_BOUND_TO_VALUE - A bind variable has not been bound to a value.}
     \item{UNKNOWN - The value returned if the actual value is not exposed by the requested API
version.}
   }
  }
}

\strong{String_ValueMapEntry}

This represents an entry in a map with a key of type String and value of
type Value.
\href{https://developers.google.com/doubleclick-publishers/docs/reference/v201508/ProductPackageService.String_ValueMapEntry}{Google Documentation for String_ValueMapEntry}
\describe{
 \item{key}{a string}
 \item{value}{a Value}
}

\strong{StringLengthError}

Errors for Strings which do not meet given length constraints.
\href{https://developers.google.com/doubleclick-publishers/docs/reference/v201508/ProductPackageService.StringLengthError}{Google Documentation for StringLengthError}
\describe{
 \item{extends ApiError}{see documentation for ApiError}
 \item{reason}{a StringLengthError.Reason - This can take one of the following values:
   \itemize{
     \item{TOO_LONG}
     \item{TOO_SHORT}
     \item{UNKNOWN - The value returned if the actual value is not exposed by the requested API
version.}
   }
  }
}

\strong{TextValue}

Contains a string value.
\href{https://developers.google.com/doubleclick-publishers/docs/reference/v201508/ProductPackageService.TextValue}{Google Documentation for TextValue}
\describe{
 \item{extends Value}{see documentation for Value}
 \item{value}{a string -  The string value.}
}

\strong{UnarchiveProductPackages}

The action used to un-archive product packages.
\href{https://developers.google.com/doubleclick-publishers/docs/reference/v201508/ProductPackageService.UnarchiveProductPackages}{Google Documentation for UnarchiveProductPackages}
\describe{
 \item{extends ProductPackageAction}{see documentation for ProductPackageAction}
}

\strong{UniqueError}

An error for a field which must satisfy a uniqueness constraint
\href{https://developers.google.com/doubleclick-publishers/docs/reference/v201508/ProductPackageService.UniqueError}{Google Documentation for UniqueError}
\describe{
 \item{extends ApiError}{see documentation for ApiError}
}

\strong{UpdateResult}

Represents the result of performing an action on objects.
\href{https://developers.google.com/doubleclick-publishers/docs/reference/v201508/ProductPackageService.UpdateResult}{Google Documentation for UpdateResult}
\describe{
 \item{numChanges}{a integer -  The number of objects that were changed as a result of performing the
action.}
}

\strong{Value}

Value represents a value.
\href{https://developers.google.com/doubleclick-publishers/docs/reference/v201508/ProductPackageService.Value}{Google Documentation for Value}
\describe{
}
}

