% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfp_utils.R
\name{dfp_report_url_to_dataframe}
\alias{dfp_report_url_to_dataframe}
\title{Take report URL and convert to data.frame}
\usage{
dfp_report_url_to_dataframe(report_url, exportFormat='CSV_DUMP')
}
\arguments{
\item{report_url}{a URL character string returned from the 
function \link{dfp_getReportDownloadURL}}

\item{exportFormat}{a character string naming what type of exportFormat was 
provided to \link{dfp_getReportDownloadURL}. This is used to determine how to parse the results.}
}
\value{
a \code{data.frame} of report results from the specified URL
}
\description{
Receive a URL (usually from the ReportService) and 
download data from that URL. Currently, the exportFormat
must have been set for CSV_DUMP
}
