% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PublisherQueryLanguageService.R
\name{dfp_select}
\alias{dfp_select}
\title{select}
\usage{
dfp_select()
}
\value{
a \code{list} containing all the elements of a selectResponse
}
\description{
Retrieves rows of data that satisfy the given Statement query from the system. specify what data needs to returned
}

