% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfp_utils.R
\name{dfp_set_above_attributes}
\alias{dfp_set_above_attributes}
\title{Reformat parsed list to resubmit}
\usage{
dfp_set_above_attributes(list)
}
\arguments{
\item{list}{a \code{list} with attributes to be placed 
back appropriately in the hierarchy}
}
\value{
a \code{list} formatted with attributes at 
the top of the respective list they came from
}
\description{
Receive a list (usually from the API service that was parsed and 
and contains .attr nodes) and push any .attr nodes to top
of their respective list element so the SOAP request can 
be formatted with the proper types
}
\keyword{internal}

