% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ActivityService.R
\name{dfp_updateActivities}
\alias{dfp_updateActivities}
\title{updateActivities}
\usage{
dfp_updateActivities(request_data)
}
\arguments{
\item{request_data}{a \code{list} or \code{data.frame} of data elements
to be formatted for a SOAP request (XML format, but passed as character string)}
}
\value{
a \code{list} containing all the elements of a updateActivitiesResponse
}
\description{
Updates the specified Activity objects.
}
\seealso{
\href{https://developers.google.com/doubleclick-publishers/docs/reference/v201508/ActivityService#updateActivities}{Google Documentation for updateActivities}
}

