% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ActivityGroupService.R
\name{dfp_updateActivityGroups}
\alias{dfp_updateActivityGroups}
\title{updateActivityGroups}
\usage{
dfp_updateActivityGroups(request_data)
}
\arguments{
\item{request_data}{a \code{list} or \code{data.frame} of data elements
to be formatted for a SOAP request (XML format, but passed as character string)}
}
\value{
a \code{list} containing all the elements of a updateActivityGroupsResponse
}
\description{
Updates the specified ActivityGroup objects.
}
\seealso{
\href{https://developers.google.com/doubleclick-publishers/docs/reference/v201508/ActivityGroupService#updateActivityGroups}{Google Documentation for updateActivityGroups}
}

