% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AdRuleService.R
\name{dfp_updateAdRules}
\alias{dfp_updateAdRules}
\title{updateAdRules}
\usage{
dfp_updateAdRules(request_data)
}
\arguments{
\item{request_data}{a \code{list} or \code{data.frame} of data elements
to be formatted for a SOAP request (XML format, but passed as character string)#' @return a \code{list} containing all the elements of a updateAdRulesResponse}
}
\description{
Updates the specified AdRule objects.
}

