% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InventoryService.R
\name{dfp_updateAdUnits}
\alias{dfp_updateAdUnits}
\title{updateAdUnits}
\usage{
dfp_updateAdUnits(request_data)
}
\arguments{
\item{request_data}{a \code{list} or \code{data.frame} of data elements
to be formatted for a SOAP request (XML format, but passed as character string)}
}
\value{
a \code{list} containing all the elements of a updateAdUnitsResponse
}
\description{
Updates the specified AdUnit objects.
}

