% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AudienceSegmentService.R
\name{dfp_updateAudienceSegments}
\alias{dfp_updateAudienceSegments}
\title{updateAudienceSegments}
\usage{
dfp_updateAudienceSegments(request_data, as_df=FALSE)
}
\arguments{
\item{request_data}{a \code{list} or \code{data.frame} of data elements
to be formatted for a SOAP request (XML format, but passed as character string)}

\item{as_df}{a boolean indicating whether to attempt to parse the result into a \code{data.frame}}
}
\value{
a \code{list} or \code{data.frame} containing all the elements of a updateAudienceSegmentsResponse
}
\description{
Updates the given RuleBasedFirstPartyAudienceSegment objects.
}
\seealso{
\href{https://developers.google.com/doubleclick-publishers/docs/reference/v201508/AudienceSegmentService#updateAudienceSegments}{Google Documentation for updateAudienceSegments}
}

