% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CdnConfigurationService.R
\name{dfp_updateCdnConfigurations}
\alias{dfp_updateCdnConfigurations}
\title{updateCdnConfigurations}
\usage{
dfp_updateCdnConfigurations(request_data, as_df = TRUE, verbose = FALSE)
}
\arguments{
\item{request_data}{a \code{list} or \code{data.frame} of data elements
to be formatted for a SOAP
request (XML format, but passed as character string)}

\item{as_df}{a boolean indicating whether to attempt to parse the result into
a \code{data.frame}}

\item{verbose}{a boolean indicating whether to print the service URL and POSTed XML}
}
\value{
a \code{data.frame} or \code{list} containing all the elements of a updateCdnConfigurationsResponse
}
\description{
Updates the specified CdnConfiguration objects. Updates the specified CdnConfiguration objects. Updates the specified CdnConfiguration objects.
}
\examples{
\dontrun{
 res <- dfp_updateCdnConfigurations(request_data)
}
}
\seealso{
\href{https://developers.google.com/doubleclick-publishers/docs/reference/v201802/CdnConfigurationService#updateCdnConfigurations}{Google Documentation for updateCdnConfigurations}
}
