% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CompanyService.R
\name{dfp_updateCompanies}
\alias{dfp_updateCompanies}
\title{updateCompanies}
\usage{
dfp_updateCompanies(request_data)
}
\arguments{
\item{request_data}{a \code{list} or \code{data.frame} of data elements
to be formatted for a SOAP request (XML format, but passed as character string)}
}
\value{
a \code{list} containing all the elements of a updateCompaniesResponse
}
\description{
Updates the specified Company objects.
}

