% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ContactService.R
\name{dfp_updateContacts}
\alias{dfp_updateContacts}
\title{updateContacts}
\usage{
dfp_updateContacts(request_data)
}
\arguments{
\item{request_data}{a \code{list} or \code{data.frame} of data elements
to be formatted for a SOAP request (XML format, but passed as character string)}
}
\value{
a \code{list} containing all the elements of a updateContactsResponse
}
\description{
Updates the specified Contact objects.
}

