% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ContentMetadataKeyHierarchyService.R
\name{dfp_updateContentMetadataKeyHierarchies}
\alias{dfp_updateContentMetadataKeyHierarchies}
\title{updateContentMetadataKeyHierarchies}
\usage{
dfp_updateContentMetadataKeyHierarchies(request_data)
}
\arguments{
\item{request_data}{a \code{list} or \code{data.frame} of data elements
to be formatted for a SOAP request (XML format, but passed as character string)}
}
\value{
a \code{list} containing all the elements of a updateContentMetadataKeyHierarchiesResponse
}
\description{
Updates the specified ContentMetadataKeyHierarchy objects.
}
\seealso{
\href{https://developers.google.com/doubleclick-publishers/docs/reference/v201508/ContentMetadataKeyHierarchyService#updateContentMetadataKeyHierarchies}{Google Documentation for updateContentMetadataKeyHierarchies}
}

