% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreativeSetService.R
\name{dfp_updateCreativeSet}
\alias{dfp_updateCreativeSet}
\title{updateCreativeSet}
\usage{
dfp_updateCreativeSet(request_data)
}
\arguments{
\item{request_data}{a \code{list} or \code{data.frame} of data elements
to be formatted for a SOAP request (XML format, but passed as character string)#' @return a \code{list} containing all the elements of a updateCreativeSetResponse}
}
\description{
Updates the specified CreativeSet.
}

