% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreativeWrapperService.R
\name{dfp_updateCreativeWrappers}
\alias{dfp_updateCreativeWrappers}
\title{updateCreativeWrappers}
\usage{
dfp_updateCreativeWrappers(request_data)
}
\arguments{
\item{request_data}{a \code{list} or \code{data.frame} of data elements
to be formatted for a SOAP request (XML format, but passed as character string)}
}
\value{
a \code{list} containing all the elements of a updateCreativeWrappersResponse
}
\description{
Updates the specified CreativeWrapper objects.
}

