% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CustomFieldService.R
\name{dfp_updateCustomFieldOptions}
\alias{dfp_updateCustomFieldOptions}
\title{updateCustomFieldOptions}
\usage{
dfp_updateCustomFieldOptions(request_data)
}
\arguments{
\item{request_data}{a \code{list} or \code{data.frame} of data elements
to be formatted for a SOAP request (XML format, but passed as character string)#' @return a \code{list} containing all the elements of a updateCustomFieldOptionsResponse}
}
\description{
Updates the specified CustomFieldOption objects.
}

