% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CustomFieldService.R
\name{dfp_updateCustomFields}
\alias{dfp_updateCustomFields}
\title{updateCustomFields}
\usage{
dfp_updateCustomFields(request_data)
}
\arguments{
\item{request_data}{a \code{list} or \code{data.frame} of data elements
to be formatted for a SOAP request (XML format, but passed as character string)}
}
\value{
a \code{list} containing all the elements of a updateCustomFieldsResponse
}
\description{
Updates the specified CustomField objects.
}

