% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CustomTargetingService.R
\name{dfp_updateCustomTargetingKeys}
\alias{dfp_updateCustomTargetingKeys}
\title{updateCustomTargetingKeys}
\usage{
dfp_updateCustomTargetingKeys(request_data)
}
\arguments{
\item{request_data}{a \code{list} or \code{data.frame} of data elements
to be formatted for a SOAP request (XML format, but passed as character string)}
}
\value{
a \code{list} containing all the elements of a updateCustomTargetingKeysResponse
}
\description{
Updates the specified CustomTargetingKey objects.
}
\seealso{
\href{https://developers.google.com/doubleclick-publishers/docs/reference/v201508/CustomTargetingService#updateCustomTargetingKeys}{Google Documentation for updateCustomTargetingKeys}
}

