% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExchangeRateService.R
\name{dfp_updateExchangeRates}
\alias{dfp_updateExchangeRates}
\title{updateExchangeRates}
\usage{
dfp_updateExchangeRates(request_data)
}
\arguments{
\item{request_data}{a \code{list} or \code{data.frame} of data elements
to be formatted for a SOAP request (XML format, but passed as character string)#' @return a \code{list} containing all the elements of a updateExchangeRatesResponse}
}
\description{
Updates the specified ExchangeRate objects.
}

