% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LabelService.R
\name{dfp_updateLabels}
\alias{dfp_updateLabels}
\title{updateLabels}
\usage{
dfp_updateLabels(request_data, as_df=TRUE)
}
\arguments{
\item{request_data}{a \code{list} or \code{data.frame} of data elements
to be formatted for a SOAP request (XML format, but passed as character string)}

\item{as_df}{a boolean indicating whether to attempt to parse the result into a \code{data.frame}}
}
\value{
a \code{data.frame} or \code{list} containing all the elements of a updateLabelsResponse
}
\description{
Updates the specified Label objects.
}
\seealso{
\href{https://developers.google.com/doubleclick-publishers/docs/reference/v201508/LabelService#updateLabels}{Google Documentation for updateLabels}
}

