% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LineItemCreativeAssociationService.R
\name{dfp_updateLineItemCreativeAssociations}
\alias{dfp_updateLineItemCreativeAssociations}
\title{updateLineItemCreativeAssociations}
\usage{
dfp_updateLineItemCreativeAssociations(request_data)
}
\arguments{
\item{request_data}{a \code{list} or \code{data.frame} of data elements
to be formatted for a SOAP request (XML format, but passed as character string)#' @return a \code{list} containing all the elements of a updateLineItemCreativeAssociationsResponse}
}
\description{
Updates the specified LineItemCreativeAssociation objects update
}

