% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LineItemService.R
\name{dfp_updateLineItems}
\alias{dfp_updateLineItems}
\title{updateLineItems}
\usage{
dfp_updateLineItems(request_data)
}
\arguments{
\item{request_data}{a \code{list} or \code{data.frame} of data elements
to be formatted for a SOAP request (XML format, but passed as character string)}
}
\value{
a \code{list} containing all the elements of a updateLineItemsResponse
}
\description{
Updates the specified LineItem objects.
}

