% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LiveStreamEventService.R
\name{dfp_updateLiveStreamEvents}
\alias{dfp_updateLiveStreamEvents}
\title{updateLiveStreamEvents}
\usage{
dfp_updateLiveStreamEvents(request_data)
}
\arguments{
\item{request_data}{a \code{list} or \code{data.frame} of data elements
to be formatted for a SOAP request (XML format, but passed as character string)}
}
\value{
a \code{list} containing all the elements of a updateLiveStreamEventsResponse
}
\description{
Updates the specified LiveStreamEvent objects.
}
\seealso{
\href{https://developers.google.com/doubleclick-publishers/docs/reference/v201508/LiveStreamEventService#updateLiveStreamEvents}{Google Documentation for updateLiveStreamEvents}
}

