% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NetworkService.R
\name{dfp_updateNetwork}
\alias{dfp_updateNetwork}
\title{updateNetwork}
\usage{
dfp_updateNetwork(request_data)
}
\arguments{
\item{request_data}{a \code{list} or \code{data.frame} of data elements
to be formatted for a SOAP request (XML format, but passed as character string)#' @return a \code{list} containing all the elements of a updateNetworkResponse}
}
\description{
Updates the specified network. Currently, only the network display name can be updated.
}

