% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlacementService.R
\name{dfp_updatePlacements}
\alias{dfp_updatePlacements}
\title{updatePlacements}
\usage{
dfp_updatePlacements(request_data)
}
\arguments{
\item{request_data}{a \code{list} or \code{data.frame} of data elements
to be formatted for a SOAP request (XML format, but passed as character string)}
}
\value{
a \code{list} containing all the elements of a updatePlacementsResponse
}
\description{
Updates the specified Placement objects.
}

