% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PremiumRateService.R
\name{dfp_updatePremiumRates}
\alias{dfp_updatePremiumRates}
\title{updatePremiumRates}
\usage{
dfp_updatePremiumRates(request_data)
}
\arguments{
\item{request_data}{a \code{list} or \code{data.frame} of data elements
to be formatted for a SOAP request (XML format, but passed as character string)}
}
\value{
a \code{list} containing all the elements of a updatePremiumRatesResponse
}
\description{
Updates the specified PremiumRate objects.
}

