% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProductPackageService.R
\name{dfp_updateProductPackages}
\alias{dfp_updateProductPackages}
\title{updateProductPackages}
\usage{
dfp_updateProductPackages(request_data)
}
\arguments{
\item{request_data}{a \code{list} or \code{data.frame} of data elements
to be formatted for a SOAP request (XML format, but passed as character string)}
}
\value{
a \code{list} containing all the elements of a updateProductPackagesResponse
}
\description{
Updates the specified ProductPackage objects.
}
\seealso{
\href{https://developers.google.com/doubleclick-publishers/docs/reference/v201508/ProductPackageService#updateProductPackages}{Google Documentation for updateProductPackages}
}

