% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProductService.R
\name{dfp_updateProducts}
\alias{dfp_updateProducts}
\title{updateProducts}
\usage{
dfp_updateProducts(request_data)
}
\arguments{
\item{request_data}{a \code{list} or \code{data.frame} of data elements
to be formatted for a SOAP request (XML format, but passed as character string)}
}
\value{
a \code{list} containing all the elements of a updateProductsResponse
}
\description{
Updates the specified Product objects. Note non-updatable fields will not be backfilled.
}

