% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProductTemplateService.R
\name{dfp_updateProductTemplates}
\alias{dfp_updateProductTemplates}
\title{updateProductTemplates}
\usage{
dfp_updateProductTemplates(request_data)
}
\arguments{
\item{request_data}{a \code{list} or \code{data.frame} of data elements
to be formatted for a SOAP request (XML format, but passed as character string)}
}
\value{
a \code{list} containing all the elements of a updateProductTemplatesResponse
}
\description{
Updates the specified ProductTemplate objects.
}
\seealso{
\href{https://developers.google.com/doubleclick-publishers/docs/reference/v201508/ProductTemplateService#updateProductTemplates}{Google Documentation for updateProductTemplates}
}

