% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProposalService.R
\name{dfp_updateProposals}
\alias{dfp_updateProposals}
\title{updateProposals}
\usage{
dfp_updateProposals(request_data, as_df=TRUE)
}
\arguments{
\item{request_data}{a \code{list} or \code{data.frame} of data elements
to be formatted for a SOAP request (XML format, but passed as character string)}

\item{as_df}{a boolean indicating whether to attempt to parse the result into a \code{data.frame}}
}
\value{
a \code{data.frame} or \code{list} containing all the elements of a updateProposalsResponse
}
\description{
Updates the specified Proposal objects.
}
\seealso{
\href{https://developers.google.com/doubleclick-publishers/docs/reference/v201508/ProposalService#updateProposals}{Google Documentation for updateProposals}
}

