% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RateCardService.R
\name{dfp_updateRateCards}
\alias{dfp_updateRateCards}
\title{updateRateCards}
\usage{
dfp_updateRateCards(request_data)
}
\arguments{
\item{request_data}{a \code{list} or \code{data.frame} of data elements
to be formatted for a SOAP request (XML format, but passed as character string)}
}
\value{
a \code{list} containing all the elements of a updateRateCardsResponse
}
\description{
Updates a list of RateCard objects.
}
\seealso{
\href{https://developers.google.com/doubleclick-publishers/docs/reference/v201508/RateCardService#updateRateCards}{Google Documentation for updateRateCards}
}

