% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReconciliationLineItemReportService.R
\name{dfp_updateReconciliationLineItemReports}
\alias{dfp_updateReconciliationLineItemReports}
\title{updateReconciliationLineItemReports}
\usage{
dfp_updateReconciliationLineItemReports(request_data)
}
\arguments{
\item{request_data}{a \code{list} or \code{data.frame} of data elements
to be formatted for a SOAP request (XML format, but passed as character string)#' @return a \code{list} containing all the elements of a updateReconciliationLineItemReportsResponse}
}
\description{
Updates a list of ReconciliationLineItemReport objects which belong to same ReconciliationReport. ReconciliationLineItemReport reconciliation line item reports to update
}

