% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReconciliationReportRowService.R
\name{dfp_updateReconciliationReportRows}
\alias{dfp_updateReconciliationReportRows}
\title{updateReconciliationReportRows}
\usage{
dfp_updateReconciliationReportRows(request_data)
}
\arguments{
\item{request_data}{a \code{list} or \code{data.frame} of data elements
to be formatted for a SOAP request (XML format, but passed as character string)#' @return a \code{list} containing all the elements of a updateReconciliationReportRowsResponse}
}
\description{
Updates a list of ReconciliationReportRow which belong to same ReconciliationReport.
}

