% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReconciliationReportService.R
\name{dfp_updateReconciliationReports}
\alias{dfp_updateReconciliationReports}
\title{updateReconciliationReports}
\usage{
dfp_updateReconciliationReports(request_data)
}
\arguments{
\item{request_data}{a \code{list} or \code{data.frame} of data elements
to be formatted for a SOAP request (XML format, but passed as character string)#' @return a \code{list} containing all the elements of a updateReconciliationReportsResponse}
}
\description{
Updates the specified ReconciliationReport objects.
}

