% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TeamService.R
\name{dfp_updateTeams}
\alias{dfp_updateTeams}
\title{updateTeams}
\usage{
dfp_updateTeams(request_data)
}
\arguments{
\item{request_data}{a \code{list} or \code{data.frame} of data elements
to be formatted for a SOAP request (XML format, but passed as character string)#' @return a \code{list} containing all the elements of a updateTeamsResponse}
}
\description{
Updates the specified Team objects.
}

