% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UserService.R
\name{dfp_updateUsers}
\alias{dfp_updateUsers}
\title{updateUsers}
\usage{
dfp_updateUsers(request_data)
}
\arguments{
\item{request_data}{a \code{list} or \code{data.frame} of data elements
to be formatted for a SOAP request (XML format, but passed as character string)#' @return a \code{list} containing all the elements of a updateUsersResponse}
}
\description{
Updates the specified User objects.
}

