% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UserTeamAssociationService.R
\name{dfp_updateUserTeamAssociations}
\alias{dfp_updateUserTeamAssociations}
\title{updateUserTeamAssociations}
\usage{
dfp_updateUserTeamAssociations(request_data)
}
\arguments{
\item{request_data}{a \code{list} or \code{data.frame} of data elements
to be formatted for a SOAP request (XML format, but passed as character string)#' @return a \code{list} containing all the elements of a updateUserTeamAssociationsResponse}
}
\description{
Updates the specified UserTeamAssociation objects.
}

