% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WorkflowRequestService.R
\name{dfp_WorkflowRequestService_object_factory}
\alias{dfp_WorkflowRequestService_object_factory}
\title{WorkflowRequestService}
\usage{
dfp_WorkflowRequestService_object_factory(obj_type, obj_data)
}
\arguments{
\item{obj_type}{a string from one of the object types described above}

\item{obj_data}{a \code{list} of \code{lists} or a \code{data.frame} with the required components to create
the obj_type specified. All input names must match the expected attributes for the object.
If submitting a list of attributes to create 1 object make sure to nest it inside top-level list, so that
it is a list of length 1 with 1 list of attributes to create the object.}
}
\value{
a string of formatted XML that can be inserted into a SOAP request body to the APT
}
\description{
Provides methods to retrieve and perform actions on WorkflowRequest objects
}
\details{
To use this service, you need to have the new sales management solution
enabled on your network. If you do not see a "Sales" tab in <a
href="https://www.google.com/dfp">DoubleClick for Publishers (DFP)</a>, you
will not be able to use this service.

WorkflowRequestService Object Factory

A function to create a variety of objects that are part of the WorkflowRequestService
Note: not all objects here are inputs to the functions of this service so 
please consult each function for the required inputs

Below is a list of objects and their required components to be created with this function:

\strong{ApiError}

The API error base class that provides details about an error that occurred
while processing a service request. The OGNL field path is provided for
parsers to identify the request data element that may have caused the error.
\href{https://developers.google.com/doubleclick-publishers/docs/reference/v201508/WorkflowRequestService.ApiError}{Google Documentation for ApiError}
\describe{
 \item{fieldPath}{a character - The OGNL field path to identify cause of error.}
 \item{trigger}{a character - The data that caused the error.}
 \item{errorString}{a character - A simple string representation of the error and reason.}
}

\strong{ApiException}

Exception class for holding a list of service errors.
\href{https://developers.google.com/doubleclick-publishers/docs/reference/v201508/WorkflowRequestService.ApiException}{Google Documentation for ApiException}
\describe{
 \item{message}{a character (inherited from ApplicationException) - Error message.}
 \item{errors}{a ApiError - List of errors.}
}

\strong{ApiVersionError}

Errors related to the usage of API versions.
\href{https://developers.google.com/doubleclick-publishers/docs/reference/v201508/WorkflowRequestService.ApiVersionError}{Google Documentation for ApiVersionError}
\describe{
 \item{extends ApiError}{see documentation for ApiError}
 \item{reason}{a ApiVersionError.Reason - This can take one of the following values: 
   \itemize{
     \item{UPDATE_TO_NEWER_VERSION - Indicates that the operation is not allowed in the version the request was
made in.}
     \item{UNKNOWN - The value returned if the actual value is not exposed by the requested API
version.}
   }
  }
}

\strong{ApplicationException}

Base class for exceptions.
\href{https://developers.google.com/doubleclick-publishers/docs/reference/v201508/WorkflowRequestService.ApplicationException}{Google Documentation for ApplicationException}
\describe{
 \item{message}{a character - Error message.}
}

\strong{ApproveWorkflowApprovalRequests}

The action used to approve WorkflowApprovalRequest workflow approval
requests.
\href{https://developers.google.com/doubleclick-publishers/docs/reference/v201508/WorkflowRequestService.ApproveWorkflowApprovalRequests}{Google Documentation for ApproveWorkflowApprovalRequests}
\describe{
 \item{extends WorkflowRequestAction}{see documentation for WorkflowRequestAction}
 \item{comment}{a character - The comment of the ApproveWorkflowApprovalRequests action. This field is
optional.}
}

\strong{AuthenticationError}

An error for an exception that occurred when authenticating.
\href{https://developers.google.com/doubleclick-publishers/docs/reference/v201508/WorkflowRequestService.AuthenticationError}{Google Documentation for AuthenticationError}
\describe{
 \item{extends ApiError}{see documentation for ApiError}
 \item{reason}{a AuthenticationError.Reason - This can take one of the following values: 
   \itemize{
     \item{AMBIGUOUS_SOAP_REQUEST_HEADER - The SOAP message contains a request header with an ambiguous definition of
the authentication header fields. This means either the authToken and
oAuthToken fields were both null or both were specified. Exactly one value
should be specified with each request.}
     \item{INVALID_EMAIL - The login provided is invalid.}
     \item{AUTHENTICATION_FAILED - Tried to authenticate with provided information, but failed.}
     \item{INVALID_OAUTH_SIGNATURE - The OAuth provided is invalid.}
     \item{INVALID_SERVICE - The specified service to use was not recognized.}
     \item{MISSING_SOAP_REQUEST_HEADER - The SOAP message is missing a request header with an authToken and optional
networkCode.}
     \item{MISSING_AUTHENTICATION_HTTP_HEADER - The HTTP request is missing a request header with an authToken}
     \item{MISSING_AUTHENTICATION - The request is missing an authToken}
     \item{NOT_WHITELISTED_FOR_API_ACCESS - The customer is not whitelisted for API access.}
     \item{NO_NETWORKS_TO_ACCESS - The user is not associated with any network.}
     \item{NETWORK_NOT_FOUND - No network for the given networkCode was found.}
     \item{NETWORK_CODE_REQUIRED - The user has access to more than one network, but did not provide a
networkCode.}
     \item{CONNECTION_ERROR - An error happened on the server side during connection to authentication
service.}
     \item{GOOGLE_ACCOUNT_ALREADY_ASSOCIATED_WITH_NETWORK - The user tried to create a test network using an account that already is
associated with a network.}
     \item{UNKNOWN - The value returned if the actual value is not exposed by the requested API
version.}
   }
  }
}

\strong{BooleanValue}

Contains a boolean value.
\href{https://developers.google.com/doubleclick-publishers/docs/reference/v201508/WorkflowRequestService.BooleanValue}{Google Documentation for BooleanValue}
\describe{
 \item{extends Value}{see documentation for Value}
 \item{value}{a boolean - The boolean value.}
}

\strong{CollectionSizeError}

Error for the size of the collection being too large
\href{https://developers.google.com/doubleclick-publishers/docs/reference/v201508/WorkflowRequestService.CollectionSizeError}{Google Documentation for CollectionSizeError}
\describe{
 \item{extends ApiError}{see documentation for ApiError}
 \item{reason}{a CollectionSizeError.Reason - This can take one of the following values: 
   \itemize{
     \item{TOO_LARGE}
     \item{UNKNOWN - The value returned if the actual value is not exposed by the requested API
version.}
   }
  }
}

\strong{CommonError}

A place for common errors that can be used across services.
\href{https://developers.google.com/doubleclick-publishers/docs/reference/v201508/WorkflowRequestService.CommonError}{Google Documentation for CommonError}
\describe{
 \item{extends ApiError}{see documentation for ApiError}
 \item{reason}{a CommonError.Reason - This can take one of the following values: 
   \itemize{
     \item{NOT_FOUND - Indicates that an attempt was made to retrieve an entity that does not
exist.}
     \item{ALREADY_EXISTS - Indicates that an attempt was made to create an entity that already exists.}
     \item{DUPLICATE_OBJECT - Indicates that two elements in the collection were identical.}
     \item{CANNOT_UPDATE - Indicates that an attempt was made to change an immutable field.}
     \item{CONCURRENT_MODIFICATION - Indicates that another request attempted to update the same data in the same
network at about the same time. Please wait and try the request again.}
     \item{UNKNOWN - The value returned if the actual value is not exposed by the requested API
version.}
   }
  }
}

\strong{Date}

Represents a date.
\href{https://developers.google.com/doubleclick-publishers/docs/reference/v201508/WorkflowRequestService.Date}{Google Documentation for Date}
\describe{
 \item{year}{a integer - Year (e.g., 2009)}
 \item{month}{a integer - Month (1..12)}
 \item{day}{a integer - Day (1..31)}
}

\strong{DateTime}

Represents a date combined with the time of day.
\href{https://developers.google.com/doubleclick-publishers/docs/reference/v201508/WorkflowRequestService.DateTime}{Google Documentation for DateTime}
\describe{
 \item{date}{a Date}
 \item{hour}{a integer}
 \item{minute}{a integer}
 \item{second}{a integer}
 \item{timeZoneID}{a character}
}

\strong{DateTimeValue}

Contains a date-time value.
\href{https://developers.google.com/doubleclick-publishers/docs/reference/v201508/WorkflowRequestService.DateTimeValue}{Google Documentation for DateTimeValue}
\describe{
 \item{extends Value}{see documentation for Value}
 \item{value}{a DateTime - The DateTime value.}
}

\strong{DateValue}

Contains a date value.
\href{https://developers.google.com/doubleclick-publishers/docs/reference/v201508/WorkflowRequestService.DateValue}{Google Documentation for DateValue}
\describe{
 \item{extends Value}{see documentation for Value}
 \item{value}{a Date - The Date value.}
}

\strong{FeatureError}

Errors related to feature management. If you attempt using a feature that is
not available to the current network you'll receive a FeatureError with the
missing feature as the trigger.
\href{https://developers.google.com/doubleclick-publishers/docs/reference/v201508/WorkflowRequestService.FeatureError}{Google Documentation for FeatureError}
\describe{
 \item{extends ApiError}{see documentation for ApiError}
 \item{reason}{a FeatureError.Reason - This can take one of the following values: 
   \itemize{
     \item{MISSING_FEATURE - A feature is being used that is not enabled on the current network.}
     \item{UNKNOWN - The value returned if the actual value is not exposed by the requested API
version.}
   }
  }
}

\strong{ForecastError}

Errors that can result from a forecast request.
\href{https://developers.google.com/doubleclick-publishers/docs/reference/v201508/WorkflowRequestService.ForecastError}{Google Documentation for ForecastError}
\describe{
 \item{extends ApiError}{see documentation for ApiError}
 \item{reason}{a ForecastError.Reason - The reason for the forecast error. This can take one of the following values: 
   \itemize{
     \item{SERVER_NOT_AVAILABLE - The forecast could not be retrieved due to a server side connection problem.
Please try again soon.}
     \item{INTERNAL_ERROR - There was an unexpected internal error.}
     \item{NO_FORECAST_YET - The forecast could not be retrieved because there is not enough forecasting
data available yet. It may take up to one week before enough data is
available.}
     \item{NOT_ENOUGH_INVENTORY - There's not enough inventory for the requested reservation.}
     \item{SUCCESS - No error from forecast.}
     \item{ZERO_LENGTH_RESERVATION - The requested reservation is of zero length. No forecast is returned.}
     \item{EXCEEDED_QUOTA - The number of requests made per second is too high and has exceeded the
allowable limit. The recommended approach to handle this error is to wait
about 5 seconds and then retry the request. Note that this does not
guarantee the request will succeed. If it fails again, try increasing the
wait time.  Another way to mitigate this error is to limit requests to 2 per
second. Once again this does not guarantee that every request will succeed,
but may help reduce the number of times you receive this error.}
     \item{UNKNOWN - The value returned if the actual value is not exposed by the requested API
version.}
   }
  }
}

\strong{InternalApiError}

Indicates that a server-side error has occured. InternalApiErrors are
generally not the result of an invalid request or message sent by the
client.
\href{https://developers.google.com/doubleclick-publishers/docs/reference/v201508/WorkflowRequestService.InternalApiError}{Google Documentation for InternalApiError}
\describe{
 \item{extends ApiError}{see documentation for ApiError}
 \item{reason}{a InternalApiError.Reason - The error reason represented by an enum. This can take one of the following values: 
   \itemize{
     \item{UNEXPECTED_INTERNAL_API_ERROR - API encountered an unexpected internal error.}
     \item{TRANSIENT_ERROR - A temporary error occurred during the request. Please retry.}
     \item{UNKNOWN - The cause of the error is not known or only defined in newer versions.}
     \item{DOWNTIME - The API is currently unavailable for a planned downtime.}
   }
  }
}

\strong{LineItemOperationError}

Lists all errors for executing operations on line items
\href{https://developers.google.com/doubleclick-publishers/docs/reference/v201508/WorkflowRequestService.LineItemOperationError}{Google Documentation for LineItemOperationError}
\describe{
 \item{extends ApiError}{see documentation for ApiError}
 \item{reason}{a LineItemOperationError.Reason - The error reason represented by an enum. This can take one of the following values: 
   \itemize{
     \item{NOT_ALLOWED - The operation is not allowed due to lack of permissions.}
     \item{NOT_APPLICABLE - The operation is not applicable for the current state of the LineItem.}
     \item{HAS_COMPLETED - The LineItem is completed. A LineItemAction cannot be applied to a line item
that is completed.}
     \item{HAS_NO_ACTIVE_CREATIVES - The LineItem has no active creatives. A line item cannot be activated with
no active creatives.}
     \item{CANNOT_ACTIVATE_LEGACY_DFP_LINE_ITEM - A LineItem of type LineItemType LEGACY_DFP cannot be Activated.}
     \item{CANNOT_DELETE_DELIVERED_LINE_ITEM - Deleting an LineItem that has delivered is not allowed}
     \item{CANNOT_RESERVE_COMPANY_CREDIT_STATUS_NOT_ACTIVE - Reservation cannot be made for line item because the LineItem advertiserId
it is associated with has Company creditStatus that is not ACTIVE or
ON_HOLD.}
     \item{CANNOT_ACTIVATE_INVALID_COMPANY_CREDIT_STATUS - Cannot activate line item because the LineItem advertiserId it is associated
with has Company creditStatus that is not ACTIVE, INACTIVE, or ON_HOLD.}
     \item{UNKNOWN - The value returned if the actual value is not exposed by the requested API
version.}
   }
  }
}

\strong{NotNullError}

Caused by supplying a null value for an attribute that cannot be null.
\href{https://developers.google.com/doubleclick-publishers/docs/reference/v201508/WorkflowRequestService.NotNullError}{Google Documentation for NotNullError}
\describe{
 \item{extends ApiError}{see documentation for ApiError}
 \item{reason}{a NotNullError.Reason - The error reason represented by an enum. This can take one of the following values: 
   \itemize{
     \item{ARG1_NULL - Assuming that a method will not have more than 3 arguments, if it does,
return NULL}
     \item{ARG2_NULL}
     \item{ARG3_NULL}
     \item{NULL}
     \item{UNKNOWN - The value returned if the actual value is not exposed by the requested API
version.}
   }
  }
}

\strong{NumberValue}

Contains a numeric value.
\href{https://developers.google.com/doubleclick-publishers/docs/reference/v201508/WorkflowRequestService.NumberValue}{Google Documentation for NumberValue}
\describe{
 \item{extends Value}{see documentation for Value}
 \item{value}{a character - The numeric value represented as a string.}
}

\strong{ObjectValue}

Contains an object value.  This object is experimental! ObjectValue is an
experimental, innovative, and rapidly changing new feature for DFP.
Unfortunately, being on the bleeding edge means that we may make
backwards-incompatible changes to ObjectValue. We will inform the community
when this feature is no longer experimental.
\href{https://developers.google.com/doubleclick-publishers/docs/reference/v201508/WorkflowRequestService.ObjectValue}{Google Documentation for ObjectValue}
\describe{
 \item{extends Value}{see documentation for Value}
}

\strong{ParseError}

Lists errors related to parsing.
\href{https://developers.google.com/doubleclick-publishers/docs/reference/v201508/WorkflowRequestService.ParseError}{Google Documentation for ParseError}
\describe{
 \item{extends ApiError}{see documentation for ApiError}
 \item{reason}{a ParseError.Reason - The error reason represented by an enum. This can take one of the following values: 
   \itemize{
     \item{UNPARSABLE - Indicates an error in parsing an attribute.}
     \item{UNKNOWN - The value returned if the actual value is not exposed by the requested API
version.}
   }
  }
}

\strong{PermissionError}

Errors related to incorrect permission.
\href{https://developers.google.com/doubleclick-publishers/docs/reference/v201508/WorkflowRequestService.PermissionError}{Google Documentation for PermissionError}
\describe{
 \item{extends ApiError}{see documentation for ApiError}
 \item{reason}{a PermissionError.Reason - This can take one of the following values: 
   \itemize{
     \item{PERMISSION_DENIED - User does not have the required permission for the request.}
     \item{UNKNOWN - The value returned if the actual value is not exposed by the requested API
version.}
   }
  }
}

\strong{ProposalActionError}

Lists all errors associated with performing actions on Proposal objects.
\href{https://developers.google.com/doubleclick-publishers/docs/reference/v201508/WorkflowRequestService.ProposalActionError}{Google Documentation for ProposalActionError}
\describe{
 \item{extends ApiError}{see documentation for ApiError}
 \item{reason}{a ProposalActionError.Reason - The error reason represented by an enum. This can take one of the following values: 
   \itemize{
     \item{NOT_APPLICABLE - The operation is not applicable to the current state.}
     \item{IS_ARCHIVED - The operation cannot be applied because the proposal is archived.}
     \item{UNKNOWN - The value returned if the actual value is not exposed by the requested API
version.}
   }
  }
}

\strong{PublisherQueryLanguageContextError}

An error that occurs while executing a PQL query contained in a Statement
object.
\href{https://developers.google.com/doubleclick-publishers/docs/reference/v201508/WorkflowRequestService.PublisherQueryLanguageContextError}{Google Documentation for PublisherQueryLanguageContextError}
\describe{
 \item{extends ApiError}{see documentation for ApiError}
 \item{reason}{a PublisherQueryLanguageContextError.Reason - The error reason represented by an enum. This can take one of the following values: 
   \itemize{
     \item{UNEXECUTABLE - Indicates that there was an error executing the PQL.}
     \item{UNKNOWN - The value returned if the actual value is not exposed by the requested API
version.}
   }
  }
}

\strong{PublisherQueryLanguageSyntaxError}

An error that occurs while parsing a PQL query contained in a Statement
object.
\href{https://developers.google.com/doubleclick-publishers/docs/reference/v201508/WorkflowRequestService.PublisherQueryLanguageSyntaxError}{Google Documentation for PublisherQueryLanguageSyntaxError}
\describe{
 \item{extends ApiError}{see documentation for ApiError}
 \item{reason}{a PublisherQueryLanguageSyntaxError.Reason - The error reason represented by an enum. This can take one of the following values: 
   \itemize{
     \item{UNPARSABLE - Indicates that there was a PQL syntax error.}
     \item{UNKNOWN - The value returned if the actual value is not exposed by the requested API
version.}
   }
  }
}

\strong{QuotaError}

Describes a client-side error on which a user is attempting to perform an
action to which they have no quota remaining.
\href{https://developers.google.com/doubleclick-publishers/docs/reference/v201508/WorkflowRequestService.QuotaError}{Google Documentation for QuotaError}
\describe{
 \item{extends ApiError}{see documentation for ApiError}
 \item{reason}{a QuotaError.Reason - This can take one of the following values: 
   \itemize{
     \item{EXCEEDED_QUOTA - The number of requests made per second is too high and has exceeded the
allowable limit. The recommended approach to handle this error is to wait
about 5 seconds and then retry the request. Note that this does not
guarantee the request will succeed. If it fails again, try increasing the
wait time.  Another way to mitigate this error is to limit requests to 2 per
second for Small Business networks, or 8 per second for Premium networks.
Once again this does not guarantee that every request will succeed, but may
help reduce the number of times you receive this error.}
     \item{UNKNOWN - The value returned if the actual value is not exposed by the requested API
version.}
   }
  }
}

\strong{RejectWorkflowApprovalRequests}

The action used to reject WorkflowApprovalRequest workflow approval
requests.
\href{https://developers.google.com/doubleclick-publishers/docs/reference/v201508/WorkflowRequestService.RejectWorkflowApprovalRequests}{Google Documentation for RejectWorkflowApprovalRequests}
\describe{
 \item{extends WorkflowRequestAction}{see documentation for WorkflowRequestAction}
 \item{comment}{a character - The comment of the RejectWorkflowApprovalRequests action. This field is
optional.}
}

\strong{ServerError}

Errors related to the server.
\href{https://developers.google.com/doubleclick-publishers/docs/reference/v201508/WorkflowRequestService.ServerError}{Google Documentation for ServerError}
\describe{
 \item{extends ApiError}{see documentation for ApiError}
 \item{reason}{a ServerError.Reason - This can take one of the following values: 
   \itemize{
     \item{SERVER_ERROR - Indicates that an unexpected error occured.}
     \item{SERVER_BUSY - Indicates that the server is currently experiencing a high load. Please wait
and try your request again.}
     \item{UNKNOWN - The value returned if the actual value is not exposed by the requested API
version.}
   }
  }
}

\strong{SetValue}

Contains a set of Value Values. May not contain duplicates.
\href{https://developers.google.com/doubleclick-publishers/docs/reference/v201508/WorkflowRequestService.SetValue}{Google Documentation for SetValue}
\describe{
 \item{extends Value}{see documentation for Value}
 \item{values}{a Value - The values. They must all be the same type of Value and not contain
duplicates.}
}

\strong{SkipWorkflowExternalConditionRequests}

The action to skip WorkflowExternalConditionRequest workflow external
condition requests. Doing so means that the rule did not trigger.
\href{https://developers.google.com/doubleclick-publishers/docs/reference/v201508/WorkflowRequestService.SkipWorkflowExternalConditionRequests}{Google Documentation for SkipWorkflowExternalConditionRequests}
\describe{
 \item{extends WorkflowRequestAction}{see documentation for WorkflowRequestAction}
}

\strong{SoapRequestHeader}

Represents the SOAP request header used by API requests.
\href{https://developers.google.com/doubleclick-publishers/docs/reference/v201508/WorkflowRequestService.SoapRequestHeader}{Google Documentation for SoapRequestHeader}
\describe{
 \item{networkCode}{a character - The network code to use in the context of a request.}
 \item{applicationName}{a character - The name of client library application.}
}

\strong{SoapResponseHeader}

Represents the SOAP request header used by API responses.
\href{https://developers.google.com/doubleclick-publishers/docs/reference/v201508/WorkflowRequestService.SoapResponseHeader}{Google Documentation for SoapResponseHeader}
\describe{
 \item{requestId}{a character}
 \item{responseTime}{a integer}
}

\strong{Statement}

Holds the query in PQL syntax. See link for how to form PQL.
\href{https://developers.google.com/doubleclick-publishers/docs/reference/v201508/WorkflowRequestService.Statement}{Google Documentation for Statement}
\describe{
 \item{query}{a character - Holds the query in PQL syntax. See link for how to form PQL.}
 \item{values}{a String_ValueMapEntry - Holds keys and values for bind variables and their values. The key is the
name of the bind variable. The value is the literal value of the variable.
In the example "WHERE status = :bindStatus ORDER BY id LIMIT 30", the bind
variable, represented by :bindStatus is named bindStatus, which would also
be the parameter map key. The bind variable's value would be represented by
a parameter map value of type TextValue. The final result, for example,
would be an entry of "bindStatus" => StringParam("ACTIVE").}
}

\strong{StatementError}

An error that occurs while parsing Statement objects.
\href{https://developers.google.com/doubleclick-publishers/docs/reference/v201508/WorkflowRequestService.StatementError}{Google Documentation for StatementError}
\describe{
 \item{extends ApiError}{see documentation for ApiError}
 \item{reason}{a StatementError.Reason - The error reason represented by an enum. This can take one of the following values: 
   \itemize{
     \item{VARIABLE_NOT_BOUND_TO_VALUE - A bind variable has not been bound to a value.}
     \item{UNKNOWN - The value returned if the actual value is not exposed by the requested API
version.}
   }
  }
}

\strong{String_ValueMapEntry}

This represents an entry in a map with a key of type String and value of
type Value.
\href{https://developers.google.com/doubleclick-publishers/docs/reference/v201508/WorkflowRequestService.String_ValueMapEntry}{Google Documentation for String_ValueMapEntry}
\describe{
 \item{key}{a character}
 \item{value}{a Value}
}

\strong{TextValue}

Contains a string value.
\href{https://developers.google.com/doubleclick-publishers/docs/reference/v201508/WorkflowRequestService.TextValue}{Google Documentation for TextValue}
\describe{
 \item{extends Value}{see documentation for Value}
 \item{value}{a character - The string value.}
}

\strong{TriggerWorkflowExternalConditionRequests}

The action to trigger WorkflowExternalConditionRequest workflow external
condition requests.
\href{https://developers.google.com/doubleclick-publishers/docs/reference/v201508/WorkflowRequestService.TriggerWorkflowExternalConditionRequests}{Google Documentation for TriggerWorkflowExternalConditionRequests}
\describe{
 \item{extends WorkflowRequestAction}{see documentation for WorkflowRequestAction}
}

\strong{UpdateResult}

Represents the result of performing an action on objects.
\href{https://developers.google.com/doubleclick-publishers/docs/reference/v201508/WorkflowRequestService.UpdateResult}{Google Documentation for UpdateResult}
\describe{
 \item{numChanges}{a integer - The number of objects that were changed as a result of performing the
action.}
}

\strong{Value}

Value represents a value.
\href{https://developers.google.com/doubleclick-publishers/docs/reference/v201508/WorkflowRequestService.Value}{Google Documentation for Value}
\describe{
}

\strong{WorkflowActionError}

Lists all errors associated with performing actions within WorkflowAction.
\href{https://developers.google.com/doubleclick-publishers/docs/reference/v201508/WorkflowRequestService.WorkflowActionError}{Google Documentation for WorkflowActionError}
\describe{
 \item{extends ApiError}{see documentation for ApiError}
 \item{reason}{a WorkflowActionError.Reason - The error reason represented by an enum. This can take one of the following values: 
   \itemize{
     \item{NOT_APPLICABLE - The action does not exist or is not applicable to the current state.}
     \item{WORKFLOW_DEFINITION_NOT_FOUND - Means there's no workflow definition found for the entity.}
     \item{EMPTY_ACTION_LIST - Means no action is given, when user approve/reject a proposal, the action id
list cannot be empty.}
     \item{NOT_ACTION_APPROVER - Means the user is not an approver of this action.}
     \item{WORKFLOW_ALREADY_COMPLETED - Means the workflow is already completed.}
     \item{WORKFLOW_ALREADY_FAILED - Means the workflow is already failed.}
     \item{WORKFLOW_ALREADY_CANCELED - Means the workflow is already canceled.}
     \item{ACTION_COMPLETED - Means the action is already completed.}
     \item{ACTION_FAILED - Means the action is already failed.}
     \item{ACTION_CANCELED - Means the action is already canceled.}
     \item{ACTION_NOT_ACTIVE - Means the action currently is not active.}
     \item{UNKNOWN - The value returned if the actual value is not exposed by the requested API
version.}
   }
  }
}

\strong{WorkflowApprovalRequest}

A WorkflowApprovalRequest represents a workflow action unit that requires
user approval.
\href{https://developers.google.com/doubleclick-publishers/docs/reference/v201508/WorkflowRequestService.WorkflowApprovalRequest}{Google Documentation for WorkflowApprovalRequest}
\describe{
 \item{id}{a integer (inherited from WorkflowRequest) - The unique ID of the workflow request. This attribute is readonly and is
assigned by Google.}
 \item{workflowRuleName}{a character (inherited from WorkflowRequest) - The name of the workflow rule that generated the request. This attribute is
readonly and is assigned by Google.}
 \item{entityId}{a integer (inherited from WorkflowRequest) - The ID of entity associated to this workflow request.}
 \item{entityType}{a WorkflowEntityType (inherited from WorkflowRequest) - The WorkflowEntityType associated to this workflow request.}
 \item{type}{a WorkflowRequestType (inherited from WorkflowRequest) - The WorkflowRequestType of this workflow request. This attribute is readonly
and is assigned by Google.}
 \item{status}{a WorkflowApprovalRequestStatus - The status of the WorkflowApprovalRequest. This attribute is readonly. This can take one of the following values: 
   \itemize{
     \item{PENDING_APPROVAL - The request is pending approval.}
     \item{APPROVED - The workflow approval request has been approved.}
     \item{REJECTED - The workflow approval request has been rejected.}
     \item{RETRACTED - The workflow request was retracted because the proposal was retracted.}
     \item{UNKNOWN - The value returned if the actual value is not exposed by the requested API
version.}
   }
  }
}

\strong{WorkflowExternalConditionRequest}

A WorkflowExternalConditionRequest represents a workflow condition that
requires external system to determine the result.
\href{https://developers.google.com/doubleclick-publishers/docs/reference/v201508/WorkflowRequestService.WorkflowExternalConditionRequest}{Google Documentation for WorkflowExternalConditionRequest}
\describe{
 \item{id}{a integer (inherited from WorkflowRequest) - The unique ID of the workflow request. This attribute is readonly and is
assigned by Google.}
 \item{workflowRuleName}{a character (inherited from WorkflowRequest) - The name of the workflow rule that generated the request. This attribute is
readonly and is assigned by Google.}
 \item{entityId}{a integer (inherited from WorkflowRequest) - The ID of entity associated to this workflow request.}
 \item{entityType}{a WorkflowEntityType (inherited from WorkflowRequest) - The WorkflowEntityType associated to this workflow request.}
 \item{type}{a WorkflowRequestType (inherited from WorkflowRequest) - The WorkflowRequestType of this workflow request. This attribute is readonly
and is assigned by Google.}
 \item{status}{a WorkflowExternalConditionStatus - The status of the WorkflowExternalConditionRequest. This attribute is
readonly. This can take one of the following values: 
   \itemize{
     \item{PENDING - The workflow external condition is pending evaluation.}
     \item{TRIGGERED - The workflow external condition has been triggered.}
     \item{NOT_TRIGGERED - The workflow external condition has not been triggered.}
     \item{CANCELLED - The associated proposal was retracted, this rule is no longer valid.}
     \item{UNKNOWN - The value returned if the actual value is not exposed by the requested API
version.}
   }
  }
}

\strong{WorkflowRequest}

A WorkflowRequest represents a workflow action unit that requires external
or manual interference.
\href{https://developers.google.com/doubleclick-publishers/docs/reference/v201508/WorkflowRequestService.WorkflowRequest}{Google Documentation for WorkflowRequest}
\describe{
 \item{id}{a integer - The unique ID of the workflow request. This attribute is readonly and is
assigned by Google.}
 \item{workflowRuleName}{a character - The name of the workflow rule that generated the request. This attribute is
readonly and is assigned by Google.}
 \item{entityId}{a integer - The ID of entity associated to this workflow request.}
 \item{entityType}{a WorkflowEntityType - The WorkflowEntityType associated to this workflow request. This can take one of the following values: 
   \itemize{
     \item{PROPOSAL - Represents a Proposal.}
     \item{UNKNOWN - The value returned if the actual value is not exposed by the requested API
version.}
   }
  }
 \item{type}{a WorkflowRequestType - The WorkflowRequestType of this workflow request. This attribute is readonly
and is assigned by Google. This can take one of the following values: 
   \itemize{
     \item{WORKFLOW_APPROVAL_REQUEST - This is a request for an approval action. Either by a frontend user or an
API request.}
     \item{WORKFLOW_EXTERNAL_CONDITION_REQUEST - This is an external condition request. It is pending an external system to
determine whether an action should trigger or not.}
     \item{UNKNOWN}
   }
  }
}

\strong{WorkflowRequestAction}

Represents the actions that can be performed on WorkflowRequest objects. May take one of the actions listed below.
\href{https://developers.google.com/doubleclick-publishers/docs/reference/v201508/WorkflowRequestService.WorkflowRequestAction}{Google Documentation for WorkflowRequestAction}
\describe{
 \item{ApproveWorkflowApprovalRequests}{The action used to approve WorkflowApprovalRequest workflow approval
requests.}
 \item{SkipWorkflowExternalConditionRequests}{The action to skip WorkflowExternalConditionRequest workflow external
condition requests. Doing so means that the rule did not trigger.}
 \item{TriggerWorkflowExternalConditionRequests}{The action to trigger WorkflowExternalConditionRequest workflow external
condition requests.}
 \item{RejectWorkflowApprovalRequests}{The action used to reject WorkflowApprovalRequest workflow approval
requests.}
}

\strong{WorkflowRequestError}

Lists errors associated with WorkflowRequest workflow requests.
\href{https://developers.google.com/doubleclick-publishers/docs/reference/v201508/WorkflowRequestService.WorkflowRequestError}{Google Documentation for WorkflowRequestError}
\describe{
 \item{extends ApiError}{see documentation for ApiError}
 \item{reason}{a WorkflowRequestError.Reason - The error reason represented by an enum. This can take one of the following values: 
   \itemize{
     \item{REQUEST_ALREADY_PROCESSED - The request has already been processed.}
     \item{WORKFLOW_NOT_IN_PROGRESS - The workflow is not in progress.}
     \item{INVALID_REQUEST_ID - The request ID is invalid.}
     \item{INVALID_ACTION - The action could not be performed on a request of that type.}
     \item{TYPE_REQUIRED - Querying for workflow requests without specifying type is not allowed.}
     \item{UNKNOWN - The value returned if the actual value is not exposed by the requested API
version.}
   }
  }
}

\strong{WorkflowRequestPage}

Captures a page of WorkflowRequest objects.
\href{https://developers.google.com/doubleclick-publishers/docs/reference/v201508/WorkflowRequestService.WorkflowRequestPage}{Google Documentation for WorkflowRequestPage}
\describe{
 \item{results}{a WorkflowRequest - The collection of workflow requests contained within this page.}
 \item{startIndex}{a integer - The absolute index in the total result set on which this page begins.}
 \item{totalResultSetSize}{a integer - The size of the total result set to which this page belongs.}
}
}

