% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fapiao.R
\name{get_fp_acctNumber}
\alias{get_fp_acctNumber}
\title{get the acctNumber from fp}
\usage{
get_fp_acctNumber(bankNumber)
}
\arguments{
\item{bankNumber}{bank number}
}
\value{
acctNumber
}
\description{
get the acctNumber from fp
}
\examples{
get_fp_acctNumber()
}
