% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fapiao.R
\name{get_fp_addr}
\alias{get_fp_addr}
\title{get the fp addr}
\usage{
get_fp_addr(addrPhone)
}
\arguments{
\item{addrPhone}{phoneNumber}
}
\value{
return address
}
\description{
get the fp addr
}
\examples{
get_fp_addr
}
