% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fapiao.R
\name{get_fp_bank}
\alias{get_fp_bank}
\title{get the fp bankinto}
\usage{
get_fp_bank(bankNumber)
}
\arguments{
\item{bankNumber}{bankno}
}
\value{
bank
}
\description{
get the fp bankinto
}
\examples{
get_fp_bank()
}
