% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fapiao.R
\name{get_fp_phone}
\alias{get_fp_phone}
\title{get the phone info from fp}
\usage{
get_fp_phone(addrPhone)
}
\arguments{
\item{addrPhone}{addrphone}
}
\value{
phone
}
\description{
get the phone info from fp
}
\examples{
get_fp_phone()
}
