% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfp_utils.R
\name{make_request_body}
\alias{make_request_body}
\title{Format SOAP Request Body}
\usage{
make_request_body(service, root_name, data = NULL, verbose = T)
}
\arguments{
\item{service}{a character string matching one of the API
services}

\item{root_name}{a character string to be put in the 
topmost level of the created XML hierarchy}

\item{data}{a \code{list} or \code{data.frame} to create
XML in the request}

\item{verbose}{a boolean indicating whether the xml body should be printed
during the function call}
}
\value{
a character string of XML with service name
as an attribute
}
\description{
Receive data for a service and build the Body of text 
to include in a SOAP request.
}
\keyword{internal}

