% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfp_utils_xml.R
\name{parse_soap_response}
\alias{parse_soap_response}
\title{Parse SOAP Response Body}
\usage{
parse_soap_response(httr_response, resp_element, as_df)
}
\arguments{
\item{httr_response}{an httr response from a POST to the API}

\item{resp_element}{character; the name of the response body element}

\item{as_df}{logical; indicating whether the results should be a list or data.frame}
}
\value{
a \code{list} or \code{data.frame}
}
\description{
Parse the returned XML in the SOAP Response Body.
}
\note{
This function is meant to be used internally. Only use when debugging.
}
\keyword{internal}
