% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdfp.R
\docType{package}
\name{rdfp}
\alias{rdfp}
\alias{rdfp-package}
\title{\code{rdfp} package}
\description{
DoubleClick for Publishers API from R
}
\details{
The DoubleClick for Publishers (DFP) API consists of roughly 50 services. 
Each service is written to handle a particular set of operations within the 
API and grouped.

The official documentation, provided by Google, is available at:
\url{https://developers.google.com/doubleclick-publishers/docs/rel_notes}

This package attempts to scrape the functionality and documentation from the 
references and WSDL to provide an interface via R. Each function has its own
documentation, but additional material can be found in the 
\href{https://github.com/reportmort/rdfp}{README} on GitHub
}
