% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfp_auth.R
\name{token_exists}
\alias{token_exists}
\title{Check if authorization currently in force}
\usage{
token_exists(verbose = TRUE)
}
\value{
logical
}
\description{
Check if authorization currently in force
}
\note{
This function is meant to be used internally. Only use when debugging.
}
\keyword{internal}
