% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AdRuleService.R
\name{dfp_createAdRules}
\alias{dfp_createAdRules}
\title{AdRuleService}
\usage{
dfp_createAdRules(request_data, as_df = TRUE, verbose = FALSE)
}
\arguments{
\item{request_data}{a \code{list} or \code{data.frame} of data elements
to be formatted for a SOAP
request (XML format, but passed as character string)}

\item{as_df}{a boolean indicating whether to attempt to parse the result into
a \code{data.frame}}

\item{verbose}{a boolean indicating whether to print the service URL and POSTed XML}
}
\value{
a \code{data.frame} or \code{list} containing all the elements of a createAdRulesResponse
}
\description{
Provides methods for creating, updating and retrieving AdRule objects.
}
\details{
Ad rules contain data that the ad server uses to generate a playlist of
video ads.

createAdRules

Creates new AdRule objects.
}
\examples{
\dontrun{
 res <- dfp_createAdRules(request_data)
}
}
\seealso{
\href{https://developers.google.com/doubleclick-publishers/docs/reference/v201802/AdRuleService#createAdRules}{Google Documentation for createAdRules}
}
