% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CompanyService.R
\name{dfp_createCompanies}
\alias{dfp_createCompanies}
\title{createCompanies}
\usage{
dfp_createCompanies(request_data)
}
\arguments{
\item{request_data}{a \code{list} or \code{data.frame} of data elements
to be formatted for a SOAP request (XML format, but passed as character string)#' @return a \code{list} containing all the elements of a createCompaniesResponse}
}
\description{
Creates new Company objects.
}

