% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ContentMetadataKeyHierarchyService.R
\name{dfp_createContentMetadataKeyHierarchies}
\alias{dfp_createContentMetadataKeyHierarchies}
\title{createContentMetadataKeyHierarchies}
\usage{
dfp_createContentMetadataKeyHierarchies(request_data, as_df=TRUE)
}
\arguments{
\item{request_data}{a \code{list} or \code{data.frame} of data elements
to be formatted for a SOAP request (XML format, but passed as character string)}

\item{as_df}{a boolean indicating whether to attempt to parse the result into a \code{data.frame}}
}
\value{
a \code{data.frame} or \code{list} containing all the elements of a createContentMetadataKeyHierarchiesResponse
}
\description{
Creates new ContentMetadataKeyHierarchy objects. The following fields are required: <ul> <li>ContentMetadataKeyHierarchy id</li> <li>ContentMetadataKeyHierarchy name</li> <li>ContentMetadataKeyHierarchy hierarchyLevels</li> </ul>
}
\seealso{
\href{https://developers.google.com/doubleclick-publishers/docs/reference/v201508/ContentMetadataKeyHierarchyService#createContentMetadataKeyHierarchies}{Google Documentation for createContentMetadataKeyHierarchies}
}

