% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreativeSetService.R
\name{dfp_createCreativeSet}
\alias{dfp_createCreativeSet}
\title{createCreativeSet}
\usage{
dfp_createCreativeSet(request_data)
}
\arguments{
\item{request_data}{a \code{list} or \code{data.frame} of data elements
to be formatted for a SOAP request (XML format, but passed as character string)}
}
\value{
a \code{list} containing all the elements of a createCreativeSetResponse
}
\description{
Creates a new CreativeSet.
}

