% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreativeSetService.R
\name{dfp_createCreativeSet}
\alias{dfp_createCreativeSet}
\title{CreativeSetService}
\usage{
dfp_createCreativeSet(request_data, as_df = TRUE, verbose = FALSE)
}
\arguments{
\item{request_data}{a \code{list} or \code{data.frame} of data elements
to be formatted for a SOAP
request (XML format, but passed as character string)}

\item{as_df}{a boolean indicating whether to attempt to parse the result into
a \code{data.frame}}

\item{verbose}{a boolean indicating whether to print the service URL and POSTed XML}
}
\value{
a \code{data.frame} or \code{list} containing all the elements of a createCreativeSetResponse
}
\description{
Provides methods for adding, updating and retrieving CreativeSet objects.
}
\details{
createCreativeSet

Creates a new CreativeSet.
}
\examples{
\dontrun{
 res <- dfp_createCreativeSet(request_data)
}
}
\seealso{
\href{https://developers.google.com/ad-manager/api/reference/v201905/CreativeSetService#createCreativeSet}{Google Documentation for createCreativeSet}
}
