% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CustomFieldService.R
\name{dfp_createCustomFieldOptions}
\alias{dfp_createCustomFieldOptions}
\title{createCustomFieldOptions}
\usage{
dfp_createCustomFieldOptions(request_data, as_df=TRUE, verbose=FALSE)
}
\arguments{
\item{request_data}{a \code{list} or \code{data.frame} of data elements
to be formatted for a SOAP request (XML format, but passed as character string)}

\item{as_df}{a boolean indicating whether to attempt to parse the result into a \code{data.frame}}

\item{verbose}{a boolean indicating whether to print the service URL and POSTed XML}
}
\value{
a \code{data.frame} or \code{list} containing all the elements of a createCustomFieldOptionsResponse
}
\description{
Creates new CustomFieldOption objects. The following fields are required: <ul> <li>CustomFieldOption displayName</li> <li>CustomFieldOption customFieldId</li> </ul>
}
\seealso{
\href{https://developers.google.com/doubleclick-publishers/docs/reference/v201508/CustomFieldService#createCustomFieldOptions}{Google Documentation for createCustomFieldOptions}
}

