% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OrderService.R
\name{dfp_createOrders}
\alias{dfp_createOrders}
\title{createOrders}
\usage{
dfp_createOrders(request_data)
}
\arguments{
\item{request_data}{a \code{list} or \code{data.frame} of data elements
to be formatted for a SOAP request (XML format, but passed as character string)#' @return a \code{list} containing all the elements of a createOrdersResponse}
}
\description{
Creates new Order objects.
}

