% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlacementService.R
\name{dfp_createPlacements}
\alias{dfp_createPlacements}
\title{createPlacements}
\usage{
dfp_createPlacements(request_data)
}
\arguments{
\item{request_data}{a \code{list} or \code{data.frame} of data elements
to be formatted for a SOAP request (XML format, but passed as character string)}
}
\value{
a \code{list} containing all the elements of a createPlacementsResponse
}
\description{
Creates new Placement objects.
}
\seealso{
\href{https://developers.google.com/doubleclick-publishers/docs/reference/v201508/PlacementService#createPlacements}{Google Documentation for createPlacements}
}

