% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProductPackageItemService.R
\name{dfp_createProductPackageItems}
\alias{dfp_createProductPackageItems}
\title{createProductPackageItems}
\usage{
dfp_createProductPackageItems(request_data)
}
\arguments{
\item{request_data}{a \code{list} or \code{data.frame} of data elements
to be formatted for a SOAP request (XML format, but passed as character string)#' @return a \code{list} containing all the elements of a createProductPackageItemsResponse}
}
\description{
Creates new ProductPackageItem objects.
}

