% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProductPackageService.R
\name{dfp_createProductPackages}
\alias{dfp_createProductPackages}
\title{createProductPackages}
\usage{
dfp_createProductPackages(request_data)
}
\arguments{
\item{request_data}{a \code{list} or \code{data.frame} of data elements
to be formatted for a SOAP request (XML format, but passed as character string)}
}
\value{
a \code{list} containing all the elements of a createProductPackagesResponse
}
\description{
Creates new ProductPackage objects.
}

