% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProductTemplateService.R
\name{dfp_createProductTemplates}
\alias{dfp_createProductTemplates}
\title{createProductTemplates}
\usage{
dfp_createProductTemplates(request_data, as_df=FALSE)
}
\arguments{
\item{request_data}{a \code{list} or \code{data.frame} of data elements
to be formatted for a SOAP request (XML format, but passed as character string)}

\item{as_df}{a boolean indicating whether to attempt to parse the result into a \code{data.frame}}
}
\value{
a \code{list} or \code{data.frame} containing all the elements of a createProductTemplatesResponse
}
\description{
Creates new ProductTemplate objects.
}
\seealso{
\href{https://developers.google.com/doubleclick-publishers/docs/reference/v201508/ProductTemplateService#createProductTemplates}{Google Documentation for createProductTemplates}
}

