% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RateCardService.R
\name{dfp_createRateCards}
\alias{dfp_createRateCards}
\title{createRateCards}
\usage{
dfp_createRateCards(request_data)
}
\arguments{
\item{request_data}{a \code{list} or \code{data.frame} of data elements
to be formatted for a SOAP request (XML format, but passed as character string)}
}
\value{
a \code{list} containing all the elements of a createRateCardsResponse
}
\description{
Creates a list of RateCard objects. Rate cards must be activated before being associated with proposal line items and products.
}

