% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TeamService.R
\name{dfp_createTeams}
\alias{dfp_createTeams}
\title{createTeams}
\usage{
dfp_createTeams(request_data)
}
\arguments{
\item{request_data}{a \code{list} or \code{data.frame} of data elements
to be formatted for a SOAP request (XML format, but passed as character string)#' @return a \code{list} containing all the elements of a createTeamsResponse}
}
\description{
Creates new Team objects. The following fields are required: <ul> <li>Team name</li> </ul>
}

