% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UserTeamAssociationService.R
\name{dfp_createUserTeamAssociations}
\alias{dfp_createUserTeamAssociations}
\title{createUserTeamAssociations}
\usage{
dfp_createUserTeamAssociations(request_data, as_df=FALSE)
}
\arguments{
\item{request_data}{a \code{list} or \code{data.frame} of data elements
to be formatted for a SOAP request (XML format, but passed as character string)}

\item{as_df}{a boolean indicating whether to attempt to parse the result into a \code{data.frame}}
}
\value{
a \code{list} or \code{data.frame} containing all the elements of a createUserTeamAssociationsResponse
}
\description{
Creates new UserTeamAssociation objects.
}
\seealso{
\href{https://developers.google.com/doubleclick-publishers/docs/reference/v201508/UserTeamAssociationService#createUserTeamAssociations}{Google Documentation for createUserTeamAssociations}
}

