% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NetworkService.R
\name{dfp_getAllNetworks}
\alias{dfp_getAllNetworks}
\title{getAllNetworks}
\usage{
dfp_getAllNetworks(as_df=FALSE)
}
\arguments{
\item{as_df}{a boolean indicating whether to attempt to parse the result into a \code{data.frame}}
}
\value{
a \code{list} or \code{data.frame} containing all the elements of a getAllNetworksResponse
}
\description{
Returns the list of Network objects to which the current login has access. <p> Intended to be used without a network code in the SOAP header when the login may have more than one network associated with it. </p>
}
\seealso{
\href{https://developers.google.com/doubleclick-publishers/docs/reference/v201508/NetworkService#getAllNetworks}{Google Documentation for getAllNetworks}
}

