% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NetworkService.R
\name{dfp_getAllNetworks}
\alias{dfp_getAllNetworks}
\title{getAllNetworks}
\usage{
dfp_getAllNetworks()
}
\value{
a \code{list} containing all the elements of a getAllNetworksResponse
}
\description{
Returns the list of Network objects to which the current login has access. <p> Intended to be used without a network code in the SOAP header when the login may have more than one network associated with it. </p>
}

