% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UserService.R
\name{dfp_getAllRoles}
\alias{dfp_getAllRoles}
\title{getAllRoles}
\usage{
dfp_getAllRoles()
}
\value{
a \code{list} containing all the elements of a getAllRolesResponse
}
\description{
Returns the Role objects that are defined for the users of the network.
}

