% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UserService.R
\name{dfp_getAllRoles}
\alias{dfp_getAllRoles}
\title{getAllRoles}
\usage{
dfp_getAllRoles(as_df = TRUE, verbose = FALSE)
}
\arguments{
\item{as_df}{a boolean indicating whether to attempt to parse the result into
a \code{data.frame}}

\item{verbose}{a boolean indicating whether to print the service URL and POSTed XML}
}
\value{
a \code{data.frame} or \code{list} containing all the elements of a getAllRolesResponse
}
\description{
Returns the Role objects that are defined for the users of the network. Returns the Role objects that are defined for the users of the network. @return the roles defined for the user's network Returns the Role objects that are defined for the users of the network. @return the roles defined for the user's network
}
\examples{
\dontrun{
 res <- dfp_getAllRoles()
}
}
\seealso{
\href{https://developers.google.com/doubleclick-publishers/docs/reference/v201802/UserService#getAllRoles}{Google Documentation for getAllRoles}
}
