% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AudienceSegmentService.R
\name{dfp_getAudienceSegmentsByStatement}
\alias{dfp_getAudienceSegmentsByStatement}
\title{getAudienceSegmentsByStatement}
\usage{
dfp_getAudienceSegmentsByStatement(request_data, as_df=FALSE)
}
\arguments{
\item{request_data}{a \code{list} or \code{data.frame} of data elements
to be formatted for a SOAP request (XML format, but passed as character string)}

\item{as_df}{a boolean indicating whether to attempt to parse the result into a \code{data.frame}}
}
\value{
a \code{list} or \code{data.frame} containing all the elements of a getAudienceSegmentsByStatementResponse
}
\description{
Gets an AudienceSegmentPage of AudienceSegment objects that satisfy the given Statement query. The following fields are supported for filtering:  segments 
\itemize{
  \item{id}
  \item{name}
  \item{status}
  \item{type}
  \item{size}
  \item{dataProviderName}
  \item{approvalStatus}
  \item{cost}
  \item{startDateTime}
  \item{endDateTime}
}
}
\seealso{
\href{https://developers.google.com/doubleclick-publishers/docs/reference/v201508/AudienceSegmentService#getAudienceSegmentsByStatement}{Google Documentation for getAudienceSegmentsByStatement}
}

