% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ForecastService.R
\name{dfp_getAvailabilityForecast}
\alias{dfp_getAvailabilityForecast}
\title{getAvailabilityForecast}
\usage{
dfp_getAvailabilityForecast()
}
\value{
a \code{list} containing all the elements of a getAvailabilityForecastResponse
}
\description{
Gets the availability forecast for a ProspectiveLineItem. An availability forecast reports the maximum number of available units that the line item can book, and the total number of units matching the line item's targeting. <p>Note: Beginning in v201502, this replaces the previous getForecast method.
}

