% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ContentService.R
\name{dfp_getContentByStatementAndCustomTargetingValue}
\alias{dfp_getContentByStatementAndCustomTargetingValue}
\title{getContentByStatementAndCustomTargetingValue}
\usage{
dfp_getContentByStatementAndCustomTargetingValue(as_df=FALSE)
}
\arguments{
\item{as_df}{a boolean indicating whether to attempt to parse the result into a \code{data.frame}}
}
\value{
a \code{list} or \code{data.frame} containing all the elements of a getContentByStatementAndCustomTargetingValueResponse
}
\description{
Gets a ContentPage of Content objects that satisfy the given Statement query. Additionally, filters on the given value ID and key ID that the value belongs to. The following fields are supported for filtering:  filter a set of content 
\itemize{
  \item{id}
  \item{status}
  \item{name}
  \item{lastModifiedDateTime}
}
}
\seealso{
\href{https://developers.google.com/doubleclick-publishers/docs/reference/v201508/ContentService#getContentByStatementAndCustomTargetingValue}{Google Documentation for getContentByStatementAndCustomTargetingValue}
}

