% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreativeTemplateService.R
\name{dfp_getCreativeTemplatesByStatement}
\alias{dfp_getCreativeTemplatesByStatement}
\title{getCreativeTemplatesByStatement}
\usage{
dfp_getCreativeTemplatesByStatement(request_data)
}
\arguments{
\item{request_data}{a \code{list} or \code{data.frame} of data elements
to be formatted for a SOAP request (XML format, but passed as character string)}
}
\value{
a \code{list} containing all the elements of a getCreativeTemplatesByStatementResponse
}
\description{
Gets a CreativeTemplatePage of CreativeTemplate objects that satisfy the given Statement query. The following fields are supported for filtering:  a set of creative templates. 
\itemize{
  \item{id}
  \item{name}
  \item{type}
  \item{status}
}
}

