% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NetworkService.R
\name{dfp_getCurrentNetwork}
\alias{dfp_getCurrentNetwork}
\title{getCurrentNetwork}
\usage{
dfp_getCurrentNetwork()
}
\value{
a \code{list} containing all the elements of a getCurrentNetworkResponse
}
\description{
Returns the current network for which requests are being made.
}
\seealso{
\href{https://developers.google.com/doubleclick-publishers/docs/reference/v201508/NetworkService#getCurrentNetwork}{Google Documentation for getCurrentNetwork}
}

