% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NetworkService.R
\name{dfp_getCurrentNetwork}
\alias{dfp_getCurrentNetwork}
\title{getCurrentNetwork}
\usage{
dfp_getCurrentNetwork(as_df = TRUE, verbose = FALSE)
}
\arguments{
\item{as_df}{a boolean indicating whether to attempt to parse the result into
a \code{data.frame}}

\item{verbose}{a boolean indicating whether to print the service URL and POSTed XML}
}
\value{
a \code{data.frame} or \code{list} containing all the elements of a getCurrentNetworkResponse
}
\description{
Returns the current network for which requests are being made. Returns the current network for which requests are being made. @return the network for which the user is currently making the request Returns the current network for which requests are being made. @return the network for which the user is currently making the request
}
\examples{
\dontrun{
 res <- dfp_getCurrentNetwork()
}
}
\seealso{
\href{https://developers.google.com/ad-manager/api/reference/v201905/NetworkService#getCurrentNetwork}{Google Documentation for getCurrentNetwork}
}
