% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UserService.R
\name{dfp_getCurrentUser}
\alias{dfp_getCurrentUser}
\title{getCurrentUser}
\usage{
dfp_getCurrentUser(as_df=TRUE)
}
\arguments{
\item{as_df}{a boolean indicating whether to attempt to parse the result into a \code{data.frame}}
}
\value{
a \code{data.frame} or \code{list} containing all the elements of a getCurrentUserResponse
}
\description{
Returns the current User.
}
\seealso{
\href{https://developers.google.com/doubleclick-publishers/docs/reference/v201508/UserService#getCurrentUser}{Google Documentation for getCurrentUser}
}

