% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UserService.R
\name{dfp_getCurrentUser}
\alias{dfp_getCurrentUser}
\title{getCurrentUser}
\usage{
dfp_getCurrentUser(as_df = TRUE, verbose = FALSE)
}
\arguments{
\item{as_df}{a boolean indicating whether to attempt to parse the result into
a \code{data.frame}}

\item{verbose}{a boolean indicating whether to print the service URL and POSTed XML}
}
\value{
a \code{data.frame} or \code{list} containing all the elements of a getCurrentUserResponse
}
\description{
Returns the current User. Returns the current User. @return the current user Returns the current User. @return the current user
}
\examples{
\dontrun{
 res <- dfp_getCurrentUser()
}
}
\seealso{
\href{https://developers.google.com/ad-manager/api/reference/v201811/UserService#getCurrentUser}{Google Documentation for getCurrentUser}
}
