% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CustomFieldService.R
\name{dfp_getCustomFieldOption}
\alias{dfp_getCustomFieldOption}
\title{getCustomFieldOption}
\usage{
dfp_getCustomFieldOption(request_data, as_df = TRUE, verbose = FALSE)
}
\arguments{
\item{request_data}{a \code{list} or \code{data.frame} of data elements
to be formatted for a SOAP
request (XML format, but passed as character string)}

\item{as_df}{a boolean indicating whether to attempt to parse the result into
a \code{data.frame}}

\item{verbose}{a boolean indicating whether to print the service URL and POSTed XML}
}
\value{
a \code{data.frame} or \code{list} containing all the elements of a getCustomFieldOptionResponse
}
\description{
Returns the CustomFieldOption uniquely identified by the given ID.
}
\examples{
\dontrun{
 res <- dfp_getCustomFieldOption(request_data)
}
}
\seealso{
\href{https://developers.google.com/doubleclick-publishers/docs/reference/v201802/CustomFieldService#getCustomFieldOption}{Google Documentation for getCustomFieldOption}
}
