% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CustomFieldService.R
\name{dfp_getCustomFieldsByStatement}
\alias{dfp_getCustomFieldsByStatement}
\title{getCustomFieldsByStatement}
\usage{
dfp_getCustomFieldsByStatement(request_data)
}
\arguments{
\item{request_data}{a \code{list} or \code{data.frame} of data elements
to be formatted for a SOAP request (XML format, but passed as character string)#' @return a \code{list} containing all the elements of a getCustomFieldsByStatementResponse}
}
\description{
Gets a CustomFieldPage of CustomField objects that satisfy the given Statement query. The following fields are supported for filtering:  a set of custom fields. 
\itemize{
  \item{id}
  \item{entityType}
  \item{name}
  \item{isActive}
  \item{visibility}
}
}

