% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CustomTargetingService.R
\name{dfp_getCustomTargetingKeysByStatement}
\alias{dfp_getCustomTargetingKeysByStatement}
\title{getCustomTargetingKeysByStatement}
\usage{
dfp_getCustomTargetingKeysByStatement(request_data, as_df=TRUE)
}
\arguments{
\item{request_data}{a \code{list} or \code{data.frame} of data elements
to be formatted for a SOAP request (XML format, but passed as character string)}

\item{as_df}{a boolean indicating whether to attempt to parse the result into a \code{data.frame}}
}
\value{
a \code{data.frame} or \code{list} containing all the elements of a getCustomTargetingKeysByStatementResponse
}
\description{
Gets a CustomTargetingKeyPage of CustomTargetingKey objects that satisfy the given Statement query. The following fields are supported for filtering:  a set of custom targeting keys 
\itemize{
  \item{NA}
  \item{id}
  \item{name}
  \item{displayName}
  \item{type}
}
}
\seealso{
\href{https://developers.google.com/doubleclick-publishers/docs/reference/v201508/CustomTargetingService#getCustomTargetingKeysByStatement}{Google Documentation for getCustomTargetingKeysByStatement}
}

