% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LineItemTemplateService.R
\name{dfp_getLineItemTemplatesByStatement}
\alias{dfp_getLineItemTemplatesByStatement}
\title{getLineItemTemplatesByStatement}
\usage{
dfp_getLineItemTemplatesByStatement(request_data)
}
\arguments{
\item{request_data}{a \code{list} or \code{data.frame} of data elements
to be formatted for a SOAP request (XML format, but passed as character string)#' @return a \code{list} containing all the elements of a getLineItemTemplatesByStatementResponse}
}
\description{
Gets a LineItemTemplatePage of LineItemTemplate objects that satisfy the given Statement query. The following fields are supported for filtering:  a set of line item templates 
\itemize{
  \item{id}
}
}

