% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LineItemTemplateService.R
\name{dfp_getLineItemTemplatesByStatement}
\alias{dfp_getLineItemTemplatesByStatement}
\title{LineItemTemplateService}
\usage{
dfp_getLineItemTemplatesByStatement(request_data, as_df = TRUE,
  verbose = FALSE)
}
\arguments{
\item{request_data}{a \code{list} or \code{data.frame} of data elements
to be formatted for a SOAP
request (XML format, but passed as character string)}

\item{as_df}{a boolean indicating whether to attempt to parse the result into
a \code{data.frame}}

\item{verbose}{a boolean indicating whether to print the service URL and POSTed XML}
}
\value{
a \code{data.frame} or \code{list} containing all the elements of a getLineItemTemplatesByStatementResponse
}
\description{
Provides methods for creating, updating and retrieving LineItemTemplate
objects.
}
\details{
getLineItemTemplatesByStatement

Gets a LineItemTemplatePage of LineItemTemplate objects that satisfy the given Statement query. The following fields are supported for filtering: 
\itemize{
  \item{id}
}
}
\examples{
\dontrun{
 dat <- list(filterStatement=list('query'="WHERE status='ACTIVE'")) 
 res <- dfp_getLineItemTemplatesByStatement(dat)
}
}
\seealso{
\href{https://developers.google.com/ad-manager/api/reference/v201811/LineItemTemplateService#getLineItemTemplatesByStatement}{Google Documentation for getLineItemTemplatesByStatement}
}
