% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlacementService.R
\name{dfp_getPlacementsByStatement}
\alias{dfp_getPlacementsByStatement}
\title{getPlacementsByStatement}
\usage{
dfp_getPlacementsByStatement(request_data, as_df = TRUE, verbose = FALSE)
}
\arguments{
\item{request_data}{a \code{list} or \code{data.frame} of data elements
to be formatted for a SOAP
request (XML format, but passed as character string)}

\item{as_df}{a boolean indicating whether to attempt to parse the result into
a \code{data.frame}}

\item{verbose}{a boolean indicating whether to print the service URL and POSTed XML}
}
\value{
a \code{data.frame} or \code{list} containing all the elements of a getPlacementsByStatementResponse
}
\description{
Gets a PlacementPage of Placement objects that satisfy the given Statement query. The following fields are supported for filtering: 
\itemize{
  \item{description}
  \item{id}
  \item{isAdSenseTargetingEnabled}
  \item{name}
  \item{placementCode}
  \item{status}
  \item{lastModifiedDateTime}
}
}
\examples{
\dontrun{
 dat <- list(filterStatement=list('query'="WHERE status='ACTIVE'")) 
 res <- dfp_getPlacementsByStatement(dat)
}
}
\seealso{
\href{https://developers.google.com/doubleclick-publishers/docs/reference/v201802/PlacementService#getPlacementsByStatement}{Google Documentation for getPlacementsByStatement}
}
