% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LineItemCreativeAssociationService.R
\name{dfp_getPreviewUrl}
\alias{dfp_getPreviewUrl}
\title{getPreviewUrl}
\usage{
dfp_getPreviewUrl(as_df=FALSE)
}
\arguments{
\item{as_df}{a boolean indicating whether to attempt to parse the result into a \code{data.frame}}
}
\value{
a \code{list} or \code{data.frame} containing all the elements of a getPreviewUrlResponse
}
\description{
Returns an insite preview URL that references the specified site URL with the specified creative from the association served to it. For Creative Set previewing you may specify the master creative Id. creative served to it
}
\seealso{
\href{https://developers.google.com/doubleclick-publishers/docs/reference/v201508/LineItemCreativeAssociationService#getPreviewUrl}{Google Documentation for getPreviewUrl}
}

