% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LineItemCreativeAssociationService.R
\name{dfp_getPreviewUrl}
\alias{dfp_getPreviewUrl}
\title{getPreviewUrl}
\usage{
dfp_getPreviewUrl(as_df = TRUE, verbose = FALSE)
}
\arguments{
\item{as_df}{a boolean indicating whether to attempt to parse the result into
a \code{data.frame}}

\item{verbose}{a boolean indicating whether to print the service URL and POSTed XML}
}
\value{
a \code{data.frame} or \code{list} containing all the elements of a getPreviewUrlResponse
}
\description{
Returns an insite preview URL that references the specified site URL with the specified creative from the association served to it. For Creative Set previewing you may specify the master creative Id.
}
\examples{
\dontrun{
 res <- dfp_getPreviewUrl()
}
}
\seealso{
\href{https://developers.google.com/ad-manager/api/reference/v201905/LineItemCreativeAssociationService#getPreviewUrl}{Google Documentation for getPreviewUrl}
}
