% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProductTemplateService.R
\name{dfp_getProductTemplatesByStatement}
\alias{dfp_getProductTemplatesByStatement}
\title{getProductTemplatesByStatement}
\usage{
dfp_getProductTemplatesByStatement(request_data, as_df = TRUE,
  verbose = FALSE)
}
\arguments{
\item{request_data}{a \code{list} or \code{data.frame} of data elements
to be formatted for a SOAP
request (XML format, but passed as character string)}

\item{as_df}{a boolean indicating whether to attempt to parse the result into
a \code{data.frame}}

\item{verbose}{a boolean indicating whether to print the service URL and POSTed XML}
}
\value{
a \code{data.frame} or \code{list} containing all the elements of a getProductTemplatesByStatementResponse
}
\description{
Gets a ProductTemplatePage of ProductTemplate objects that satisfy the filtering criteria specified by given Statement query. The following fields are supported for filtering: 
\itemize{
  \item{id}
  \item{name}
  \item{nameMacro}
  \item{description}
  \item{status}
  \item{lastModifiedDateTime}
  \item{lineItemType}
  \item{productType}
  \item{rateType}
}
}
\examples{
\dontrun{
 dat <- list(filterStatement=list('query'="WHERE status='ACTIVE'")) 
 res <- dfp_getProductTemplatesByStatement(dat)
}
}
\seealso{
\href{https://developers.google.com/ad-manager/api/reference/v201905/ProductTemplateService#getProductTemplatesByStatement}{Google Documentation for getProductTemplatesByStatement}
}
