% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RateCardService.R
\name{dfp_getRateCardsByStatement}
\alias{dfp_getRateCardsByStatement}
\title{getRateCardsByStatement}
\usage{
dfp_getRateCardsByStatement(request_data, as_df = TRUE,
  verbose = FALSE)
}
\arguments{
\item{request_data}{a \code{list} or \code{data.frame} of data elements
to be formatted for a SOAP
request (XML format, but passed as character string)}

\item{as_df}{a boolean indicating whether to attempt to parse the result into
a \code{data.frame}}

\item{verbose}{a boolean indicating whether to print the service URL and POSTed XML}
}
\value{
a \code{data.frame} or \code{list} containing all the elements of a getRateCardsByStatementResponse
}
\description{
Gets a RateCardPage of RateCard objects that satisfy the given Statement query. The following fields are supported for filtering: 
\itemize{
  \item{ForMarketplace}
  \item{Id}
  \item{LastModifiedDateTime}
  \item{Name}
  \item{Status}
}
}
\examples{
\dontrun{
 dat <- list(filterStatement=list('query'="WHERE status='ACTIVE'")) 
 res <- dfp_getRateCardsByStatement(dat)
}
}
\seealso{
\href{https://developers.google.com/ad-manager/api/reference/v201905/RateCardService#getRateCardsByStatement}{Google Documentation for getRateCardsByStatement}
}
