% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReconciliationLineItemReportService.R
\name{dfp_getReconciliationLineItemReportsByStatement}
\alias{dfp_getReconciliationLineItemReportsByStatement}
\title{getReconciliationLineItemReportsByStatement}
\usage{
dfp_getReconciliationLineItemReportsByStatement(request_data, as_df=FALSE)
}
\arguments{
\item{request_data}{a \code{list} or \code{data.frame} of data elements
to be formatted for a SOAP request (XML format, but passed as character string)}

\item{as_df}{a boolean indicating whether to attempt to parse the result into a \code{data.frame}}
}
\value{
a \code{list} or \code{data.frame} containing all the elements of a getReconciliationLineItemReportsByStatementResponse
}
\description{
Gets a ReconciliationLineItemReportPage of ReconciliationLineItemReport objects that satisfy the given Statement query. The following fields are supported for filtering:  
\itemize{
  \item{id}
  \item{reconciliationReportId}
  \item{orderId}
  \item{proposalId}
  \item{lineItemId}
  \item{proposalLineItemId}
}
}
\seealso{
\href{https://developers.google.com/doubleclick-publishers/docs/reference/v201508/ReconciliationLineItemReportService#getReconciliationLineItemReportsByStatement}{Google Documentation for getReconciliationLineItemReportsByStatement}
}

