% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReconciliationLineItemReportService.R
\name{dfp_getReconciliationLineItemReportsByStatement}
\alias{dfp_getReconciliationLineItemReportsByStatement}
\title{ReconciliationLineItemReportService}
\usage{
dfp_getReconciliationLineItemReportsByStatement(request_data,
  as_df = TRUE, verbose = FALSE)
}
\arguments{
\item{request_data}{a \code{list} or \code{data.frame} of data elements
to be formatted for a SOAP
request (XML format, but passed as character string)}

\item{as_df}{a boolean indicating whether to attempt to parse the result into
a \code{data.frame}}

\item{verbose}{a boolean indicating whether to print the service URL and POSTed XML}
}
\value{
a \code{data.frame} or \code{list} containing all the elements of a getReconciliationLineItemReportsByStatementResponse
}
\description{
Provides methods for retrieving and updating ReconciliationLineItemReport
objects.
}
\details{
getReconciliationLineItemReportsByStatement

Gets a ReconciliationLineItemReportPage of ReconciliationLineItemReport objects that satisfy the given Statement query. The following fields are supported for filtering: 
\itemize{
  \item{id}
  \item{reconciliationReportId}
  \item{orderId}
  \item{proposalId}
  \item{lineItemId}
  \item{proposalLineItemId}
}
}
\examples{
\dontrun{
 dat <- list(filterStatement=list('query'="WHERE status='ACTIVE'")) 
 res <- dfp_getReconciliationLineItemReportsByStatement(dat)
}
}
\seealso{
\href{https://developers.google.com/ad-manager/api/reference/v201811/ReconciliationLineItemReportService#getReconciliationLineItemReportsByStatement}{Google Documentation for getReconciliationLineItemReportsByStatement}
}
