% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReconciliationOrderReportService.R
\name{dfp_getReconciliationOrderReportsByStatement}
\alias{dfp_getReconciliationOrderReportsByStatement}
\title{ReconciliationOrderReportService}
\usage{
dfp_getReconciliationOrderReportsByStatement(request_data, as_df = TRUE,
  verbose = FALSE)
}
\arguments{
\item{request_data}{a \code{list} or \code{data.frame} of data elements
to be formatted for a SOAP
request (XML format, but passed as character string)}

\item{as_df}{a boolean indicating whether to attempt to parse the result into
a \code{data.frame}}

\item{verbose}{a boolean indicating whether to print the service URL and POSTed XML}
}
\value{
a \code{data.frame} or \code{list} containing all the elements of a getReconciliationOrderReportsByStatementResponse
}
\description{
Provides methods for retrieving, reconciling, and reverting
ReconciliationOrderReport objects.
}
\details{
getReconciliationOrderReportsByStatement

Gets an ReconciliationOrderReportPage of ReconciliationOrderReport objects that satisfy the given Statement query. The following fields are supported for filtering: 
\itemize{
  \item{reconciliationReportId}
  \item{id}
  \item{orderId}
  \item{proposalId}
  \item{status}
  \item{submissionDateTime}
  \item{submitterId}
}
}
\examples{
\dontrun{
 dat <- list(filterStatement=list('query'="WHERE status='ACTIVE'")) 
 res <- dfp_getReconciliationOrderReportsByStatement(dat)
}
}
\seealso{
\href{https://developers.google.com/doubleclick-publishers/docs/reference/v201802/ReconciliationOrderReportService#getReconciliationOrderReportsByStatement}{Google Documentation for getReconciliationOrderReportsByStatement}
}
