% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReconciliationReportService.R
\name{dfp_getReconciliationReportsByStatement}
\alias{dfp_getReconciliationReportsByStatement}
\title{ReconciliationReportService}
\usage{
dfp_getReconciliationReportsByStatement(request_data, as_df = TRUE,
  verbose = FALSE)
}
\arguments{
\item{request_data}{a \code{list} or \code{data.frame} of data elements
to be formatted for a SOAP
request (XML format, but passed as character string)}

\item{as_df}{a boolean indicating whether to attempt to parse the result into
a \code{data.frame}}

\item{verbose}{a boolean indicating whether to print the service URL and POSTed XML}
}
\value{
a \code{data.frame} or \code{list} containing all the elements of a getReconciliationReportsByStatementResponse
}
\description{
Provides methods for retrieving, submitting and reverting the
ReconciliationReport objects.
A ReconciliationReport is a group of
ReconciliationReportRow objects.
}
\details{
getReconciliationReportsByStatement

Gets an ReconciliationReportPage of ReconciliationReport objects that satisfy the given Statement query. The following fields are supported for filtering.  
\itemize{
  \item{id}
  \item{status}
  \item{startDate}
}
}
\examples{
\dontrun{
 dat <- list(filterStatement=list('query'="WHERE status='ACTIVE'")) 
 res <- dfp_getReconciliationReportsByStatement(dat)
}
}
\seealso{
\href{https://developers.google.com/ad-manager/api/reference/v201811/ReconciliationReportService#getReconciliationReportsByStatement}{Google Documentation for getReconciliationReportsByStatement}
}
