% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReportService.R
\name{dfp_getReportDownloadURL}
\alias{dfp_getReportDownloadURL}
\title{getReportDownloadURL}
\usage{
dfp_getReportDownloadURL(request_data, as_df=FALSE)
}
\arguments{
\item{request_data}{a \code{list} or \code{data.frame} of data elements
to be formatted for a SOAP request (XML format, but passed as character string)}

\item{as_df}{a boolean indicating whether to attempt to parse the result into a \code{data.frame}}
}
\value{
a \code{list} or \code{data.frame} containing all the elements of a getReportDownloadURLResponse
}
\description{
Returns the URL at which the report file can be downloaded. <p> The report will be generated as a gzip archive, containing the report file itself.
}
\seealso{
\href{https://developers.google.com/doubleclick-publishers/docs/reference/v201508/ReportService#getReportDownloadURL}{Google Documentation for getReportDownloadURL}
}

