% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReportService.R
\name{dfp_getReportJobStatus}
\alias{dfp_getReportJobStatus}
\title{getReportJobStatus}
\usage{
dfp_getReportJobStatus()
}
\value{
a \code{list} containing all the elements of a getReportJobStatusResponse
}
\description{
Returns the ReportJobStatus of the report job with the specified ID.
}
\seealso{
\href{https://developers.google.com/doubleclick-publishers/docs/reference/v201508/ReportService#getReportJobStatus}{Google Documentation for getReportJobStatus}
}

