% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SharedAdUnitService.R
\name{dfp_getSharedAdUnitsByStatement}
\alias{dfp_getSharedAdUnitsByStatement}
\title{getSharedAdUnitsByStatement}
\usage{
dfp_getSharedAdUnitsByStatement(request_data)
}
\arguments{
\item{request_data}{a \code{list} or \code{data.frame} of data elements
to be formatted for a SOAP request (XML format, but passed as character string)}
}
\value{
a \code{list} containing all the elements of a getSharedAdUnitsByStatementResponse
}
\description{
Gets a SharedAdUnitPage of SharedAdUnit objects that satisfy the given Statement query. The following fields are supported for filtering:  a set of shared ad units 
\itemize{
  \item{id}
  \item{name}
}
}

