% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WorkflowRequestService.R
\name{dfp_getWorkflowRequestsByStatement}
\alias{dfp_getWorkflowRequestsByStatement}
\title{WorkflowRequestService}
\usage{
dfp_getWorkflowRequestsByStatement(request_data, as_df = TRUE,
  verbose = FALSE)
}
\arguments{
\item{request_data}{a \code{list} or \code{data.frame} of data elements
to be formatted for a SOAP
request (XML format, but passed as character string)}

\item{as_df}{a boolean indicating whether to attempt to parse the result into
a \code{data.frame}}

\item{verbose}{a boolean indicating whether to print the service URL and POSTed XML}
}
\value{
a \code{data.frame} or \code{list} containing all the elements of a getWorkflowRequestsByStatementResponse
}
\description{
Provides methods to retrieve and perform actions on WorkflowRequest objects
}
\details{
To use this service, you need to have the new sales management solution
enabled on your network. If you do not see a "Sales" tab in
\href{https://www.google.com/dfp}{DoubleClick for Publishers (DFP)}, you
will not be able to use this service.

getWorkflowRequestsByStatement

Gets a list of WorkflowRequest objects that satisfy the given Statement query. The following fields are supported for filtering: 
\itemize{
  \item{id}
  \item{workflowRuleName}
  \item{entityType}
  \item{entityId}
  \item{approvalStatus}
  \item{conditionStatus}
  \item{type}
}
}
\examples{
\dontrun{
 dat <- list(filterStatement=list('query'="WHERE status='ACTIVE'")) 
 res <- dfp_getWorkflowRequestsByStatement(dat)
}
}
\seealso{
\href{https://developers.google.com/ad-manager/api/reference/v201811/WorkflowRequestService#getWorkflowRequestsByStatement}{Google Documentation for getWorkflowRequestsByStatement}
}
