% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AdExclusionRuleService.R
\name{dfp_performAdExclusionRuleAction}
\alias{dfp_performAdExclusionRuleAction}
\title{performAdExclusionRuleAction}
\usage{
dfp_performAdExclusionRuleAction(request_data, as_df=FALSE)
}
\arguments{
\item{request_data}{a \code{list} or \code{data.frame} of data elements
to be formatted for a SOAP request (XML format, but passed as character string)}

\item{as_df}{a boolean indicating whether to attempt to parse the result into a \code{data.frame}}
}
\value{
a \code{list} or \code{data.frame} containing all the elements of a performAdExclusionRuleActionResponse
}
\description{
Performs action on AdExclusionRule objects that satisfy the given Statement query. a set of ad exclusion rules
}
\seealso{
\href{https://developers.google.com/doubleclick-publishers/docs/reference/v201508/AdExclusionRuleService#performAdExclusionRuleAction}{Google Documentation for performAdExclusionRuleAction}
}

