% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AdExclusionRuleService.R
\name{dfp_performAdExclusionRuleAction}
\alias{dfp_performAdExclusionRuleAction}
\title{performAdExclusionRuleAction}
\usage{
dfp_performAdExclusionRuleAction(request_data, as_df = TRUE,
  verbose = FALSE)
}
\arguments{
\item{request_data}{a \code{list} or \code{data.frame} of data elements
to be formatted for a SOAP
request (XML format, but passed as character string)}

\item{as_df}{a boolean indicating whether to attempt to parse the result into
a \code{data.frame}}

\item{verbose}{a boolean indicating whether to print the service URL and POSTed XML}
}
\value{
a \code{data.frame} or \code{list} containing all the elements of a performAdExclusionRuleActionResponse
}
\description{
Performs action on AdExclusionRule objects that satisfy the given Statement query.
}
\examples{
\dontrun{
 res <- dfp_performAdExclusionRuleAction(request_data)
}
}
\seealso{
\href{https://developers.google.com/doubleclick-publishers/docs/reference/v201802/AdExclusionRuleService#performAdExclusionRuleAction}{Google Documentation for performAdExclusionRuleAction}
}
