% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InventoryService.R
\name{dfp_performAdUnitAction}
\alias{dfp_performAdUnitAction}
\title{performAdUnitAction}
\usage{
dfp_performAdUnitAction(request_data)
}
\arguments{
\item{request_data}{a \code{list} or \code{data.frame} of data elements
to be formatted for a SOAP request (XML format, but passed as character string)}
}
\value{
a \code{list} containing all the elements of a performAdUnitActionResponse
}
\description{
Performs actions on AdUnit objects that match the given Statement query. a set of ad units
}
\seealso{
\href{https://developers.google.com/doubleclick-publishers/docs/reference/v201508/InventoryService#performAdUnitAction}{Google Documentation for performAdUnitAction}
}

