% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AudienceSegmentService.R
\name{dfp_performAudienceSegmentAction}
\alias{dfp_performAudienceSegmentAction}
\title{performAudienceSegmentAction}
\usage{
dfp_performAudienceSegmentAction()
}
\value{
a \code{list} containing all the elements of a performAudienceSegmentActionResponse
}
\description{
Performs the given AudienceSegmentAction on the set of segments identified by the given statement. segments
}
\seealso{
\href{https://developers.google.com/doubleclick-publishers/docs/reference/v201508/AudienceSegmentService#performAudienceSegmentAction}{Google Documentation for performAudienceSegmentAction}
}

