% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BaseRateService.R
\name{dfp_performBaseRateAction}
\alias{dfp_performBaseRateAction}
\title{performBaseRateAction}
\usage{
dfp_performBaseRateAction(request_data)
}
\arguments{
\item{request_data}{a \code{list} or \code{data.frame} of data elements
to be formatted for a SOAP request (XML format, but passed as character string)}
}
\value{
a \code{list} containing all the elements of a performBaseRateActionResponse
}
\description{
Performs actions on BaseRate objects that satisfy the given Statement query. a set of base rates.
}
\seealso{
\href{https://developers.google.com/doubleclick-publishers/docs/reference/v201508/BaseRateService#performBaseRateAction}{Google Documentation for performBaseRateAction}
}

