% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ContentBundleService.R
\name{dfp_performContentBundleAction}
\alias{dfp_performContentBundleAction}
\title{performContentBundleAction}
\usage{
dfp_performContentBundleAction(request_data, as_df=FALSE)
}
\arguments{
\item{request_data}{a \code{list} or \code{data.frame} of data elements
to be formatted for a SOAP request (XML format, but passed as character string)}

\item{as_df}{a boolean indicating whether to attempt to parse the result into a \code{data.frame}}
}
\value{
a \code{list} or \code{data.frame} containing all the elements of a performContentBundleActionResponse
}
\description{
Performs actions on ContentBundle objects that match the given Statement query. a set of content bundles
}
\seealso{
\href{https://developers.google.com/doubleclick-publishers/docs/reference/v201508/ContentBundleService#performContentBundleAction}{Google Documentation for performContentBundleAction}
}

