% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ContentMetadataKeyHierarchyService.R
\name{dfp_performContentMetadataKeyHierarchyAction}
\alias{dfp_performContentMetadataKeyHierarchyAction}
\title{performContentMetadataKeyHierarchyAction}
\usage{
dfp_performContentMetadataKeyHierarchyAction(request_data)
}
\arguments{
\item{request_data}{a \code{list} or \code{data.frame} of data elements
to be formatted for a SOAP request (XML format, but passed as character string)}
}
\value{
a \code{list} containing all the elements of a performContentMetadataKeyHierarchyActionResponse
}
\description{
Performs actions on ContentMetadataKeyHierarchy objects that match the given Statement query. a set of hierarchies
}
\seealso{
\href{https://developers.google.com/doubleclick-publishers/docs/reference/v201508/ContentMetadataKeyHierarchyService#performContentMetadataKeyHierarchyAction}{Google Documentation for performContentMetadataKeyHierarchyAction}
}

