% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ContentMetadataKeyHierarchyService.R
\name{dfp_performContentMetadataKeyHierarchyAction}
\alias{dfp_performContentMetadataKeyHierarchyAction}
\title{performContentMetadataKeyHierarchyAction}
\usage{
dfp_performContentMetadataKeyHierarchyAction()
}
\value{
a \code{list} containing all the elements of a performContentMetadataKeyHierarchyActionResponse
}
\description{
Performs actions on ContentMetadataKeyHierarchy objects that match the given Statement query. a set of hierarchies
}

