% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreativeWrapperService.R
\name{dfp_performCreativeWrapperAction}
\alias{dfp_performCreativeWrapperAction}
\title{performCreativeWrapperAction}
\usage{
dfp_performCreativeWrapperAction(request_data)
}
\arguments{
\item{request_data}{a \code{list} or \code{data.frame} of data elements
to be formatted for a SOAP request (XML format, but passed as character string)}
}
\value{
a \code{list} containing all the elements of a performCreativeWrapperActionResponse
}
\description{
Performs actions on CreativeWrapper objects that match the given Statement query. a set of labels
}
\seealso{
\href{https://developers.google.com/doubleclick-publishers/docs/reference/v201508/CreativeWrapperService#performCreativeWrapperAction}{Google Documentation for performCreativeWrapperAction}
}

