% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CustomTargetingService.R
\name{dfp_performCustomTargetingKeyAction}
\alias{dfp_performCustomTargetingKeyAction}
\title{performCustomTargetingKeyAction}
\usage{
dfp_performCustomTargetingKeyAction()
}
\value{
a \code{list} containing all the elements of a performCustomTargetingKeyActionResponse
}
\description{
Performs actions on CustomTargetingKey objects that match the given Statement query. a set of custom targeting keys
}

