% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CustomTargetingService.R
\name{dfp_performCustomTargetingValueAction}
\alias{dfp_performCustomTargetingValueAction}
\title{performCustomTargetingValueAction}
\usage{
dfp_performCustomTargetingValueAction()
}
\value{
a \code{list} containing all the elements of a performCustomTargetingValueActionResponse
}
\description{
Performs actions on CustomTargetingValue objects that match the given Statement query. a set of ad units
}
\seealso{
\href{https://developers.google.com/doubleclick-publishers/docs/reference/v201508/CustomTargetingService#performCustomTargetingValueAction}{Google Documentation for performCustomTargetingValueAction}
}

