% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CustomTargetingService.R
\name{dfp_performCustomTargetingValueAction}
\alias{dfp_performCustomTargetingValueAction}
\title{performCustomTargetingValueAction}
\usage{
dfp_performCustomTargetingValueAction(request_data, as_df=TRUE)
}
\arguments{
\item{request_data}{a \code{list} or \code{data.frame} of data elements
to be formatted for a SOAP request (XML format, but passed as character string)}

\item{as_df}{a boolean indicating whether to attempt to parse the result into a \code{data.frame}}
}
\value{
a \code{data.frame} or \code{list} containing all the elements of a performCustomTargetingValueActionResponse
}
\description{
Performs actions on CustomTargetingValue objects that match the given Statement query. a set of ad units
}
\seealso{
\href{https://developers.google.com/doubleclick-publishers/docs/reference/v201508/CustomTargetingService#performCustomTargetingValueAction}{Google Documentation for performCustomTargetingValueAction}
}

