% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LineItemCreativeAssociationService.R
\name{dfp_performLineItemCreativeAssociationAction}
\alias{dfp_performLineItemCreativeAssociationAction}
\title{performLineItemCreativeAssociationAction}
\usage{
dfp_performLineItemCreativeAssociationAction(request_data, as_df=FALSE)
}
\arguments{
\item{request_data}{a \code{list} or \code{data.frame} of data elements
to be formatted for a SOAP request (XML format, but passed as character string)}

\item{as_df}{a boolean indicating whether to attempt to parse the result into a \code{data.frame}}
}
\value{
a \code{list} or \code{data.frame} containing all the elements of a performLineItemCreativeAssociationActionResponse
}
\description{
Performs actions on LineItemCreativeAssociation objects that match the given Statement query. a set of line item creative associations
}
\seealso{
\href{https://developers.google.com/doubleclick-publishers/docs/reference/v201508/LineItemCreativeAssociationService#performLineItemCreativeAssociationAction}{Google Documentation for performLineItemCreativeAssociationAction}
}

