% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LiveStreamEventService.R
\name{dfp_performLiveStreamEventAction}
\alias{dfp_performLiveStreamEventAction}
\title{performLiveStreamEventAction}
\usage{
dfp_performLiveStreamEventAction()
}
\value{
a \code{list} containing all the elements of a performLiveStreamEventActionResponse
}
\description{
Performs actions on LiveStreamEvent objects that match the given Statement query. a set of live stream events
}
\seealso{
\href{https://developers.google.com/doubleclick-publishers/docs/reference/v201508/LiveStreamEventService#performLiveStreamEventAction}{Google Documentation for performLiveStreamEventAction}
}

