% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LiveStreamEventService.R
\name{dfp_performLiveStreamEventAction}
\alias{dfp_performLiveStreamEventAction}
\title{performLiveStreamEventAction}
\usage{
dfp_performLiveStreamEventAction(request_data)
}
\arguments{
\item{request_data}{a \code{list} or \code{data.frame} of data elements
to be formatted for a SOAP request (XML format, but passed as character string)}
}
\value{
a \code{list} containing all the elements of a performLiveStreamEventActionResponse
}
\description{
Performs actions on LiveStreamEvent objects that match the given Statement query. a set of live stream events
}
\seealso{
\href{https://developers.google.com/doubleclick-publishers/docs/reference/v201508/LiveStreamEventService#performLiveStreamEventAction}{Google Documentation for performLiveStreamEventAction}
}

