% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProductService.R
\name{dfp_performProductAction}
\alias{dfp_performProductAction}
\title{performProductAction}
\usage{
dfp_performProductAction()
}
\value{
a \code{list} containing all the elements of a performProductActionResponse
}
\description{
Performs action on Product objects that satisfy the given Statement.
}

