% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProductPackageItemService.R
\name{dfp_performProductPackageItemAction}
\alias{dfp_performProductPackageItemAction}
\title{performProductPackageItemAction}
\usage{
dfp_performProductPackageItemAction()
}
\value{
a \code{list} containing all the elements of a performProductPackageItemActionResponse
}
\description{
Performs actions on ProductPackageItem objects that satisfy the given Statement query. a set of product package items
}

