% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProductTemplateService.R
\name{dfp_performProductTemplateAction}
\alias{dfp_performProductTemplateAction}
\title{performProductTemplateAction}
\usage{
dfp_performProductTemplateAction()
}
\value{
a \code{list} containing all the elements of a performProductTemplateActionResponse
}
\description{
Performs action on ProductTemplate objects that satisfy the given Statement query. a set of product templates
}
\seealso{
\href{https://developers.google.com/doubleclick-publishers/docs/reference/v201508/ProductTemplateService#performProductTemplateAction}{Google Documentation for performProductTemplateAction}
}

