% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProposalService.R
\name{dfp_performProposalAction}
\alias{dfp_performProposalAction}
\title{performProposalAction}
\usage{
dfp_performProposalAction(request_data)
}
\arguments{
\item{request_data}{a \code{list} or \code{data.frame} of data elements
to be formatted for a SOAP request (XML format, but passed as character string)}
}
\value{
a \code{list} containing all the elements of a performProposalActionResponse
}
\description{
Performs actions on Proposal objects that match the given Statement query. The following fields are also required when submitting proposals for approval: <ul> <li>Proposal advertiser</li> <li>Proposal primarySalesperson</li> <li>Proposal primaryTraffickerId</li> </ul>
}
\seealso{
\href{https://developers.google.com/doubleclick-publishers/docs/reference/v201508/ProposalService#performProposalAction}{Google Documentation for performProposalAction}
}

