% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RateCardService.R
\name{dfp_performRateCardAction}
\alias{dfp_performRateCardAction}
\title{performRateCardAction}
\usage{
dfp_performRateCardAction()
}
\value{
a \code{list} containing all the elements of a performRateCardActionResponse
}
\description{
Performs action on RateCard objects that satisfy the given Statement query. a set of rate cards.
}

