% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReconciliationOrderReportService.R
\name{dfp_performReconciliationOrderReportAction}
\alias{dfp_performReconciliationOrderReportAction}
\title{performReconciliationOrderReportAction}
\usage{
dfp_performReconciliationOrderReportAction(request_data, as_df = TRUE,
  verbose = FALSE)
}
\arguments{
\item{request_data}{a \code{list} or \code{data.frame} of data elements
to be formatted for a SOAP
request (XML format, but passed as character string)}

\item{as_df}{a boolean indicating whether to attempt to parse the result into
a \code{data.frame}}

\item{verbose}{a boolean indicating whether to print the service URL and POSTed XML}
}
\value{
a \code{data.frame} or \code{list} containing all the elements of a performReconciliationOrderReportActionResponse
}
\description{
Performs actions on the ReconciliationOrderReport objects that match the given Statement query. The following fields are supported for filtering: 
\itemize{
  \item{orderId}
  \item{proposalId}
  \item{reconciliationReportId}
}
}
\examples{
\dontrun{
 res <- dfp_performReconciliationOrderReportAction(request_data)
}
}
\seealso{
\href{https://developers.google.com/ad-manager/api/reference/v201905/ReconciliationOrderReportService#performReconciliationOrderReportAction}{Google Documentation for performReconciliationOrderReportAction}
}
