% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SharedAdUnitService.R
\name{dfp_performSharedAdUnitAction}
\alias{dfp_performSharedAdUnitAction}
\title{performSharedAdUnitAction}
\usage{
dfp_performSharedAdUnitAction(request_data)
}
\arguments{
\item{request_data}{a \code{list} or \code{data.frame} of data elements
to be formatted for a SOAP request (XML format, but passed as character string)}
}
\value{
a \code{list} containing all the elements of a performSharedAdUnitActionResponse
}
\description{
Performs actions on shared ad unit objects that match the given Statement query. a set of shared ad units on which the action was performed.
}
\seealso{
\href{https://developers.google.com/doubleclick-publishers/docs/reference/v201508/SharedAdUnitService#performSharedAdUnitAction}{Google Documentation for performSharedAdUnitAction}
}

