% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UserService.R
\name{dfp_performUserAction}
\alias{dfp_performUserAction}
\title{performUserAction}
\usage{
dfp_performUserAction(request_data, as_df=FALSE)
}
\arguments{
\item{request_data}{a \code{list} or \code{data.frame} of data elements
to be formatted for a SOAP request (XML format, but passed as character string)}

\item{as_df}{a boolean indicating whether to attempt to parse the result into a \code{data.frame}}
}
\value{
a \code{list} or \code{data.frame} containing all the elements of a performUserActionResponse
}
\description{
Performs actions on User objects that match the given Statement query. a set of users
}
\seealso{
\href{https://developers.google.com/doubleclick-publishers/docs/reference/v201508/UserService#performUserAction}{Google Documentation for performUserAction}
}

