% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UserService.R
\name{dfp_performUserAction}
\alias{dfp_performUserAction}
\title{performUserAction}
\usage{
dfp_performUserAction(request_data, as_df = TRUE, verbose = FALSE)
}
\arguments{
\item{request_data}{a \code{list} or \code{data.frame} of data elements
to be formatted for a SOAP
request (XML format, but passed as character string)}

\item{as_df}{a boolean indicating whether to attempt to parse the result into
a \code{data.frame}}

\item{verbose}{a boolean indicating whether to print the service URL and POSTed XML}
}
\value{
a \code{data.frame} or \code{list} containing all the elements of a performUserActionResponse
}
\description{
Performs actions on User objects that match the given Statement query.
}
\examples{
\dontrun{
 res <- dfp_performUserAction(request_data)
}
}
\seealso{
\href{https://developers.google.com/ad-manager/api/reference/v201905/UserService#performUserAction}{Google Documentation for performUserAction}
}
