% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UserTeamAssociationService.R
\name{dfp_performUserTeamAssociationAction}
\alias{dfp_performUserTeamAssociationAction}
\title{performUserTeamAssociationAction}
\usage{
dfp_performUserTeamAssociationAction(request_data)
}
\arguments{
\item{request_data}{a \code{list} or \code{data.frame} of data elements
to be formatted for a SOAP request (XML format, but passed as character string)}
}
\value{
a \code{list} containing all the elements of a performUserTeamAssociationActionResponse
}
\description{
Performs actions on UserTeamAssociation objects that match the given Statement query. a set of user team associations
}
\seealso{
\href{https://developers.google.com/doubleclick-publishers/docs/reference/v201508/UserTeamAssociationService#performUserTeamAssociationAction}{Google Documentation for performUserTeamAssociationAction}
}

