% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WorkflowRequestService.R
\name{dfp_performWorkflowRequestAction}
\alias{dfp_performWorkflowRequestAction}
\title{performWorkflowRequestAction}
\usage{
dfp_performWorkflowRequestAction()
}
\value{
a \code{list} containing all the elements of a performWorkflowRequestActionResponse
}
\description{
Perform actions on WorkflowRequest objects that match the given Statement query. requests
}

