% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PublisherQueryLanguageService.R
\name{dfp_select}
\alias{dfp_select}
\title{select}
\usage{
dfp_select(request_data, as_df=FALSE)
}
\arguments{
\item{request_data}{a \code{list} or \code{data.frame} of data elements
to be formatted for a SOAP request (XML format, but passed as character string)}

\item{as_df}{a boolean indicating whether to attempt to parse the result into a \code{data.frame}}
}
\value{
a \code{data.frame} or \code{list} containing all the elements of a selectResponse
}
\description{
Retrieves rows of data that satisfy the given Statement query from the system. specify what data needs to returned
}
\seealso{
\href{https://developers.google.com/doubleclick-publishers/docs/reference/v201508/PublisherQueryLanguageService#select}{Google Documentation for select}
}

