% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AudienceSegmentService.R
\name{dfp_updateAudienceSegments}
\alias{dfp_updateAudienceSegments}
\title{updateAudienceSegments}
\usage{
dfp_updateAudienceSegments(request_data)
}
\arguments{
\item{request_data}{a \code{list} or \code{data.frame} of data elements
to be formatted for a SOAP request (XML format, but passed as character string)}
}
\value{
a \code{list} containing all the elements of a updateAudienceSegmentsResponse
}
\description{
Updates the given RuleBasedFirstPartyAudienceSegment objects.
}

