% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CustomTargetingService.R
\name{dfp_updateCustomTargetingValues}
\alias{dfp_updateCustomTargetingValues}
\title{updateCustomTargetingValues}
\usage{
dfp_updateCustomTargetingValues(request_data)
}
\arguments{
\item{request_data}{a \code{list} or \code{data.frame} of data elements
to be formatted for a SOAP request (XML format, but passed as character string)}
}
\value{
a \code{list} containing all the elements of a updateCustomTargetingValuesResponse
}
\description{
Updates the specified CustomTargetingValue objects.
}
\seealso{
\href{https://developers.google.com/doubleclick-publishers/docs/reference/v201508/CustomTargetingService#updateCustomTargetingValues}{Google Documentation for updateCustomTargetingValues}
}

