% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CustomTargetingService.R
\name{dfp_updateCustomTargetingValues}
\alias{dfp_updateCustomTargetingValues}
\title{updateCustomTargetingValues}
\usage{
dfp_updateCustomTargetingValues(request_data, as_df = TRUE,
  verbose = FALSE)
}
\arguments{
\item{request_data}{a \code{list} or \code{data.frame} of data elements
to be formatted for a SOAP
request (XML format, but passed as character string)}

\item{as_df}{a boolean indicating whether to attempt to parse the result into
a \code{data.frame}}

\item{verbose}{a boolean indicating whether to print the service URL and POSTed XML}
}
\value{
a \code{data.frame} or \code{list} containing all the elements of a updateCustomTargetingValuesResponse
}
\description{
Updates the specified CustomTargetingValue objects.
}
\examples{
\dontrun{
 res <- dfp_updateCustomTargetingValues(request_data)
}
}
\seealso{
\href{https://developers.google.com/ad-manager/api/reference/v201905/CustomTargetingService#updateCustomTargetingValues}{Google Documentation for updateCustomTargetingValues}
}
