% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExchangeRateService.R
\name{dfp_updateExchangeRates}
\alias{dfp_updateExchangeRates}
\title{updateExchangeRates}
\usage{
dfp_updateExchangeRates(request_data, as_df = TRUE, verbose = FALSE)
}
\arguments{
\item{request_data}{a \code{list} or \code{data.frame} of data elements
to be formatted for a SOAP
request (XML format, but passed as character string)}

\item{as_df}{a boolean indicating whether to attempt to parse the result into
a \code{data.frame}}

\item{verbose}{a boolean indicating whether to print the service URL and POSTed XML}
}
\value{
a \code{data.frame} or \code{list} containing all the elements of a updateExchangeRatesResponse
}
\description{
Updates the specified ExchangeRate objects.
}
\examples{
\dontrun{
 res <- dfp_updateExchangeRates(request_data)
}
}
\seealso{
\href{https://developers.google.com/ad-manager/api/reference/v201811/ExchangeRateService#updateExchangeRates}{Google Documentation for updateExchangeRates}
}
