% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LineItemService.R
\name{dfp_updateLineItems}
\alias{dfp_updateLineItems}
\title{updateLineItems}
\usage{
dfp_updateLineItems(request_data, as_df = TRUE, verbose = FALSE)
}
\arguments{
\item{request_data}{a \code{list} or \code{data.frame} of data elements
to be formatted for a SOAP
request (XML format, but passed as character string)}

\item{as_df}{a boolean indicating whether to attempt to parse the result into
a \code{data.frame}}

\item{verbose}{a boolean indicating whether to print the service URL and POSTed XML}
}
\value{
a \code{data.frame} or \code{list} containing all the elements of a updateLineItemsResponse
}
\description{
Updates the specified LineItem objects.
}
\examples{
\dontrun{
 res <- dfp_updateLineItems(request_data)
}
}
\seealso{
\href{https://developers.google.com/ad-manager/api/reference/v201811/LineItemService#updateLineItems}{Google Documentation for updateLineItems}
}
