% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfp_auth.R
\name{get_google_token}
\alias{get_google_token}
\title{Retrieve Google token from environment}
\usage{
get_google_token()
}
\description{
Get token if it's previously stored, else prompt user to get one.
}
\note{
This function is meant to be used internally. Only use when debugging.
}
\keyword{internal}
